RCCL: Swapped legacy net_ib with net_ib from NCCL 2.29
