+++ projects/rccl/src/CMakeLists.txt
@@ -368,13 +368,23 @@ set(SRC_FILES
   transport/coll_net.cc
   transport/generic.cc
   transport/net.cc
-  transport/net_ib.cc
   transport/net_ib_limits.h
+  transport/net_ib/common.cc
+  transport/net_ib/common.h
+  transport/net_ib/connect.cc
+  transport/net_ib/connect.h
+  transport/net_ib/gdr.cc
+  transport/net_ib/gin.cc
+  transport/net_ib/gin.h
+  transport/net_ib/init.cc
+  transport/net_ib/p2p.cc
+  transport/net_ib/p2p.h
+  transport/net_ib/p2p_resiliency.cc
+  transport/net_ib/p2p_resiliency.h
+  transport/net_ib/reg.cc
   transport/net_ib_cast/net_ib_cast_inspect.h
   transport/net_ib_cast/net_ib_fault_inject.h
-  transport/net_ib/gin.h
-  # net_ib_rocm.cc is generated by rocmIb.cmake directly into the hipify staging area
-  # so it is not listed here - it's added to HIP_SOURCES separately below
+  transport/net_ib/gdaki/gin_host_gdaki.h
   transport/net_socket.cc
   transport/nvls.cc
   transport/p2p.cc
@@ -382,7 +392,6 @@ set(SRC_FILES
   transport/shm.cc
   gin/gin_host.cc
   gin/gin_host_proxy.cc
-  transport/net_ib/gdaki/gin_host_gdaki.h
   include/latency_profiler/CollTrace.h
   include/latency_profiler/CollTraceEvent.h
   include/latency_profiler/CollTraceFunc.h
@@ -504,9 +513,23 @@ add_custom_target(hipify_all DEPENDS ${HIP_SOURCES})
 # could proceed before hipify_all stage had completed.
 add_dependencies(copy_nccl_device_headers hipify_all)
 
-# Add the generated net_ib_rocm.cc file (created by rocmIb.cmake in the staging area)
-# This file is patched and transformed directly into HIPIFY_DIR, bypassing the hipify step
-list(APPEND HIP_SOURCES "${HIPIFY_DIR}/src/transport/net_ib_rocm.cc")
+# net_ib/ headers collide with other basenames (common.h, gin.h, p2p.h) and are hipified
+# with _tmp suffixes. Create symlinks so that #include "common.h" etc. resolve correctly
+# when compiling files in hipify/src/transport/net_ib/.
+set(NET_IB_HIPIFY_DIR "${HIPIFY_DIR}/src/transport/net_ib")
+set(NET_IB_HEADER_SYMLINKS "")
+foreach(_hdr common gin p2p)
+  add_custom_command(
+    OUTPUT  "${NET_IB_HIPIFY_DIR}/${_hdr}.h"
+    DEPENDS "${NET_IB_HIPIFY_DIR}/${_hdr}_tmp.h"
+    COMMAND ${CMAKE_COMMAND} -E create_symlink
+            "${_hdr}_tmp.h" "${NET_IB_HIPIFY_DIR}/${_hdr}.h"
+    COMMENT "Symlinking net_ib/${_hdr}_tmp.h -> ${_hdr}.h"
+  )
+  list(APPEND NET_IB_HEADER_SYMLINKS "${NET_IB_HIPIFY_DIR}/${_hdr}.h")
+endforeach()
+add_custom_target(net_ib_header_symlinks DEPENDS ${NET_IB_HEADER_SYMLINKS})
+
 
 # Generate device/host tables and all the collective functions that are going to be in librccl.so
 #==================================================================================================
@@ -602,7 +625,7 @@ add_library(rccl ${HIP_SOURCES})
 
 ## Set RCCL dependencies
 ## Ensure git version is updated before building rccl
-add_dependencies(rccl update_git_version copy_nccl_device_headers)
+add_dependencies(rccl update_git_version copy_nccl_device_headers net_ib_header_symlinks)
 
 ## Set RCCL include directories
 target_include_directories(rccl PRIVATE ${PROJECT_BINARY_DIR}/include)        # for generated rccl.h header
@@ -616,6 +639,8 @@ target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/src/include/nccl_device)
 target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/src/include/ionic)
 target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/src/include/plugin)
 target_include_directories(rccl PRIVATE ${NET_IB_CAST_INCLUDE_DIRS})
+target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/src/transport)
+target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/src/transport/net_ib)
 target_include_directories(rccl PRIVATE ${HIPIFY_DIR}/gensrc)
 target_include_directories(rccl PRIVATE ${HSA_INCLUDE_PATH})
 target_include_directories(rccl PRIVATE ${SMI_INCLUDE_DIR})
