+++ projects/rccl/src/include/net.h
@@ -39,9 +39,7 @@ extern ncclResult_t rcclNetP2pPolicy(void* handle, int isP2p);
 extern int64_t ncclParamDmaBufEnable();
 
 #if defined(__HIP_PLATFORM_AMD__) || defined(__HIPCC__)
-extern ncclNet_t rocmNetIb;
 extern ncclNet_t netIbCast;
-extern ncclResult_t rcclRocmNetP2pPolicy(void* handle, int isP2p);
 extern ncclResult_t rcclCastNetP2pPolicy(void* handle, int isP2p);
 
 enum rcclIBNicType {
