--- projects/rccl/src/transport/net_ib.cc
@@ -1,3769 +1,0 @@
-/*************************************************************************
- * Copyright (c) 2016-2022, NVIDIA CORPORATION. All rights reserved.
- * Modifications Copyright (c) 2019-2023 Advanced Micro Devices, Inc. All rights reserved.
- *
- * See LICENSE.txt for license information
- ************************************************************************/
-
-#include "nccl.h"
-#include "core.h"
-#include "socket.h"
-#include "net.h"
-#include "graph.h"
-#include "utils.h"
-#include "param.h"
-#include "profiler/net_ib.h"
-
-#include <assert.h>
-#include <pthread.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <poll.h>
-#include <sys/types.h>
-#include <unistd.h>
-#include <mutex>
-#define ENABLE_TIMER 0
-#include "timer.h"
-#include <sys/utsname.h>
-
-#include "ibvwrap.h"
-#include "mlx5/mlx5dvwrap.h"
-#include "graph/xml.h"
-
-#define MAXSUFFIXSIZE 16
-#define MAXNAMESIZE (64 + MAXSUFFIXSIZE)
-static char ncclIbIfName[MAX_IF_NAME_SIZE+1];
-static union ncclSocketAddress ncclIbIfAddr;
-
-static ncclNetCommConfig_t ibContext;
-
-struct ncclIbMr {
-  uintptr_t addr;
-  size_t pages;
-  int refs;
-  ibv_mr *mr;
-};
-
-struct ncclIbMrCache {
-  struct ncclIbMr *slots;
-  int capacity, population;
-};
-
-static int ncclNMergedIbDevs = -1;
-#define NCCL_IB_MAX_DEVS_PER_NIC 4
-#define MAX_MERGED_DEV_NAME (MAXNAMESIZE*NCCL_IB_MAX_DEVS_PER_NIC)+NCCL_IB_MAX_DEVS_PER_NIC
-struct alignas(64) ncclIbMergedDev {
-  ncclNetVDeviceProps_t vProps;
-  int speed;
-  char devName[MAX_MERGED_DEV_NAME]; // Up to NCCL_IB_MAX_DEVS_PER_NIC * name size, and a character for each '+'
-};
-
-struct ncclIbStats {
-  int fatalErrorCount;
-};
-
-enum ncclIbProvider {
-  IB_PROVIDER_NONE = 0,
-  IB_PROVIDER_MLX5 = 1,
-  IB_PROVIDER_MAX = 2,
-};
-
-const char* ibProviderName[] = {
-  "None",
-  "Mlx5",
-};
-
-static int ncclNIbDevs = -1;
-struct alignas(64) ncclIbDev {
-  std::mutex mutex;
-  int device;
-  uint64_t guid;
-  uint8_t portNum;
-  uint8_t link;
-  int speed;
-  ibv_context* context;
-  int pdRefs;
-  ibv_pd* pd;
-  char devName[MAXNAMESIZE];
-  char* pciPath;
-  char* virtualPciPath;
-  int realPort;
-  int maxQp;
-  float latency;
-  struct ncclIbMrCache mrCache;
-  int ar; // ADAPTIVE_ROUTING
-  struct ibv_port_attr portAttr;
-  struct ncclIbStats stats;
-  int dmaBufSupported;
-  enum ncclIbProvider ibProvider;
-  union {
-    struct {
-      int dataDirect;
-    } mlx5;
-  } capsProvider;
-};
-
-#define MAX_IB_DEVS  32
-#define MAX_IB_VDEVS MAX_IB_DEVS*8
-struct ncclIbMergedDev ncclIbMergedDevs[MAX_IB_VDEVS];
-struct ncclIbDev ncclIbDevs[MAX_IB_DEVS];
-static std::mutex ncclIbMutex;
-static int ncclIbRelaxedOrderingEnabled = 0;
-
-// With ncclNet_v11_t the NCCL core initializes the network plugin per-communicator
-// rather than once for all communicators. However, the internal plugin implementation
-// still assumes the plugin is initialized only once across all communicators. The ref
-// counter makes sure the plugin internally initializes only once. When per communicator
-// context support is added to the plugin the ref counter can be removed.
-static int netRefCount;
-
-#define NCCL_IB_LLSTR(ll) (((ll) == IBV_LINK_LAYER_INFINIBAND) ? "IB" : (((ll) == IBV_LINK_LAYER_ETHERNET) ? "RoCE" : "UNSPECIFIED"))
-
-#define NCCL_IB_SL_DEFAULT 0
-#define NCCL_IB_TC_DEFAULT 0
-
-NCCL_PARAM(IbGidIndex, "IB_GID_INDEX", -1);
-NCCL_PARAM(IbRoutableFlidIbGidIndex, "IB_ROUTABLE_FLID_GID_INDEX", 1);
-NCCL_PARAM(IbRoceVersionNum, "IB_ROCE_VERSION_NUM", 2);
-NCCL_PARAM(IbTimeout, "IB_TIMEOUT", 20);
-NCCL_PARAM(IbRetryCnt, "IB_RETRY_CNT", 7);
-NCCL_PARAM(IbPkey, "IB_PKEY", 0);
-NCCL_PARAM(IbUseInline, "IB_USE_INLINE", 0);
-NCCL_PARAM(IbSl, "IB_SL", -1);
-NCCL_PARAM(IbTc, "IB_TC", -1);
-NCCL_PARAM(IbArThreshold, "IB_AR_THRESHOLD", 8192);
-NCCL_PARAM(IbPciRelaxedOrdering, "IB_PCI_RELAXED_ORDERING", 2);
-NCCL_PARAM(IbAdaptiveRouting, "IB_ADAPTIVE_ROUTING", -2);
-NCCL_PARAM(IbFifoTc, "IB_FIFO_TC", -1);
-NCCL_PARAM(IbAsyncEvents,"IB_RETURN_ASYNC_EVENTS",1);
-NCCL_PARAM(IbEceEnable,"IB_ECE_ENABLE",1);
-NCCL_PARAM(IbDataDirect,"IB_DATA_DIRECT",1);
-NCCL_PARAM(IbQpsPerConn, "IB_QPS_PER_CONNECTION", 1);
-RCCL_PARAM(IbQpsPerP2p, "IB_QPS_PER_P2P", 0);
-
-static ncclResult_t ncclIbStatsInit(struct ncclIbStats* stat) {
-  __atomic_store_n(&stat->fatalErrorCount, 0, __ATOMIC_RELAXED);
-  return ncclSuccess;
-}
-static void ncclIbStatsFatalError(struct ncclIbStats* stat){
-  __atomic_fetch_add(&stat->fatalErrorCount, 1, __ATOMIC_RELAXED);
-}
-static ncclResult_t ncclIbStatsCheckFatalCount(struct ncclIbStats* stat, const char* funcName) {
-  if (ncclParamIbAsyncEvents() && __atomic_load_n(&stat->fatalErrorCount, __ATOMIC_RELAXED)) {
-    ERROR("RCCL encountered a communication fatal error (detected in %s)\n", funcName);
-    ERROR("RCCL cannot recover from this network failure and now exiting. Please check the network health.");
-    return ncclSystemError;
-  }
-  return ncclSuccess;
-}
-static void ncclIbQpFatalError(struct ibv_qp* qp) {
-  ncclIbStatsFatalError((struct ncclIbStats*)qp->qp_context);
-}
-static void ncclIbCqFatalError(struct ibv_cq* cq) {
-  ncclIbStatsFatalError((struct ncclIbStats*)cq->cq_context);
-}
-// Calculate number of QPs based on P2P flag and device counts
-static int ncclIbCalculateNqps(int isP2p, int localNdevs, int remoteNdevs, const char* funcName) {
-  auto qp_multiplier = (rcclParamIbQpsPerP2p() > 0 && isP2p) ? 
-                       rcclParamIbQpsPerP2p() : ncclParamIbQpsPerConn();
-  int localNqps = qp_multiplier * localNdevs;
-  int remoteNqps = qp_multiplier * remoteNdevs;
-  int maxNqps = (remoteNqps > localNqps) ? remoteNqps : localNqps;
-  INFO(NCCL_NET, "NET/IB: %s Max Nqps=%d, localNqps=%d, remoteNqps=%d", 
-       funcName, maxNqps, localNqps, remoteNqps);
-  return maxNqps;
-}
-
-static void ncclIbDevFatalError(struct ncclIbDev* dev) {
-  ncclIbStatsFatalError(&dev->stats);
-}
-
-// [RCCL] NCCL 2.29.7 promoted ibvWcStatusStr / ibvWcOpcodeStr to public
-// helpers in ibvwrap.{h,cc}. The local static definitions that used to live
-// here would otherwise collide with the new ibvwrap.h declarations.
-pthread_t ncclIbAsyncThread;
-static void* ncclIbAsyncThreadMain(void* args) {
-  struct ncclIbDev* dev = (struct ncclIbDev*)args;
-  while (1) {
-    struct ibv_async_event event;
-    if (ncclSuccess != wrap_ibv_get_async_event(dev->context, &event)) { break; }
-    char *str;
-    struct ibv_cq* cq = event.element.cq;    // only valid if CQ error
-    struct ibv_qp* qp = event.element.qp;    // only valid if QP error
-    struct ibv_srq* srq = event.element.srq; // only valid if SRQ error
-    if (ncclSuccess != wrap_ibv_event_type_str(&str, event.event_type)) { break; }
-    switch (event.event_type) {
-    case IBV_EVENT_DEVICE_FATAL:
-      // the above is device fatal error
-      WARN("NET/IB : %s:%d async fatal event: %s", dev->devName, dev->portNum, str);
-      ncclIbDevFatalError(dev);
-      break;
-    case IBV_EVENT_CQ_ERR:
-      // the above is a CQ fatal error
-      WARN("NET/IB : %s:%d async fatal event on CQ (%p): %s", dev->devName, dev->portNum, cq, str);
-      ncclIbCqFatalError(cq);
-      break;
-    case IBV_EVENT_QP_FATAL:
-    case IBV_EVENT_QP_REQ_ERR:
-    case IBV_EVENT_QP_ACCESS_ERR:
-      // the above are QP fatal errors
-      WARN("NET/IB : %s:%d async fatal event on QP (%p): %s", dev->devName, dev->portNum, qp, str);
-      ncclIbQpFatalError(qp);
-      break;
-    case IBV_EVENT_SRQ_ERR:
-      // SRQ are not used in NCCL
-      WARN("NET/IB : %s:%d async fatal event on SRQ, unused for now (%p): %s", dev->devName, dev->portNum, srq, str);
-      break;
-    case IBV_EVENT_GID_CHANGE:
-      WARN("NET/IB : %s:%d GID table changed", dev->devName, dev->portNum);
-      break;
-    case IBV_EVENT_PATH_MIG_ERR:
-    case IBV_EVENT_PORT_ERR:
-    case IBV_EVENT_PATH_MIG:
-    case IBV_EVENT_PORT_ACTIVE:
-    case IBV_EVENT_SQ_DRAINED:
-    case IBV_EVENT_LID_CHANGE:
-    case IBV_EVENT_PKEY_CHANGE:
-    case IBV_EVENT_SM_CHANGE:
-    case IBV_EVENT_QP_LAST_WQE_REACHED:
-    case IBV_EVENT_CLIENT_REREGISTER:
-    case IBV_EVENT_SRQ_LIMIT_REACHED:
-      // the above are non-fatal
-      WARN("NET/IB : %s:%d Got async event: %s", dev->devName, dev->portNum, str);
-      break;
-    case IBV_EVENT_COMM_EST:
-      break;
-    default:
-      WARN("NET/IB : %s:%d unknown event type (%d)", dev->devName, dev->portNum, event.event_type);
-      break;
-    }
-    // acknowledgment needs to happen last to avoid user-after-free
-    if (ncclSuccess != wrap_ibv_ack_async_event(&event)) { break; }
-  }
-  return NULL;
-}
-
-static sa_family_t envIbAddrFamily(void) {
-  sa_family_t family = AF_INET;
-  const char* env = ncclGetEnv("NCCL_IB_ADDR_FAMILY");
-  if (env == NULL || strlen(env) == 0) {
-    return family;
-  }
-
-  INFO(NCCL_ENV, "NCCL_IB_ADDR_FAMILY set by environment to %s", env);
-
-  if (strcmp(env, "AF_INET") == 0) {
-    family = AF_INET;
-  } else if (strcmp(env, "AF_INET6") == 0) {
-    family = AF_INET6;
-  }
-
-  return family;
-}
-
-static void* envIbAddrRange(sa_family_t af, int* mask) {
-  *mask = 0;
-  static struct in_addr addr;
-  static struct in6_addr addr6;
-  void *ret = (af == AF_INET) ? (void *)&addr : (void *)&addr6;
-
-  const char* env = ncclGetEnv("NCCL_IB_ADDR_RANGE");
-  if (NULL == env || strlen(env) == 0) {
-    return NULL;
-  }
-
-  INFO(NCCL_ENV, "NCCL_IB_ADDR_RANGE set by environment to %s", env);
-
-  char addrString[128] = { 0 };
-  snprintf(addrString, 128, "%s", env);
-  char *addrStrPtr = addrString;
-  char *maskStrPtr = strstr(addrString, "/");
-  if (NULL == maskStrPtr) {
-    return NULL;
-  }
-  *(maskStrPtr++) = '\0';
-
-  if (inet_pton(af, addrStrPtr, ret) == 0) {
-    INFO(NCCL_INIT|NCCL_NET, "NET/IB: Ip address '%s' is invalid for family %s, ignoring address", addrStrPtr, (af == AF_INET) ? "AF_INET" : "AF_INET6");
-    return NULL;
-  }
-
-  *mask = (int)strtol(maskStrPtr, NULL, 10);
-  if (af == AF_INET && *mask > 32) {
-    INFO(NCCL_INIT|NCCL_NET, "NET/IB: Ip address mask '%d' is invalid for family %s, ignoring mask", *mask, (af == AF_INET) ? "AF_INET" : "AF_INET6");
-    *mask = 0;
-    ret = NULL;
-  } else if (af == AF_INET6 && *mask > 128) {
-    INFO(NCCL_INIT|NCCL_NET, "NET/IB: Ip address mask '%d' is invalid for family %s, ignoring mask", *mask, (af == AF_INET) ? "AF_INET" : "AF_INET6");
-    *mask = 0;
-    ret = NULL;
-  }
-
-  return ret;
-}
-
-static sa_family_t getGidAddrFamily(union ibv_gid* gid) {
-  const struct in6_addr *a = (struct in6_addr *)gid->raw;
-  bool isIpV4Mapped = ((a->s6_addr32[0] | a->s6_addr32[1]) | (a->s6_addr32[2] ^ htonl(0x0000ffff))) == 0UL;
-  bool isIpV4MappedMulticast = (a->s6_addr32[0] == htonl(0xff0e0000) && ((a->s6_addr32[1] | (a->s6_addr32[2] ^ htonl(0x0000ffff))) == 0UL));
-  return (isIpV4Mapped || isIpV4MappedMulticast) ? AF_INET : AF_INET6;
-}
-
-static bool matchGidAddrPrefix(sa_family_t af, void* prefix, int prefixlen, union ibv_gid* gid) {
-  struct in_addr *base = NULL;
-  struct in6_addr *base6 = NULL;
-  struct in6_addr *addr6 = NULL;;
-  if (af == AF_INET) {
-    base = (struct in_addr *)prefix;
-  } else {
-    base6 = (struct in6_addr *)prefix;
-  }
-  addr6 = (struct in6_addr *)gid->raw;
-
-#define NETMASK(bits) (htonl(0xffffffff ^ ((1 << (32 - bits)) - 1)))
-
-  int i = 0;
-  while (prefixlen > 0 && i < 4) {
-    if (af == AF_INET) {
-      int mask = NETMASK(prefixlen);
-      if ((base->s_addr & mask) ^ (addr6->s6_addr32[3] & mask)) {
-        break;
-      }
-      prefixlen = 0;
-      break;
-    } else {
-      if (prefixlen >= 32) {
-        if (base6->s6_addr32[i] ^ addr6->s6_addr32[i]) {
-          break;
-        }
-        prefixlen -= 32;
-        ++i;
-      } else {
-        int mask = NETMASK(prefixlen);
-        if ((base6->s6_addr32[i] & mask) ^ (addr6->s6_addr32[i] & mask)) {
-          break;
-        }
-        prefixlen = 0;
-      }
-    }
-  }
-
-  return (prefixlen == 0) ? true : false;
-}
-
-static bool configuredGid(union ibv_gid* gid) {
-  const struct in6_addr *a = (struct in6_addr *)gid->raw;
-  int trailer = (a->s6_addr32[1] | a->s6_addr32[2] | a->s6_addr32[3]);
-  if (((a->s6_addr32[0] | trailer) == 0UL) || ((a->s6_addr32[0] == htonl(0xfe800000)) && (trailer == 0UL))) {
-    return false;
-  }
-  return true;
-}
-
-static bool linkLocalGid(union ibv_gid* gid) {
-  const struct in6_addr *a = (struct in6_addr *)gid->raw;
-  if (a->s6_addr32[0] == htonl(0xfe800000) && a->s6_addr32[1] == 0UL) {
-    return true;
-  }
-  return false;
-}
-
-static bool validGid(union ibv_gid* gid) {
-  return (configuredGid(gid) && !linkLocalGid(gid));
-}
-
-static ncclResult_t ncclIbRoceGetVersionNum(const char* deviceName, int portNum, int gidIndex, int* version) {
-  char gidRoceVerStr[16] = { 0 };
-  char roceTypePath[PATH_MAX] = { 0 };
-  snprintf(roceTypePath, sizeof(roceTypePath), "/sys/class/infiniband/%s/ports/%d/gid_attrs/types/%d", deviceName, portNum, gidIndex);
-
-  int fd = open(roceTypePath, O_RDONLY);
-  if (fd == -1) {
-    WARN("NET/IB: open failed in ncclIbRoceGetVersionNum: %s", strerror(errno));
-    return ncclSystemError;
-  }
-  int ret = read(fd, gidRoceVerStr, 15);
-  close(fd);
-
-  if (ret == -1) {
-    // In containerized environments, read could return EINVAL if the GID index is not mapped to the
-    // container sysfs. In this case return ncclSuccess and let the caller move to next GID index.
-    if (errno == EINVAL) return ncclSuccess;
-    WARN("NET/IB: read failed in ncclIbRoceGetVersionNum: %s", strerror(errno));
-    return ncclSystemError;
-  }
-
-  if (strlen(gidRoceVerStr)) {
-    if (strncmp(gidRoceVerStr, "IB/RoCE v1", strlen("IB/RoCE v1")) == 0 || strncmp(gidRoceVerStr, "RoCE v1", strlen("RoCE v1")) == 0) {
-      *version = 1;
-    } else if (strncmp(gidRoceVerStr, "RoCE v2", strlen("RoCE v2")) == 0) {
-      *version = 2;
-    }
-  }
-
-  return ncclSuccess;
-}
-
-static ncclResult_t ncclUpdateGidIndex(struct ibv_context* context, uint8_t portNum, sa_family_t af, void* prefix, int prefixlen, int roceVer, int gidIndexCandidate, int* gidIndex) {
-  union ibv_gid gid, gidCandidate;
-  NCCLCHECK(wrap_ibv_query_gid(context, portNum, *gidIndex, &gid));
-  NCCLCHECK(wrap_ibv_query_gid(context, portNum, gidIndexCandidate, &gidCandidate));
-
-  sa_family_t usrFam = af;
-  sa_family_t gidFam = getGidAddrFamily(&gid);
-  sa_family_t gidCandidateFam = getGidAddrFamily(&gidCandidate);
-  bool gidCandidateMatchSubnet = matchGidAddrPrefix(usrFam, prefix, prefixlen, &gidCandidate);
-
-  if (gidCandidateFam != gidFam && gidCandidateFam == usrFam && gidCandidateMatchSubnet) {
-    *gidIndex = gidIndexCandidate;
-  } else {
-    if (gidCandidateFam != usrFam || !validGid(&gidCandidate) || !gidCandidateMatchSubnet) {
-      return ncclSuccess;
-    }
-    int usrRoceVer = roceVer;
-    int gidRoceVerNum, gidRoceVerNumCandidate = -1;
-    const char* deviceName = wrap_ibv_get_device_name(context->device);
-    NCCLCHECK(ncclIbRoceGetVersionNum(deviceName, portNum, *gidIndex, &gidRoceVerNum));
-    NCCLCHECK(ncclIbRoceGetVersionNum(deviceName, portNum, gidIndexCandidate, &gidRoceVerNumCandidate));
-    if ((gidRoceVerNum != gidRoceVerNumCandidate || !validGid(&gid)) && gidRoceVerNumCandidate == usrRoceVer) {
-      *gidIndex = gidIndexCandidate;
-    }
-  }
-
-  return ncclSuccess;
-}
-
-// GID Format
-// global:  |              64b  - subnet-prefix                |                 64b - EUI                          |
-// raw   :  | 10b fixed | 22b 0 | 16b FLID | 16b subnet-prefix |                 64b - EUI                          |
-static uint16_t ncclIbExtractLocalSubnetPrefix(uint64_t subnet_prefix)
-{
-  return (be64toh(subnet_prefix) & 0xffff);
-}
-
-static int ncclIbExtractFlid (union ibv_gid *gid)
-{
-  return ntohs(*((uint16_t*)((uintptr_t)(gid->raw) + 4)));
-}
-
-static ncclResult_t ncclIbGetGidIndex(struct ibv_context *context, uint8_t portNum, struct ibv_port_attr* portAttr, int *gidIndex) {
-  int gidTblLen = portAttr->gid_tbl_len;
-
-  //for IB, choose GID Index that will have routable FLID if present
-  if (portAttr->link_layer == IBV_LINK_LAYER_INFINIBAND) {
-    union ibv_gid gid;
-    int routableGidIndex = ncclParamIbRoutableFlidIbGidIndex();
-    if (routableGidIndex < gidTblLen) {
-      NCCLCHECK(wrap_ibv_query_gid(context, portNum, routableGidIndex, &gid));
-      if (ncclIbExtractFlid(&gid) != 0) {
-        *gidIndex = routableGidIndex;
-        return ncclSuccess;
-      }
-    }
-    *gidIndex = 0;
-    return ncclSuccess;
-  }
-
-  //for ROCE
-  *gidIndex = ncclParamIbGidIndex();
-  if (*gidIndex >= 0) {
-    return ncclSuccess;
-  }
-
-  sa_family_t userAddrFamily = envIbAddrFamily();
-  int userRoceVersion = ncclParamIbRoceVersionNum();
-  int prefixlen;
-  void *prefix = envIbAddrRange(userAddrFamily, &prefixlen);
-
-  *gidIndex = 0;
-  for (int gidIndexNext = 1; gidIndexNext < gidTblLen; ++gidIndexNext) {
-    NCCLCHECK(ncclUpdateGidIndex(context, portNum, userAddrFamily, prefix, prefixlen, userRoceVersion, gidIndexNext, gidIndex));
-  }
-
-  return ncclSuccess;
-}
-
-NCCL_PARAM(IbDisable, "IB_DISABLE", 0);
-NCCL_PARAM(IbMergeVfs, "IB_MERGE_VFS", 1);
-NCCL_PARAM(IbMergeNics, "IB_MERGE_NICS", 1);
-
-// Returns 1 if this is the path of two VFs of the same physical device
-static int ncclIbMatchVfPath(char* path1, char* path2) {
-  // Merge multi-port NICs into the same PCI device
-  if (ncclParamIbMergeVfs()) {
-    return strncmp(path1, path2, strlen(path1)-4) == 0;
-  } else {
-    return strncmp(path1, path2, strlen(path1)-1) == 0;
-  }
-}
-
-/**
- * Assumes PCIe path ends with xxxx:xx:xx.x 
- */
- static void ncclIbNormalizePciPath(const char* in, char* out, size_t out_size) {
-  if (!in || !out || out_size == 0) return;
-  // Safe copy with truncation
-  size_t len = strnlen(in, out_size - 1);
-  memmove(out, in, len);
-  out[len] = '\0';
-  if (len < 4) return;
-  // Merge multi-port NICs (.1/.2/.3 -> .0)
-  out[len - 1] = '0';
-  // Merge VFs if enabled
-  if (ncclParamIbMergeVfs()) {
-    out[len - 3] = '0';
-    out[len - 4] = '0';
-  }
-}
-
-static ncclResult_t ncclIbGetPciPath(char* devName, char** path, int* realPort) {
-  char devicePath[PATH_MAX];
-  snprintf(devicePath, PATH_MAX, "/sys/class/infiniband/%s/device", devName);
-  char* p = realpath(devicePath, NULL);
-  // Keep real path unchanged
-  *path = p;
-  if (p == NULL) {
-    WARN("Could not find real path of %s (%s)", devName, devicePath);
-  } else {
-    char normalized[PATH_MAX];
-    ncclIbNormalizePciPath(p, normalized, PATH_MAX);
-    // Keep the real port aside (the ibv port is always 1 on recent cards)
-    *realPort = 0;
-    for (int d = 0; d < ncclNIbDevs; d++) {
-      char otherNorm[PATH_MAX];
-      ncclIbNormalizePciPath(ncclIbDevs[d].pciPath, otherNorm, PATH_MAX);
-      if (ncclIbMatchVfPath(normalized, otherNorm)) {
-        (*realPort)++;
-      }
-    }
-  }
-  return ncclSuccess;
-}
-
-static int ibvWidths[] = { 1, 4, 8, 12, 2 };
-static int ibvSpeeds[] = {
-  2500,  /* SDR */
-  5000,  /* DDR */
-  10000, /* QDR */
-  10000, /* QDR */
-  14000, /* FDR */
-  25000, /* EDR */
-  50000, /* HDR */
-  100000, /* NDR */
-  200000  /* XDR */
-};
-
-static int firstBitSet(int val, int max) {
-  int i = 0;
-  while (i<max && ((val & (1<<i)) == 0)) i++;
-  return i;
-}
-static int ncclIbWidth(int width) {
-  return ibvWidths[firstBitSet(width, sizeof(ibvWidths)/sizeof(int)-1)];
-}
-static int ncclIbSpeed(int speed) {
-  return ibvSpeeds[firstBitSet(speed, sizeof(ibvSpeeds)/sizeof(int)-1)];
-}
-
-// Determine whether RELAXED_ORDERING is enabled and possible
-static int ncclIbRelaxedOrderingCapable(void) {
-  int roMode = ncclParamIbPciRelaxedOrdering();
-  ncclResult_t r = ncclInternalError;
-  if (roMode == 1 || roMode == 2) {
-    // Query IBVERBS_1.8 API - needed for IBV_ACCESS_RELAXED_ORDERING support
-    r = wrap_ibv_reg_mr_iova2(NULL, NULL, NULL, 0, 0, 0);
-  }
-  return r == ncclInternalError ? 0 : 1;
-}
-
-static bool ncclMlx5dvDmaBufCapable(ibv_context *context){
-  ncclResult_t res;
-  int dev_fail = 0;
-
-  struct ibv_pd* pd;
-  NCCLCHECKGOTO(wrap_ibv_alloc_pd(&pd, context), res, failure);
-  // Test kernel DMA-BUF support with a dummy call (fd=-1)
-  (void)wrap_direct_ibv_reg_dmabuf_mr(pd, 0ULL /*offset*/, 0ULL /*len*/, 0ULL /*iova*/, -1 /*fd*/, 0 /*flags*/);
-  // ibv_reg_dmabuf_mr() will fail with EOPNOTSUPP/EPROTONOSUPPORT if not supported (EBADF otherwise)
-  (void)wrap_direct_mlx5dv_reg_dmabuf_mr(pd, 0ULL /*offset*/, 0ULL /*len*/, 0ULL /*iova*/, -1 /*fd*/, 0 /*flags*/, 0 /* mlx5 flags*/);
-  // mlx5dv_reg_dmabuf_mr() will fail with EOPNOTSUPP/EPROTONOSUPPORT if not supported (EBADF otherwise)
-  dev_fail |= (errno == EOPNOTSUPP) || (errno == EPROTONOSUPPORT);
-  NCCLCHECKGOTO(wrap_ibv_dealloc_pd(pd), res, failure);
-  // stop the search and goto failure
-  if (dev_fail) goto failure;
-  return true;
-failure:
-  return false;
-}
-
-/**
- * Extract the PCIe root complex (domain:bus) from a Linux sysfs PCI device path.
- *
- * This function scans a sysfs PCI device path (e.g.
- *   "/sys/devices/pci0000:80/0000:80:03.1/0000:81:00.0")
- * and extracts the first occurrence of the PCI root host bridge identifier
- * of the form "pciDDDD:BB", where:
- *   - DDDD is the PCI domain (hex)
- *   - BB   is the PCI bus number (hex)
- *
- * The extracted root is returned as a string formatted as "DDDD:BB".
- *
- * This helper is typically used to determine whether multiple NICs or devices
- * reside under the same PCIe root complex, which is important for validating
- * NIC merging and avoiding cross-root PCIe traffic.
- *
- * @param pciPath  NUL-terminated sysfs PCI device path
- * @param root     Output buffer receiving the "DDDD:BB" PCI root identifier
- * @param rootLen  Size of the output buffer; must be >= 8
- *
- * @return ncclSuccess on success
- * @return ncclInvalidUsage if inputs are invalid or no PCI root is found
- *
- * Notes:
- * - The function scans for the first valid "pciDDDD:BB" pattern in the path.
- * - The loop is bounded by the length of the string and cannot hang.
- * - This function does not allocate memory.
- */
-static ncclResult_t ncclIbGetPciRootFromPath(
-    const char* pciPath,
-    char* root,
-    size_t rootLen
-) {
-    if (pciPath == NULL || root == NULL || rootLen < 8){
-        return ncclInvalidUsage;
-    }
-    const char* p = strstr(pciPath, "pci");
-    while (p != NULL) {
-        int domain, bus;
-        int chars_read = 0;
-        if (sscanf(p, "pci%4x:%2x%n", &domain, &bus, &chars_read) == 2 &&
-            chars_read == 10) {
-            snprintf(root, rootLen, "%04x:%02x", domain, bus);
-            return ncclSuccess;
-        }
-        p = strstr(p + 1, "pci");
-    }
-    return ncclInvalidUsage;
-}
-
-/**
- * Determine the NUMA node associated with a PCI device from its sysfs path.
- *
- * This function reads the "numa_node" attribute from a PCI device's sysfs
- * directory (e.g. "<pciPath>/numa_node") and returns the NUMA node ID.
- *
- * Linux sysfs convention:
- *   - A non-negative integer indicates the NUMA node the device is local to
- *   - "-1" indicates that the device has no specific NUMA affinity
- *
- * This helper is used to validate NUMA locality when merging NICs, ensuring
- * that merged devices do not silently span NUMA nodes, which could negatively
- * impact performance.
- *
- * @param pciPath  NUL-terminated sysfs PCI device path
- *
- * @return NUMA node ID (>= 0) on success
- * @return -1 if the NUMA node cannot be determined, the file is missing,
- *         unreadable, or contains invalid data
- *
- * Notes:
- * - Uses open/read instead of stdio to avoid buffering and locale issues.
- * - Uses strtol for robust numeric parsing and overflow detection.
- * - A return value of -1 may indicate either "no NUMA affinity" or an error;
- *   callers should treat it as "unknown or unspecified".
- */
-static int ncclIbGetNumaNodeFromPath(const char* pciPath) {
-    if (pciPath == NULL) {
-        return -1;
-    }
-    char numaPath[PATH_MAX];
-    if (snprintf(numaPath, sizeof(numaPath), "%s/numa_node", pciPath) >= PATH_MAX) {
-        return -1; 
-    }
-
-    int fd = open(numaPath, O_RDONLY);
-    if (fd < 0) return -1;
-
-    char buf[32];
-    ssize_t n = read(fd, buf, sizeof(buf) - 1);
-    close(fd);
-
-    if (n <= 0) return -1;
-    buf[n] = '\0';
-    
-    char* endptr;
-    errno = 0;
-    long numa = strtol(buf, &endptr, 10);
-    if (endptr == buf || errno == ERANGE) {
-        return -1;
-    }
-    return (int)numa;
-}
-
-ncclResult_t ncclIbMakeVDeviceInternal(int* d, ncclNetVDeviceProps_t* props) {
-  if (ncclParamIbMergeNics() == 0 && props->ndevs > 1) {
-    WARN("NET/IB : Skipping makeVDevice, Please set NCCL_IB_MERGE_NICS=1");
-    return ncclInvalidUsage;
-  }
-
-  if (props->ndevs == 0) {
-      WARN("NET/IB : Can't make virtual NIC with 0 devices");
-      return ncclInvalidUsage;
-  }
-
-  if (ncclNMergedIbDevs == MAX_IB_VDEVS) {
-    WARN("NET/IB : Cannot allocate any more virtual devices (%d)", MAX_IB_VDEVS);
-    return ncclInvalidUsage;
-  }
-
-  // Always count up number of merged devices
-  ncclIbMergedDev tmp;
-  memset(&tmp,0,sizeof(tmp));
-  bool used[MAX_IB_DEVS] = {0};
-
-  for (int i = 0; i < props->ndevs; i++) {
-    if( props->devs[i]  < 0 || props->devs[i] >= ncclNIbDevs ) {
-      WARN("NET/IB : Cannot use physical device %d, max %d", props->devs[i], ncclNIbDevs);
-      return ncclInvalidUsage;
-    }
-    if(used[props->devs[i]]) continue;
-    const ncclIbDev* dev = ncclIbDevs + props->devs[i];
-    if (tmp.vProps.ndevs == NCCL_IB_MAX_DEVS_PER_NIC) return ncclInvalidUsage;
-    tmp.vProps.devs[tmp.vProps.ndevs++] = props->devs[i];
-    tmp.speed += dev->speed;
-    // Each successive time, copy the name '+' new name
-    if (tmp.vProps.ndevs > 1) {
-      size_t off = strlen(tmp.devName);
-      snprintf(tmp.devName + off, sizeof(tmp.devName) - off, "+%s", dev->devName);
-    // First time, copy the plain name
-    } else {
-      strncpy(tmp.devName, dev->devName, MAXNAMESIZE-1);
-      tmp.devName[MAXNAMESIZE-1] = '\0';
-    }
-    used[props->devs[i]] = true;
-  }
-
-  // Check link layers
-  const ncclIbDev* dev0 = ncclIbDevs + tmp.vProps.devs[0];
-  for (int i = 1; i < tmp.vProps.ndevs; i++) {
-    const ncclIbDev* dev = ncclIbDevs + tmp.vProps.devs[i];
-    if (dev->link != dev0->link) {
-      WARN("NET/IB : Attempted to merge incompatible devices: [%d]%s:%d/%s and [%d]%s:%d/%s. Try selecting NICs of only one link type using NCCL_IB_HCA",
-        tmp.vProps.devs[0], dev0->devName, dev0->portNum, NCCL_IB_LLSTR(dev0->link),tmp.vProps.devs[i], dev->devName, dev->portNum, NCCL_IB_LLSTR(dev->link));
-      return ncclInvalidUsage;
-    }
-  }
-
-  int numa0 = ncclIbGetNumaNodeFromPath(dev0->pciPath);
-  //format -> 0000:00 
-  char root0[8]; 
-  ncclIbGetPciRootFromPath(dev0->pciPath, root0, sizeof(root0));
-  for (int i = 1; i < tmp.vProps.ndevs; i++) {
-    const ncclIbDev* dev = ncclIbDevs + tmp.vProps.devs[i];
-    int numa_i = ncclIbGetNumaNodeFromPath(dev->pciPath);
-    if (numa0 >= 0 && numa_i >= 0 && numa_i != numa0) {
-      WARN("NET/IB : Merging NICs across NUMA nodes (%s numa=%d, %s numa=%d). "
-           "This may significantly reduce performance.",
-           dev0->devName, numa0, dev->devName, numa_i);
-      break;
-    }
-
-    char root_i[8];
-    ncclIbGetPciRootFromPath(dev->pciPath, root_i, sizeof(root_i));
-    if (strcmp(root_i, root0) != 0) {
-      WARN("NET/IB : Merging NICs across PCIe Root Complexes "
-           "(%s root=%s, %s root=%s). "
-           "GPUDirect RDMA and bandwidth aggregation may be impacted.",
-           dev0->devName, root0, dev->devName, root_i);
-      break;
-    }
-  }
-  ncclIbMergedDevs[ncclNMergedIbDevs] = tmp;
-  *d = ncclNMergedIbDevs++;
-  INFO(NCCL_NET, "NET/IB : Made virtual device [%d] name=%s speed=%d ndevs=%d", *d, tmp.devName, tmp.speed, tmp.vProps.ndevs);
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbMakeVDevice(int* d, ncclNetVDeviceProps_t* props) {
-  std::lock_guard<std::mutex> lock(ncclIbMutex);
-  ncclResult_t res = ncclIbMakeVDeviceInternal(d, props);
-  return res;
-
-}
-
-ncclResult_t ncclIbSetNetAttr(void *ctx, ncclNetAttr_t *netAttr) {
-  (void)ctx;
-  (void)netAttr;
-  return ncclSuccess;
-}
-
-static ncclProfilerCallback_t ncclProfilerFunction;
-
-static ncclResult_t ncclIbFinalizeDevices(void) {
-  netRefCount--;
-  return ncclSuccess;
-}
-
-static ncclResult_t ncclIbInitDevices(ncclDebugLogger_t /*logFunction*/, ncclProfilerCallback_t profFunction) {
-  if (netRefCount++) return ncclSuccess;
-  ncclResult_t ret = ncclSuccess;
-  ncclProfilerFunction = profFunction;
-  if (ncclParamIbDisable()) return ncclInternalError;
-  static int shownIbHcaEnv = 0;
-  if(wrap_ibv_symbols() != ncclSuccess) { return ncclInternalError; }
-  if(wrap_mlx5dv_symbols() != ncclSuccess) { INFO(NCCL_NET, "NET/IB : Failed to open mlx5dv symbols. Advance features like CX-8 Direct-NIC will be disabled."); }
-
-  // Detect IB cards
-  int nIbDevs = 0;
-  struct ibv_device** devices = NULL;
-
-  if (ncclNIbDevs == -1) {
-    std::lock_guard<std::mutex> lock(ncclIbMutex);
-    wrap_ibv_fork_init();
-    if (ncclNIbDevs == -1) {
-      int nIpIfs = 0;
-      ncclNIbDevs = 0;
-      ncclNMergedIbDevs = 0;
-      NCCLCHECK(ncclFindInterfaces(ncclIbIfName, &ncclIbIfAddr, MAX_IF_NAME_SIZE, 1, &nIpIfs));
-      if (nIpIfs != 1) {
-        WARN("NET/IB : No IP interface found.");
-        ret = ncclInternalError;
-        goto fail;
-      }   
-
-      // Check if user defined which IB device:port to use
-      const char* userIbEnv = ncclGetEnv("NCCL_IB_HCA");
-      if (userIbEnv != NULL && shownIbHcaEnv++ == 0) INFO(NCCL_NET|NCCL_ENV, "NCCL_IB_HCA set to %s", userIbEnv);
-      struct netIf userIfs[MAX_IB_DEVS];
-      bool searchNot = userIbEnv && userIbEnv[0] == '^';
-      if (searchNot) userIbEnv++;
-      bool searchExact = userIbEnv && userIbEnv[0] == '=';
-      if (searchExact) userIbEnv++;
-      int nUserIfs = parseStringList(userIbEnv, userIfs, MAX_IB_DEVS);
-
-      if (ncclSuccess != wrap_ibv_get_device_list(&devices, &nIbDevs)) { ret = ncclInternalError; goto fail; }
-
-      for (int d=0; d<nIbDevs && ncclNIbDevs<MAX_IB_DEVS; d++) {
-        struct ibv_context * context = NULL;
-        if (ncclSuccess != wrap_ibv_open_device(&context, devices[d]) || context == NULL) {
-          WARN("NET/IB : Unable to open device %s", devices[d]->name);
-          continue;
-        }
-        char dataDirectDevicePath[PATH_MAX] = "/sys";
-        int devCount = /*undefined*/-1, devOffset = 0;
-        enum ncclIbProvider ibProvider = wrap_mlx5dv_is_supported(devices[d]) ? IB_PROVIDER_MLX5 : IB_PROVIDER_NONE;
-
-        int nPorts = 0;
-        struct ibv_device_attr devAttr;
-        memset(&devAttr, 0, sizeof(devAttr));
-        if (ncclSuccess != wrap_ibv_query_device(context, &devAttr)) {
-          WARN("NET/IB : Unable to query device %s", devices[d]->name);
-          if (ncclSuccess != wrap_ibv_close_device(context))
-          {
-            ret = ncclInternalError;
-            goto fail;
-          }
-          continue;
-        }
-        for (int port_num = 1; port_num <= devAttr.phys_port_cnt; port_num++) {
-            struct ibv_port_attr portAttr;
-            if (ncclSuccess != wrap_ibv_query_port(context, port_num, &portAttr)) {
-              WARN("NET/IB : Unable to query port_num %d", port_num);
-              continue;
-            }
-            if (portAttr.state != IBV_PORT_ACTIVE) continue;
-            if (portAttr.link_layer != IBV_LINK_LAYER_INFINIBAND && portAttr.link_layer != IBV_LINK_LAYER_ETHERNET) continue;
-
-            // check against user specified HCAs/ports
-            if (! (matchIfList(devices[d]->name, port_num, userIfs, nUserIfs, searchExact) ^ searchNot)) {
-              continue;
-            }
-
-            // check for mlx5 data direct support only once for a each device
-            if (devCount == -1) {
-              devCount = 1;
-              devOffset = 0;
-              if (ncclParamIbDataDirect() > 0 && ibProvider == IB_PROVIDER_MLX5 && ncclMlx5dvDmaBufCapable(context)) {
-                int pathLen = strlen(dataDirectDevicePath);
-                ncclResult_t res = wrap_mlx5dv_get_data_direct_sysfs_path(context, dataDirectDevicePath + pathLen, sizeof(dataDirectDevicePath) - pathLen);
-                if (res == ncclSuccess) {
-                  // data direct devices are exposed twice: with the C2C + PCIe link and with the data direct link
-                  devCount = 2;
-                  // by default only expose the data direct NIC (devOffset = 1), unless set to 2 by the user
-                  devOffset = (ncclParamIbDataDirect() == 2) ? 0 : 1;
-                  INFO(NCCL_INIT | NCCL_NET, "NET/IB: Data Direct DMA Interface is detected for device %s", devices[d]->name);
-                } else if (res == ncclInvalidArgument) {
-                  TRACE(NCCL_NET, "NET/IB: Device %s does not support Data Direct DMA.", devices[d]->name);
-                } else {
-                  WARN("NET/IB: Error in mlx5dv_get_data_direct_sysfs_path with device %s", devices[d]->name);
-                  return res;
-                }
-              }
-            }
-            for (int dev = devOffset; dev < devCount; ++dev) {
-              ncclIbDevs[ncclNIbDevs].device = d;
-              ncclIbDevs[ncclNIbDevs].ibProvider = ibProvider;
-              ncclIbDevs[ncclNIbDevs].guid = devAttr.sys_image_guid;
-              ncclIbDevs[ncclNIbDevs].portAttr = portAttr;
-              ncclIbDevs[ncclNIbDevs].portNum = port_num;
-              ncclIbDevs[ncclNIbDevs].link = portAttr.link_layer;
-              if (portAttr.active_speed_ex) {
-                // A non-zero active_speed_ex indicates XDR rate (0x100) or higher
-                ncclIbDevs[ncclNIbDevs].speed = ncclIbSpeed(portAttr.active_speed_ex) * ncclIbWidth(portAttr.active_width);
-              } else {
-                ncclIbDevs[ncclNIbDevs].speed = ncclIbSpeed(portAttr.active_speed) * ncclIbWidth(portAttr.active_width);
-              }
-              ncclIbDevs[ncclNIbDevs].context = context;
-              ncclIbDevs[ncclNIbDevs].pdRefs = 0;
-              ncclIbDevs[ncclNIbDevs].pd = NULL;
-              if (dev == 0) {
-                strncpy(ncclIbDevs[ncclNIbDevs].devName, devices[d]->name, MAXNAMESIZE);
-                NCCLCHECKGOTO(ncclIbGetPciPath(ncclIbDevs[ncclNIbDevs].devName, &ncclIbDevs[ncclNIbDevs].pciPath, &ncclIbDevs[ncclNIbDevs].realPort), ret, fail);
-              } else {
-                snprintf(ncclIbDevs[ncclNIbDevs].devName, MAXNAMESIZE, "%s_dma", devices[d]->name);
-                NCCLCHECK(ncclCalloc(&ncclIbDevs[ncclNIbDevs].pciPath, PATH_MAX));
-                strncpy(ncclIbDevs[ncclNIbDevs].pciPath, dataDirectDevicePath, PATH_MAX);
-                ncclIbDevs[ncclNIbDevs].capsProvider.mlx5.dataDirect = 1;
-              }
-              ncclIbDevs[ncclNIbDevs].maxQp = devAttr.max_qp;
-              ncclIbDevs[ncclNIbDevs].mrCache.capacity = 0;
-              ncclIbDevs[ncclNIbDevs].mrCache.population = 0;
-              ncclIbDevs[ncclNIbDevs].mrCache.slots = NULL;
-              NCCLCHECK(ncclIbStatsInit(&ncclIbDevs[ncclNIbDevs].stats));
-
-              // Enable ADAPTIVE_ROUTING by default on IB networks
-              // But allow it to be overloaded by an env parameter
-              ncclIbDevs[ncclNIbDevs].ar = (portAttr.link_layer == IBV_LINK_LAYER_INFINIBAND) ? 1 : 0;
-              if (ncclParamIbAdaptiveRouting() != -2) ncclIbDevs[ncclNIbDevs].ar = ncclParamIbAdaptiveRouting();
-
-              INFO(NCCL_NET, "NET/IB: [%d] %s:%s:%d/%s provider=%s speed=%d context=%p pciPath=%s ar=%d", d, devices[d]->name, devices[d]->dev_name,
-                   ncclIbDevs[ncclNIbDevs].portNum, NCCL_IB_LLSTR(portAttr.link_layer), ibProviderName[ncclIbDevs[ncclNIbDevs].ibProvider], ncclIbDevs[ncclNIbDevs].speed, context,
-                   ncclIbDevs[ncclNIbDevs].pciPath, ncclIbDevs[ncclNIbDevs].ar);
-
-              PTHREADCHECKGOTO(pthread_create(&ncclIbAsyncThread, NULL, ncclIbAsyncThreadMain, ncclIbDevs + ncclNIbDevs), "pthread_create", ret, fail);
-              ncclSetThreadName(ncclIbAsyncThread, "NCCL IbAsync %2d", ncclNIbDevs);
-              PTHREADCHECKGOTO(pthread_detach(ncclIbAsyncThread), "pthread_detach", ret, fail); // will not be pthread_join()'d
-
-              ncclNIbDevs++;
-              nPorts++;
-            }
-        }
-        if (nPorts == 0 && ncclSuccess != wrap_ibv_close_device(context)) { ret = ncclInternalError; goto fail; }
-      }
-
-      if (ncclSuccess != wrap_ibv_free_device_list(devices)) { ret = ncclInternalError; goto fail; }
-    }
-    if (ncclNIbDevs == 0) {
-      INFO(NCCL_INIT|NCCL_NET, "NET/IB : No device found.");
-    }
-
-    // Print out all net devices to the user (in the same format as before)
-    char line[2048];
-    line[0] = '\0';
-    // Determine whether RELAXED_ORDERING is enabled and possible
-    ncclIbRelaxedOrderingEnabled = ncclIbRelaxedOrderingCapable();
-    for (int d = 0; d < ncclNIbDevs; d++) {
-      snprintf(line+strlen(line), sizeof(line)-strlen(line), " [%d]%s:%d/%s", d, ncclIbDevs[d].devName,
-        ncclIbDevs[d].portNum, NCCL_IB_LLSTR(ncclIbDevs[d].link));
-      // Add this plain physical device to the list of virtual devices
-      int vDev;
-      ncclNetVDeviceProps_t vProps = {0};
-      vProps.ndevs = 1;
-      vProps.devs[0] = d;
-      NCCLCHECK(ncclIbMakeVDeviceInternal(&vDev, &vProps)); 
-    }
-    char addrline[SOCKET_NAME_MAXLEN+1];
-    INFO(NCCL_INIT|NCCL_NET, "NET/IB : Using%s %s; OOB %s:%s", line, ncclIbRelaxedOrderingEnabled ? "[RO]" : "",
-          ncclIbIfName, ncclSocketToString(&ncclIbIfAddr, addrline));
-
-  }
-exit:
-  return ret;
-fail:
-  if(ncclSuccess != wrap_ibv_free_device_list(devices)){WARN("NET/IB : Unable to free device list");}
-  goto exit;
-}
-
-ncclResult_t ncclIbInit(void** ctx, uint64_t commId, ncclNetCommConfig_t* config, ncclDebugLogger_t logFunction, ncclProfilerCallback_t profFunction) {
-  ncclResult_t ret = ncclSuccess;
-  ncclNetCommConfig_t* netCommConfig = nullptr;
-  NCCLCHECK(ncclIbInitDevices(logFunction, profFunction));
-  NCCLCHECK(ncclCalloc(&netCommConfig, 1));
-  netCommConfig->trafficClass = config->trafficClass;
-  *ctx = (void *)netCommConfig;
-  return ret;
-}
-
-ncclResult_t ncclIbDevices(int* ndev) {
-  *ndev = ncclNMergedIbDevs;
-  return ncclSuccess;
-}
-
-// Introduce RCCL_FORCE_ENABLE_GDRDMA to force load GPU-NIC RDMA module
-// Use ONLY for debugging!
-RCCL_PARAM(ForceEnableGdrdma, "FORCE_ENABLE_GDRDMA", -1);
-
-// Detect whether GDR can work on a given NIC with the current CUDA device
-// Returns :
-// ncclSuccess : GDR works
-// ncclSystemError : no module or module loaded but not supported by GPU
-#define KNL_MODULE_LOADED(a) ((access(a, F_OK) == -1) ? 0 : 1)
-static int ncclIbGdrModuleLoaded = 0; // 1 = true, 0 = false
-static void ibGdrSupportInitOnce() {
-#if defined(__HIP_PLATFORM_AMD__) || defined(__HIPCC__)
-  if (rcclParamForceEnableGdrdma() == 1) {
-    // RCCL_FORCE_ENABLE_GDRDMA=1 enables GPU-NIC RDMA only from RCCL-side
-    // Requires support from NIC driver modules
-    // Use ONLY for debugging!
-    ncclIbGdrModuleLoaded = 1;
-    INFO(NCCL_INIT, "RCCL_FORCE_ENABLE_GDRDMA = 1, so explicitly setting ncclIbGdrModuleLoaded = 1");
-  }
-
-  if (ncclIbGdrModuleLoaded == 0) {
-    // Check for `memory_peers` directory containing `amdkfd/version`
-    // This `memory_peers` directory is created by NIC-GPU driver interaction
-    // On Linux kernel 5.15.0 (e.g. Ubuntu 22.04), `memory_peers` is created under `/sys/kernel/mm/`
-    // However, on newer kernels like Ubuntu 24.04.1 (Linux kernel 6.8.0) or Ubuntu 22.04.4 HWE (Linux kernel 6.5.0),
-    // this `memory_peers` directory is either not created (go to else-if condition)
-    // or created under a different path like `/sys/kernel/` or `/sys/` (depending on your ib_peer_mem module)
-    const char* memory_peers_paths[] = {"/sys/kernel/mm/memory_peers/amdkfd/version",
-                                  "/sys/kernel/memory_peers/amdkfd/version",
-                                  "/sys/memory_peers/amdkfd/version",
-                                  NULL};
-    int i = 0;
-
-    while (memory_peers_paths[i]) {
-      if (access(memory_peers_paths[i], F_OK) == 0) {
-        ncclIbGdrModuleLoaded = 1;
-        INFO(NCCL_INIT,"Found %s", memory_peers_paths[i]);
-        break;
-      } else {
-        ncclIbGdrModuleLoaded = 0;
-      }
-      ++i;
-    }
-
-    char strValue[MAX_STR_LEN];
-    ncclTopoGetStrFromSys("/sys/devices/virtual/dmi/id", "bios_version", strValue);
-    if (strncmp("Hyper-V UEFI Release", strValue, 20) == 0) {
-      int roMode = ncclParamIbPciRelaxedOrdering();
-      ncclTopoGetStrFromSys("/proc/sys/kernel", "numa_balancing", strValue);
-      if (strcmp(strValue, "1") == 0 && roMode == 0)
-        ncclIbGdrModuleLoaded = 0;
-    }
-
-    if (ncclIbGdrModuleLoaded == 0) {
-      // Check for `ib_register_peer_memory_client` symbol in `/proc/kallsyms`
-      // if your system uses native OS ib_peer module
-      char buf[256];
-      FILE *fp = NULL;
-      fp = fopen("/proc/kallsyms", "r");
-
-      if (fp == NULL) {
-        INFO(NCCL_INIT,"Could not open /proc/kallsyms");
-      } else {
-        while (fgets(buf, sizeof(buf), fp) != NULL) {
-          if (strstr(buf, "t ib_register_peer_memory_client") != NULL ||
-              strstr(buf, "T ib_register_peer_memory_client") != NULL) {
-            ncclIbGdrModuleLoaded = 1;
-            INFO(NCCL_INIT,"Found ib_register_peer_memory_client in /proc/kallsyms");
-            break;
-          }
-        }
-      }
-    }
-  }
-#else
-  // Check for the nv_peer_mem module being loaded
-  ncclIbGdrModuleLoaded = KNL_MODULE_LOADED("/sys/kernel/mm/memory_peers/nv_mem/version") ||
-                          KNL_MODULE_LOADED("/sys/kernel/mm/memory_peers/nv_mem_nc/version") ||
-                          KNL_MODULE_LOADED("/sys/module/nvidia_peermem/version");
-#endif
-}
-
-ncclResult_t ncclIbGdrSupport() {
-  static std::once_flag once;
-  std::call_once(once, ibGdrSupportInitOnce);
-  if (!ncclIbGdrModuleLoaded)
-    return ncclSystemError;
-  return ncclSuccess;
-}
-
-static __thread int ibDmaSupportInitDev; // which device to init, must be thread local
-static void ibDmaBufSupportInitOnce(){
-  ncclResult_t res;
-  int dev_fail = 0;
-
-  // This is a physical device, not a virtual one, so select from ibDevs
-  ncclIbMergedDev* mergedDev = ncclIbMergedDevs + ibDmaSupportInitDev;
-  ncclIbDev* ibDev = ncclIbDevs + mergedDev->vProps.devs[0];
-  struct ibv_pd* pd;
-  struct ibv_context* ctx = ibDev->context;
-  res = rocmLibraryInit();
-  if (res != ncclSuccess) goto failure;
-  NCCLCHECKGOTO(wrap_ibv_alloc_pd(&pd, ctx), res, failure);
-  // Test kernel DMA-BUF support with a dummy call (fd=-1)
-  (void)wrap_direct_ibv_reg_dmabuf_mr(pd, 0ULL /*offset*/, 0ULL /*len*/, 0ULL /*iova*/, -1 /*fd*/, 0 /*flags*/);
-  // ibv_reg_dmabuf_mr() will fail with EOPNOTSUPP/EPROTONOSUPPORT if not supported (EBADF otherwise)
-  dev_fail |= (errno == EOPNOTSUPP) || (errno == EPROTONOSUPPORT);
-  NCCLCHECKGOTO(wrap_ibv_dealloc_pd(pd), res, failure);
-  // stop the search and goto failure
-  if (dev_fail) goto failure;
-  ibDev->dmaBufSupported = 1;
-  return;
-failure:
-  ibDev->dmaBufSupported = -1;
-  return;
-}
-// Detect whether DMA-BUF support is present in the kernel
-// Returns :
-// ncclSuccess : DMA-BUF support is available
-// ncclSystemError : DMA-BUF is not supported by the kernel
-ncclResult_t ncclIbDmaBufSupport(int dev) {
-  static std::once_flag onces[MAX_IB_DEVS];
-  // init the device only once
-  ibDmaSupportInitDev = dev;
-  std::call_once(onces[dev], ibDmaBufSupportInitOnce);
-  ncclIbMergedDev* mergedDev = ncclIbMergedDevs + ibDmaSupportInitDev;
-  ncclIbDev* ibDev = ncclIbDevs + mergedDev->vProps.devs[0];
-  int dmaBufSupported = ibDev->dmaBufSupported;
-  if (dmaBufSupported == 1) return ncclSuccess;
-  return ncclSystemError;
-}
-
-#define NCCL_NET_IB_MAX_RECVS 8
-
-ncclResult_t ncclIbGetPhysProperties(int dev, ncclNetProperties_t* props) {
-  struct ncclIbDev* ibDev = ncclIbDevs + dev;
-  std::lock_guard<std::mutex> lock(ibDev->mutex);
-  props->name = ibDev->devName;
-  props->speed = ibDev->speed;
-  props->pciPath = ibDev->pciPath;
-  props->guid = ibDev->guid;
-  props->ptrSupport = NCCL_PTR_HOST;
-  if (ncclIbGdrSupport() == ncclSuccess) {
-    props->ptrSupport |= NCCL_PTR_CUDA; // GDR support via nv_peermem
-  }
-  props->regIsGlobal = 1;
-  if (ncclIbDmaBufSupport(dev) == ncclSuccess) {
-    props->ptrSupport |= NCCL_PTR_DMABUF; // GDR support via DMA-BUF
-  }
-  props->forceFlush = 0;
-  if (ibDev->capsProvider.mlx5.dataDirect) {
-    props->forceFlush = 1;
-  }
-  props->latency = 0; // Not set
-  props->port = ibDev->portNum + ibDev->realPort;
-  props->maxComms = ibDev->maxQp;
-  props->maxRecvs = NCCL_NET_IB_MAX_RECVS;
-  props->netDeviceType    = NCCL_NET_DEVICE_HOST;
-  props->netDeviceVersion = NCCL_NET_DEVICE_INVALID_VERSION;
-  props->maxP2pBytes = NCCL_MAX_NET_SIZE_BYTES;
-  props->maxCollBytes = MAX_COLLNET_SIZE;
-  props->maxMultiRequestSize = 1;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbGetProperties(int dev, ncclNetProperties_t* props) {
-  if (dev >= ncclNMergedIbDevs) {
-    WARN("NET/IB : Requested properties for vNic %d, only %d vNics have been created", dev, ncclNMergedIbDevs);
-    return ncclInvalidUsage;
-  }
-  struct ncclIbMergedDev* mergedDev = ncclIbMergedDevs + dev;
-  // Take the rest of the properties from an arbitrary sub-device (should be the same)
-  NCCLCHECK(ncclIbGetPhysProperties(mergedDev->vProps.devs[0], props));
-  props->name = mergedDev->devName;
-  props->speed = mergedDev->speed;
-  memcpy(&props->vProps, &mergedDev->vProps, sizeof(ncclNetVDeviceProps_t));
-  return ncclSuccess;
-}
-
-// We need to support NCCL_NET_MAX_REQUESTS for each concurrent receive
-#define MAX_REQUESTS (NCCL_NET_MAX_REQUESTS*NCCL_NET_IB_MAX_RECVS)
-static_assert(MAX_REQUESTS <= 256, "request id are encoded in wr_id and we need up to 8 requests ids per completion");
-
-#include "net_ib_limits.h"
-
-// Per-QP connection metatdata
-struct ncclIbQpInfo {
-  uint32_t qpn;
-
-  // Fields needed for ece (enhanced connection establishment)
-  struct ibv_ece ece;
-  int ece_supported;
-  int devIndex;
-};
-
-// Per-Dev connection metadata
-struct ncclIbDevInfo {
-  uint32_t lid;
-  uint8_t ib_port;
-  enum ibv_mtu mtu;
-  uint8_t link_layer;
-
-  // For RoCE and IB Rounter
-  union ibv_gid gid;
-
-  // FIFO RDMA info
-  uint32_t fifoRkey;
-
-  //remote dev info
-  union ibv_gid remoteGid;
-
-  int ibv_dev_index;
-};
-
-// Struct containing everything needed to establish connections
-struct ncclIbConnectionMetadata {
-  struct ncclIbQpInfo qpInfo[NCCL_IB_MAX_QPS];
-  struct ncclIbDevInfo devs[NCCL_IB_MAX_DEVS_PER_NIC];
-  char devName[MAX_MERGED_DEV_NAME];
-  uint64_t fifoAddr;
-  int ndevs;
-  int tc;
-  int sl;
-  int isP2p;
-};
-
-enum ncclIbCommState {
-  ncclIbCommStateStart = 0,
-  ncclIbCommStateConnect = 1,
-  ncclIbCommStateAccept = 3,
-  ncclIbCommStateSend = 4,
-  ncclIbCommStateRecv = 5,
-  ncclIbCommStateConnecting = 6,
-  ncclIbCommStateConnected = 7,
-  ncclIbCommStatePendingReady = 8,
-  ncclIbCommStateSendDevList = 9,
-  ncclIbCommStateRecvDevList = 10,
-};
-
-struct ncclIbCommStage {
-  enum ncclIbCommState state;
-  int offset;
-  void* buffer;
-  void* comm;
-};
-
-struct ncclIbHandle {
-  union ncclSocketAddress connectAddr; // Filled by the target
-  uint64_t magic; // random number to help debugging
-  int isP2p; // P2P flag
-  struct ncclIbCommStage stage; // Used by the other side when connecting
-};
-
-// Retain local RoCE address for error logging
-struct ncclIbGidInfo {
-  uint8_t link_layer;
-  union ibv_gid localGid;
-  int32_t localGidIndex;
-};
-
-#define NCCL_NET_IB_REQ_UNUSED 0
-#define NCCL_NET_IB_REQ_SEND 1
-#define NCCL_NET_IB_REQ_RECV 2
-#define NCCL_NET_IB_REQ_FLUSH 3
-#define NCCL_NET_IB_REQ_GIN_IPUT 4
-#define NCCL_NET_IB_REQ_FAILED 5
-const char* reqTypeStr[] = { "Unused", "Send", "Recv", "Flush", "IPut", "Failed" };
-
-#define MAX_QPS_PER_REQ 8
-struct ncclProfilerInfo {
-  void* qpEventHandles[MAX_QPS_PER_REQ];
-  int qpIndex[MAX_QPS_PER_REQ];
-  int nEventHandles;
-  ncclProfilerNetIbDescr_v1_t data;
-  void* pHandle;
-};
-
-struct ncclIbRequest {
-  struct ncclIbNetCommBase* base;
-  int type;
-  struct ncclSocket* sock;
-  int events[NCCL_IB_MAX_DEVS_PER_NIC];
-  struct ncclIbNetCommDevBase* devBases[NCCL_IB_MAX_DEVS_PER_NIC];
-#ifdef NCCL_ENABLE_NET_PROFILING
-  struct ncclProfilerInfo pInfo[NCCL_NET_IB_MAX_RECVS];
-#endif
-  int nreqs;
-  union {
-    struct {
-      int size;
-      void* data;
-      uint32_t lkeys[NCCL_IB_MAX_DEVS_PER_NIC];
-      int offset;
-    } send;
-    struct {
-      int* sizes;
-    } recv;
-    struct {
-      int rank;
-    } iput;
-  };
-};
-
-// Check if request has any pending events (IBV operations in flight)
-static inline bool ncclIbRequestHasEvents(struct ncclIbRequest* r) {
-  for (int i = 0; i < NCCL_IB_MAX_DEVS_PER_NIC; i++) {
-    if (r->events[i] != 0) return true;
-  }
-  return false;
-}
-
-struct ncclIbNetCommDevBase {
-  int ibDevN;
-  struct ibv_pd* pd;
-  struct ibv_cq* cq;
-  uint64_t pad[2];
-  struct ncclIbGidInfo gidInfo;
-};
-
-struct ncclIbListenComm {
-  int dev;
-  struct ncclSocket sock;
-  struct ncclIbCommStage stage;
-};
-
-struct alignas(64) ncclIbSendFifo {
-  uint64_t addr;
-  uint64_t size;
-  uint32_t rkeys[NCCL_IB_MAX_DEVS_PER_NIC];
-  uint32_t nreqs;
-  uint32_t tag;
-  uint64_t idx;
-  char padding[16];
-};
-
-struct ncclIbQp {
-  struct ibv_qp* qp;
-  int devIndex;
-  int remDevIdx;
-};
-
-struct ncclIbRemSizesFifo {
-  int elems[MAX_REQUESTS][NCCL_NET_IB_MAX_RECVS];
-  uint64_t fifoTail;
-  uint64_t addr;
-  uint32_t rkeys[NCCL_IB_MAX_DEVS_PER_NIC];
-  uint32_t flags;
-  struct ibv_mr* mrs[NCCL_IB_MAX_DEVS_PER_NIC];
-  struct ibv_sge sge;
-};
-
-// A per-dev struct for netIbSendComm
-struct alignas(8) ncclIbSendCommDev {
-  struct ncclIbNetCommDevBase base;
-  struct ibv_mr* fifoMr;
-  struct ibv_mr* putSignalScratchpadMr;
-};
-
-
-// Wrapper to track an MR per-device, if needed
-struct ncclIbMrHandle {
-  ibv_mr* mrs[NCCL_IB_MAX_DEVS_PER_NIC];
-};
-
-struct alignas(32) ncclIbNetCommBase {
-  ncclNetVDeviceProps_t vProps;
-  bool isSend;
-  struct ncclIbRequest reqs[MAX_REQUESTS];
-  struct ncclIbQp qps[NCCL_IB_MAX_QPS];
-  int nqps;
-  int qpIndex;
-  int devIndex;
-  struct ncclSocket sock;
-  int ready;
-  // Track necessary remDevInfo here
-  int nRemDevs;
-  int nDataQps;
-  struct ncclIbDevInfo remDevs[NCCL_IB_MAX_DEVS_PER_NIC];
-  // statistics about the comm
-  struct ncclIbStats stats;
-};
-
-struct ncclIbSendComm {
-  struct ncclIbNetCommBase base;
-  // Start with fifo and ibv structs as they have alignment restrictions
-  struct ncclIbSendFifo fifo[MAX_REQUESTS][NCCL_NET_IB_MAX_RECVS];
-  struct ibv_sge sges[NCCL_NET_IB_MAX_RECVS];
-  struct ibv_send_wr wrs[NCCL_NET_IB_MAX_RECVS + 1];
-  // Each dev correlates to a mergedIbDev
-  struct ncclIbSendCommDev devs[NCCL_IB_MAX_DEVS_PER_NIC];
-  struct ncclIbRequest* fifoReqs[MAX_REQUESTS][NCCL_NET_IB_MAX_RECVS];
-  struct ncclIbRemSizesFifo remSizesFifo;
-  uint64_t fifoHead;
-  int ar; // Use adaptive routing when all merged devices have it enabled
-  uint64_t putSignalScratchpad;
-};
-// The SendFifo needs to be 32-byte aligned and each element needs
-// to be a 32-byte multiple, so that an entry does not get split and
-// written out of order when IB Relaxed Ordering is enabled
-static_assert((sizeof(struct ncclIbNetCommBase) % 32) == 0, "ncclIbNetCommBase size must be 32-byte multiple to ensure fifo is at proper offset");
-static_assert((offsetof(struct ncclIbSendComm, fifo) % 32) == 0, "ncclIbSendComm fifo must be 32-byte aligned");
-static_assert((sizeof(struct ncclIbSendFifo) % 32) == 0, "ncclIbSendFifo element size must be 32-byte multiples");
-static_assert((offsetof(struct ncclIbSendComm, sges) % 32) == 0, "sges must be 32-byte aligned");
-static_assert((offsetof(struct ncclIbSendComm, wrs) % 32) == 0, "wrs must be 32-byte aligned");
-
-struct ncclIbGpuFlush {
-  struct ibv_mr* hostMr;
-  struct ibv_mr* gpuMr;
-  int* gpuFlushGpuMem;
-  struct ibv_sge sge;
-  struct ncclIbQp qp;
-  int dmabuf_fd;
-};
-
-struct ncclIbRemFifo {
-  struct ncclIbSendFifo elems[MAX_REQUESTS][NCCL_NET_IB_MAX_RECVS];
-  uint64_t fifoTail;
-  uint64_t addr;
-  uint32_t flags;
-};
-
-struct alignas(16) ncclIbRecvCommDev {
-  struct ncclIbNetCommDevBase base;
-  struct ncclIbGpuFlush gpuFlush;
-  struct ibv_mr* fifoMr;
-  struct ibv_sge fifoSge;
-  struct ibv_mr* sizesFifoMr;
-};
-
-struct ncclIbRecvComm {
-  struct ncclIbNetCommBase base;
-  struct ncclIbRecvCommDev devs[NCCL_IB_MAX_DEVS_PER_NIC];
-  struct ncclIbRemFifo remFifo;
-  int sizesFifo[MAX_REQUESTS][NCCL_NET_IB_MAX_RECVS];
-  int gpuFlushHostMem;
-  int flushEnabled;
-};
-static_assert((offsetof(struct ncclIbRecvComm, remFifo) % 32) == 0, "ncclIbRecvComm fifo must be 32-byte aligned");
-
-static void ncclIbAddEvent(struct ncclIbRequest* req, int devIndex, struct ncclIbNetCommDevBase* base) {
-  req->events[devIndex]++;
-  req->devBases[devIndex] = base;
-}
-ncclResult_t ncclIbInitCommDevBase(int ibDevN, struct ncclIbNetCommDevBase* base, void* cq_context) {
-  base->ibDevN = ibDevN;
-  ncclIbDev* ibDev = ncclIbDevs + ibDevN;
-  {
-    std::lock_guard<std::mutex> lock(ibDev->mutex);
-    if (0 == ibDev->pdRefs++) {
-      NCCLCHECK(wrap_ibv_alloc_pd(&ibDev->pd, ibDev->context));
-    }
-    base->pd = ibDev->pd;
-  }
-
-  // CQ is sized to accommodate the max SQ + RQ WQE completions. If each SQ WQE could be signaled, then,
-  // for each QP, there can be 2*MAX_REQUESTS completions for SQ and MAX_REQUESTS completions for RQ. 
-  NCCLCHECK(wrap_ibv_create_cq(&base->cq, ibDev->context, 3*MAX_REQUESTS*ncclParamIbQpsPerConn(), cq_context, NULL, 0));
-
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbDestroyBase(struct ncclIbNetCommDevBase* base) {
-  NCCLCHECK(wrap_ibv_destroy_cq(base->cq));
-
-  std::lock_guard<std::mutex> lock(ncclIbDevs[base->ibDevN].mutex);
-  if (0 == --ncclIbDevs[base->ibDevN].pdRefs) {
-    NCCLCHECK(wrap_ibv_dealloc_pd(ncclIbDevs[base->ibDevN].pd));
-  }
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbCreateQp(uint8_t ib_port, struct ncclIbNetCommDevBase* base, int access_flags, void* qp_context, struct ncclIbQp* qp) {
-  struct ibv_qp_init_attr qpInitAttr;
-  memset(&qpInitAttr, 0, sizeof(struct ibv_qp_init_attr));
-  qpInitAttr.qp_context = qp_context;
-  qpInitAttr.send_cq = base->cq;
-  qpInitAttr.recv_cq = base->cq;
-  qpInitAttr.qp_type = IBV_QPT_RC;
-  // We might send 2 messages per send (RDMA and RDMA_WITH_IMM)
-  qpInitAttr.cap.max_send_wr = 2*MAX_REQUESTS;
-  qpInitAttr.cap.max_recv_wr = MAX_REQUESTS;
-  qpInitAttr.cap.max_send_sge = 1;
-  qpInitAttr.cap.max_recv_sge = 1;
-  qpInitAttr.cap.max_inline_data = ncclParamIbUseInline() ? sizeof(struct ncclIbSendFifo) : 0;
-  NCCLCHECK(wrap_ibv_create_qp(&qp->qp, base->pd, &qpInitAttr));
-  struct ibv_qp_attr qpAttr;
-  memset(&qpAttr, 0, sizeof(struct ibv_qp_attr));
-  qpAttr.qp_state = IBV_QPS_INIT;
-  qpAttr.pkey_index = ncclParamIbPkey();
-  qpAttr.port_num = ib_port;
-  qpAttr.qp_access_flags = access_flags;
-  NCCLCHECK(wrap_ibv_modify_qp(qp->qp, &qpAttr, IBV_QP_STATE | IBV_QP_PKEY_INDEX | IBV_QP_PORT | IBV_QP_ACCESS_FLAGS));
-  TRACE(NCCL_NET, "NET/IB : ncclIbCreateQp port=%d dev=%d devName=%s ndevs=%d nmdevs=%d qpn=%u pkey=%u pd=%p",
-    ib_port, base->ibDevN, ncclIbDevs[base->ibDevN].devName, ncclNIbDevs, ncclNMergedIbDevs, qp->qp->qp_num, qpAttr.pkey_index, base->pd);
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbRtrQp(struct ibv_qp* qp, struct ncclIbGidInfo* sGidInfo, uint32_t dest_qp_num, struct ncclIbDevInfo* info, bool fifoTc, int tc, int sl) {
-  struct ibv_qp_attr qpAttr;
-  memset(&qpAttr, 0, sizeof(struct ibv_qp_attr));
-  qpAttr.qp_state = IBV_QPS_RTR;
-  qpAttr.path_mtu = info->mtu;
-  qpAttr.dest_qp_num = dest_qp_num;
-  qpAttr.rq_psn = 0;
-  qpAttr.max_dest_rd_atomic = 1;
-  qpAttr.min_rnr_timer = 12;
-  if (info->link_layer == IBV_LINK_LAYER_ETHERNET) {
-    qpAttr.ah_attr.is_global = 1;
-    qpAttr.ah_attr.grh.dgid.global.subnet_prefix = info->gid.global.subnet_prefix;
-    qpAttr.ah_attr.grh.dgid.global.interface_id = info->gid.global.interface_id;
-    qpAttr.ah_attr.grh.flow_label = 0;
-    qpAttr.ah_attr.grh.sgid_index = sGidInfo->localGidIndex;
-    qpAttr.ah_attr.grh.hop_limit = 255;
-    qpAttr.ah_attr.grh.traffic_class = fifoTc && ncclParamIbFifoTc() != -1 ? ncclParamIbFifoTc() : tc;
-  } else {
-    //pick lid if subnet prefixs are same, FLID if they are not
-    if (ncclIbExtractLocalSubnetPrefix(sGidInfo->localGid.global.subnet_prefix) ==
-        ncclIbExtractLocalSubnetPrefix(info->gid.global.subnet_prefix)) {
-        qpAttr.ah_attr.is_global = 0;
-        qpAttr.ah_attr.dlid = info->lid;
-    } else {
-  uint16_t flid = ncclIbExtractFlid(&info->gid);
-        if (flid == 0) {
-          WARN("Warning: remote FLID configured as zero even when endpoints are on different subnets, using dlid as fallback");
-          qpAttr.ah_attr.dlid = info->lid;
-  } else {
-          qpAttr.ah_attr.dlid = ncclIbExtractFlid(&info->gid);
-  }
-        qpAttr.ah_attr.is_global = 1;
-        qpAttr.ah_attr.grh.dgid.global.subnet_prefix = info->gid.global.subnet_prefix;
-        qpAttr.ah_attr.grh.dgid.global.interface_id = info->gid.global.interface_id;
-        qpAttr.ah_attr.grh.sgid_index = sGidInfo->localGidIndex;
-  qpAttr.ah_attr.grh.hop_limit = 255;
-    }
-  }
-  qpAttr.ah_attr.sl = sl;
-  qpAttr.ah_attr.src_path_bits = 0;
-  qpAttr.ah_attr.port_num = info->ib_port;
-  TRACE(NCCL_NET, "NET/IB : ncclIbRtrQp qpn=%u mtu=%d dst=%u ll=%u port=%u sl: %d tc: %d", qp->qp_num, info->mtu, dest_qp_num, info->link_layer, info->ib_port, qpAttr.ah_attr.sl, qpAttr.ah_attr.grh.traffic_class);
-  NCCLCHECK(wrap_ibv_modify_qp(qp, &qpAttr, IBV_QP_STATE | IBV_QP_AV | IBV_QP_PATH_MTU | IBV_QP_DEST_QPN | IBV_QP_RQ_PSN | IBV_QP_MAX_DEST_RD_ATOMIC | IBV_QP_MIN_RNR_TIMER));
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbRtsQp(struct ibv_qp* qp) {
-  struct ibv_qp_attr qpAttr;
-  memset(&qpAttr, 0, sizeof(struct ibv_qp_attr));
-  qpAttr.qp_state = IBV_QPS_RTS;
-  qpAttr.timeout = ncclParamIbTimeout();
-  qpAttr.retry_cnt = ncclParamIbRetryCnt();
-  qpAttr.rnr_retry = 7;
-  qpAttr.sq_psn = 0;
-  qpAttr.max_rd_atomic = 1;
-  NCCLCHECK(wrap_ibv_modify_qp(qp, &qpAttr, IBV_QP_STATE | IBV_QP_TIMEOUT | IBV_QP_RETRY_CNT | IBV_QP_RNR_RETRY | IBV_QP_SQ_PSN | IBV_QP_MAX_QP_RD_ATOMIC));
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbListen(void* ctx, int dev, void* opaqueHandle, void** listenComm) {
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbListenComm* comm;
-  NCCLCHECK(ncclCalloc(&comm, 1));
-  struct ncclIbHandle* handle = (struct ncclIbHandle*) opaqueHandle;
-  static_assert(sizeof(struct ncclIbHandle) < NCCL_NET_HANDLE_MAXSIZE, "ncclIbHandle size too large");
-  memset(handle, 0, sizeof(struct ncclIbHandle));
-  comm->dev = dev;
-  handle->magic = NCCL_SOCKET_MAGIC;
-  NCCLCHECKGOTO(ncclSocketInit(&comm->sock, &ncclIbIfAddr, handle->magic, ncclSocketTypeNetIb, NULL, 1), ret, fail);
-  NCCLCHECKGOTO(ncclSocketListen(&comm->sock), ret, fail);
-  NCCLCHECKGOTO(ncclSocketGetAddr(&comm->sock, &handle->connectAddr), ret, fail);
-  *listenComm = comm;
-exit:
-  return ret;
-fail:
-  (void)ncclSocketClose(&comm->sock);
-  free(comm);
-  goto exit;
-}
-
-ncclResult_t ncclIbConnect(void* ctx, int dev, void* opaqueHandle, void** sendComm, ncclNetDeviceHandle_t** /*sendDevComm*/) {
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbHandle* handle = (struct ncclIbHandle*) opaqueHandle;
-  struct ncclIbCommStage* stage = &handle->stage;
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)stage->comm;
-  int ready;
-  uint8_t link_layer = IBV_LINK_LAYER_UNSPECIFIED;
-  int isP2p = 0; 
-  *sendComm = NULL;
-
-  if (stage->state == ncclIbCommStateConnect)      goto ib_connect_check;
-  if (stage->state == ncclIbCommStateSendDevList)  goto ib_send_dev_list;
-  if (stage->state == ncclIbCommStateRecvDevList)  goto ib_recv_dev_list;
-  if (stage->state == ncclIbCommStateSend)         goto ib_send;
-  if (stage->state == ncclIbCommStateConnecting)   goto ib_connect;
-  if (stage->state == ncclIbCommStateConnected)    goto ib_send_ready;
-  if (stage->state != ncclIbCommStateStart) {
-    WARN("Error: trying to connect already connected sendComm");
-    return ncclInternalError;
-  }
-  stage->buffer = NULL;
-
-  NCCLCHECK(ncclIbMalloc((void**)&comm, sizeof(struct ncclIbSendComm)));
-  NCCLCHECKGOTO(ncclIbStatsInit(&comm->base.stats), ret, fail);
-  NCCLCHECKGOTO(ncclSocketInit(&comm->base.sock, &handle->connectAddr, handle->magic, ncclSocketTypeNetIb, NULL, 1), ret, fail);
-  stage->comm = comm;
-  stage->state = ncclIbCommStateConnect;
-  NCCLCHECKGOTO(ncclSocketConnect(&comm->base.sock), ret, fail);
-
-ib_connect_check:
-  /* since ncclSocketConnect is async, we must check if connection is complete */
-  NCCLCHECKGOTO(ncclSocketReady(&comm->base.sock, &ready), ret, fail);
-  if (!ready) return ncclSuccess;
-
-  // IB Setup
-  struct ncclIbMergedDev* mergedDev;
-  if (dev >= ncclNMergedIbDevs) {
-    WARN("NET/IB : Trying to use non-existent virtual device %d", dev);
-    return ncclInternalError;
-  }
-
-  mergedDev = ncclIbMergedDevs + dev;
-  comm->base.vProps = mergedDev->vProps;
-  comm->base.isSend = true;
-  stage->state = ncclIbCommStateSendDevList;
-  stage->offset = 0;
-  struct ncclIbConnectionMetadata meta;
-  NCCLCHECKGOTO(ncclIbMalloc((void**)&stage->buffer, sizeof(meta)), ret, fail);
-  memcpy(stage->buffer, &mergedDev->vProps, sizeof(ncclNetVDeviceProps_t));
-
-// In the case of mismatched nDevs, we will make sure that both sides of a logical connection have the same number of RC qps
-ib_send_dev_list:
-  NCCLCHECK(ncclSocketProgress(NCCL_SOCKET_SEND, &comm->base.sock, stage->buffer, sizeof(ncclNetVDeviceProps_t), &stage->offset));
-  if (stage->offset != sizeof(ncclNetVDeviceProps_t)) return ncclSuccess;
-
-  stage->state = ncclIbCommStateRecvDevList;
-  stage->offset = 0;
-
-ib_recv_dev_list:
-  NCCLCHECK(ncclSocketProgress(NCCL_SOCKET_RECV, &comm->base.sock, stage->buffer, sizeof(ncclNetVDeviceProps_t), &stage->offset));
-  if (stage->offset != sizeof(ncclNetVDeviceProps_t)) return ncclSuccess;
-  stage->offset = 0;
-  ncclNetVDeviceProps_t remoteVProps;
-  ncclNetCommConfig_t* config;
-  memcpy(&remoteVProps, stage->buffer, sizeof(ncclNetVDeviceProps_t));
-  mergedDev = ncclIbMergedDevs + dev;
-  comm->base.vProps = mergedDev->vProps;
-
-  // Read isP2p from handle
-  isP2p = handle->isP2p;
-  INFO(NCCL_NET, "NET/IB: ncclIbConnect isP2p=%d", isP2p);
-  comm->base.nqps = ncclIbCalculateNqps(isP2p, comm->base.vProps.ndevs, 
-                                         remoteVProps.ndevs, __func__);
-
-  // Init PD, Ctx for each IB device
-  comm->ar = 1; // Set to 1 for logic
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    int ibDevN = comm->base.vProps.devs[i];
-    NCCLCHECKGOTO(ncclIbInitCommDevBase(ibDevN, &comm->devs[i].base, &comm->base.stats), ret, fail);
-    comm->ar = comm->ar && ncclIbDevs[ibDevN].ar; // ADAPTIVE_ROUTING - if all merged devs have it enabled
-  }
-
-  memset(&meta, 0, sizeof(meta));
-  meta.ndevs = comm->base.vProps.ndevs;
-  meta.isP2p = isP2p;
-  // Alternate QPs between devices
-  int devIndex;
-  devIndex = 0;
-  for (int q = 0; q < comm->base.nqps; q++) {
-    ncclIbSendCommDev* commDev = comm->devs + devIndex;
-    ncclIbDev* ibDev = ncclIbDevs + commDev->base.ibDevN;
-    NCCLCHECKGOTO(ncclIbCreateQp(ibDev->portNum, &commDev->base, IBV_ACCESS_REMOTE_WRITE, &comm->base.stats, comm->base.qps + q), ret, fail);
-    comm->base.qps[q].devIndex = devIndex;
-    meta.qpInfo[q].qpn      = comm->base.qps[q].qp->qp_num;
-    meta.qpInfo[q].devIndex = comm->base.qps[q].devIndex;
-
-    if (ncclParamIbEceEnable()) {
-      // Query ece capabilities (enhanced connection establishment)
-      NCCLCHECKGOTO(wrap_ibv_query_ece(comm->base.qps[q].qp, &meta.qpInfo[q].ece, &meta.qpInfo[q].ece_supported), ret, fail);
-    } else {
-      meta.qpInfo[q].ece_supported = 0;
-    }
-    devIndex = (devIndex + 1) % comm->base.vProps.ndevs;
-  }
-
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    ncclIbSendCommDev* commDev = comm->devs + i;
-    ncclIbDev* ibDev = ncclIbDevs + commDev->base.ibDevN;
-
-    // Write to the metadata struct via this pointer
-    ncclIbDevInfo* devInfo = meta.devs + i;
-    devInfo->ib_port       = ibDev->portNum;
-    devInfo->mtu           = ibDev->portAttr.active_mtu;
-    devInfo->lid           = ibDev->portAttr.lid;
-    devInfo->ibv_dev_index = commDev->base.ibDevN;
-    // Prepare GIN Put Signal scratchpad (for RDMA Atomic result)
-    NCCLCHECKGOTO(wrap_ibv_reg_mr(&commDev->putSignalScratchpadMr, commDev->base.pd, &comm->putSignalScratchpad, sizeof(comm->putSignalScratchpad), IBV_ACCESS_LOCAL_WRITE), ret, fail);
-
-    // Prepare my fifo
-    NCCLCHECKGOTO(wrap_ibv_reg_mr(&commDev->fifoMr, commDev->base.pd, comm->fifo, sizeof(struct ncclIbSendFifo)*MAX_REQUESTS*NCCL_NET_IB_MAX_RECVS, IBV_ACCESS_LOCAL_WRITE|IBV_ACCESS_REMOTE_WRITE|IBV_ACCESS_REMOTE_READ), ret, fail);
-    devInfo->fifoRkey = commDev->fifoMr->rkey;
-
-    // Pack local GID info
-    devInfo->link_layer = commDev->base.gidInfo.link_layer = ibDev->portAttr.link_layer;
-    NCCLCHECKGOTO(ncclIbGetGidIndex(ibDev->context, ibDev->portNum, &ibDev->portAttr, &commDev->base.gidInfo.localGidIndex), ret, fail);
-    NCCLCHECKGOTO(wrap_ibv_query_gid(ibDev->context, ibDev->portNum, commDev->base.gidInfo.localGidIndex, &commDev->base.gidInfo.localGid), ret, fail);
-    devInfo->gid.global.subnet_prefix = commDev->base.gidInfo.localGid.global.subnet_prefix;
-    devInfo->gid.global.interface_id = commDev->base.gidInfo.localGid.global.interface_id;
-
-    // info logging
-    for (int q = 0; q < comm->base.nqps; q++) {
-      // Print just the QPs for this dev
-      if (comm->base.qps[q].devIndex == i) {
-        if (devInfo->link_layer == IBV_LINK_LAYER_INFINIBAND) { // IB
-          INFO(NCCL_NET,"NET/IB: %s %d IbDev %d Port %d qpn %d mtu %d LID %d subnet-prefix %lu  FLID %d fifoRkey=0x%x fifoLkey=0x%x",
-               comm->base.vProps.ndevs > 2 ? "NCCL MergedDev" : "NCCL Dev",
-               dev, commDev->base.ibDevN, ibDev->portNum, meta.qpInfo[q].qpn, devInfo->mtu, devInfo->lid,
-               devInfo->gid.global.subnet_prefix, ncclIbExtractFlid(&devInfo->gid), devInfo->fifoRkey, commDev->fifoMr->lkey);
-        } else { // RoCE
-          INFO(NCCL_NET,"NET/IB: %s %d IbDev %d Port %d qpn %d mtu %d GID %ld (%lX/%lX) fifoRkey=0x%x fifoLkey=0x%x",
-               comm->base.vProps.ndevs > 2 ? "NCCL MergedDev" : "NCCL Dev", dev,
-               commDev->base.ibDevN, ibDev->portNum, meta.qpInfo[q].qpn, devInfo->mtu,
-               (int64_t)commDev->base.gidInfo.localGidIndex,
-               devInfo->gid.global.subnet_prefix, devInfo->gid.global.interface_id, devInfo->fifoRkey, commDev->fifoMr->lkey);
-        }
-        // Log ECE info
-        if (meta.qpInfo[q].ece_supported) {
-          INFO(NCCL_NET,"NET/IB: IbDev %d Port %d qpn %d query_ece={supported=%d, vendor_id=0x%x, options=0x%x, comp_mask=0x%x}",
-               commDev->base.ibDevN, ibDev->portNum, meta.qpInfo[q].qpn,
-               meta.qpInfo[q].ece_supported, meta.qpInfo[q].ece.vendor_id, meta.qpInfo[q].ece.options, meta.qpInfo[q].ece.comp_mask);
-        }
-      }
-    }
-    if (link_layer == IBV_LINK_LAYER_UNSPECIFIED) link_layer = devInfo->link_layer;
-    if (link_layer != devInfo->link_layer) {
-      int ibDev0 = comm->devs[0].base.ibDevN;
-      WARN("NET/IB : Attempted to connect incompatible devices: [%d]%s:%d/%s and [%d]%s:%d/%s. Try selecting NICs of only one link type using NCCL_IB_HCA",
-           commDev->base.ibDevN, ibDev->devName, ibDev->portNum, NCCL_IB_LLSTR(ibDev->portAttr.link_layer), ibDev0, ncclIbDevs[ibDev0].devName, ncclIbDevs[ibDev0].portNum, NCCL_IB_LLSTR(link_layer));
-      return ncclInternalError;
-    }
-  }
-  config = (ncclNetCommConfig_t*)ctx;
-  meta.fifoAddr = (uint64_t)comm->fifo;
-  meta.sl = (ncclParamIbSl() != -1) ? ncclParamIbSl() : (config && config->trafficClass != NCCL_NET_TRAFFIC_CLASS_UNDEF) ? config->trafficClass : NCCL_IB_SL_DEFAULT;
-  meta.tc = (ncclParamIbTc() != -1) ? ncclParamIbTc() : (config && config->trafficClass != NCCL_NET_TRAFFIC_CLASS_UNDEF) ? config->trafficClass : NCCL_IB_TC_DEFAULT;
-  strncpy(meta.devName, mergedDev->devName, MAX_MERGED_DEV_NAME);
-
-  stage->state = ncclIbCommStateSend;
-  stage->offset = 0;
-
-  memcpy(stage->buffer, &meta, sizeof(meta));
-
-ib_send:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_SEND, &comm->base.sock, stage->buffer, sizeof(meta), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(meta)) return ncclSuccess;
-
-  stage->state = ncclIbCommStateConnecting;
-  stage->offset = 0;
-  // Clear the staging buffer for re-use
-  memset(stage->buffer, 0, sizeof(meta));
-
-ib_connect:
-  struct ncclIbConnectionMetadata remMeta;
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_RECV, &comm->base.sock, stage->buffer, sizeof(ncclIbConnectionMetadata), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(remMeta)) return ncclSuccess;
-
-  memcpy(&remMeta, stage->buffer, sizeof(ncclIbConnectionMetadata));
-
-  comm->base.nRemDevs = remMeta.ndevs;
-
-  // ensure that the remote devices have the same link layer than the local devices used in the connection.
-  if (comm->base.vProps.ndevs > 0) {
-    int ibDev0 = comm->devs[0].base.ibDevN;
-    link_layer = ncclIbDevs[ibDev0].portAttr.link_layer;
-    for (int i = 0; i < remMeta.ndevs; i++) {
-      if (remMeta.devs[i].link_layer != link_layer) {
-        WARN("NET/IB : Remote %s device is incompatible with the local [%d]%s:%d/%s. Try selecting NICs of only one link type using NCCL_IB_HCA",
-             NCCL_IB_LLSTR(remMeta.devs[i].link_layer), ibDev0, ncclIbDevs[ibDev0].devName, ncclIbDevs[ibDev0].portNum, NCCL_IB_LLSTR(link_layer));
-        return ncclInternalError;
-      }
-    }
-  }
-
-  // Copy remDevInfo for things like remGidInfo, remFifoAddr, etc.
-  for (int i = 0; i < remMeta.ndevs; i++) {
-    comm->base.remDevs[i] = remMeta.devs[i];
-    comm->base.remDevs[i].remoteGid.global.interface_id = comm->base.remDevs[i].gid.global.interface_id;
-    comm->base.remDevs[i].remoteGid.global.subnet_prefix = comm->base.remDevs[i].gid.global.subnet_prefix;
-    // Retain remote sizes fifo info and prepare RDMA ops
-    comm->remSizesFifo.rkeys[i] = remMeta.devs[i].fifoRkey;
-    comm->remSizesFifo.addr = remMeta.fifoAddr;
-  }
-
-  for (int i=0; i < comm->base.vProps.ndevs; i++) {
-    NCCLCHECKGOTO(wrap_ibv_reg_mr(comm->remSizesFifo.mrs+i, comm->devs[i].base.pd, &comm->remSizesFifo.elems, sizeof(int)*MAX_REQUESTS*NCCL_NET_IB_MAX_RECVS, IBV_ACCESS_REMOTE_WRITE|IBV_ACCESS_LOCAL_WRITE|IBV_ACCESS_REMOTE_READ), ret, fail);
-  }
-  comm->base.nRemDevs = remMeta.ndevs;
-
-  for (int q = 0; q < comm->base.nqps; q++) {
-    struct ncclIbQpInfo* remQpInfo   = remMeta.qpInfo + q;
-    struct ncclIbDevInfo* remDevInfo = remMeta.devs + remQpInfo->devIndex;
-
-    // Assign per-QP remDev
-    comm->base.qps[q].remDevIdx = remQpInfo->devIndex;
-    int devIndex = comm->base.qps[q].devIndex;
-    ncclIbSendCommDev* commDev = comm->devs + devIndex;
-
-    struct ibv_qp* qp = comm->base.qps[q].qp;
-    if (remQpInfo->ece_supported) {
-      struct ncclIbQp* nqp = comm->base.qps + q;
-      int ibDevN = comm->devs[nqp->devIndex].base.ibDevN;
-      struct ncclIbDev* ibDev = ncclIbDevs + ibDevN;
-      INFO(NCCL_NET,"NET/IB: IbDev %d Port %d qpn %d set_ece={supported=%d, vendor_id=0x%x, options=0x%x, comp_mask=0x%x}",
-        ibDevN, ibDev->portNum, qp->qp_num, remMeta.qpInfo[q].ece_supported, remMeta.qpInfo[q].ece.vendor_id, remMeta.qpInfo[q].ece.options, remMeta.qpInfo[q].ece.comp_mask);
-      NCCLCHECKGOTO(wrap_ibv_set_ece(qp, &remQpInfo->ece, &remQpInfo->ece_supported), ret, fail);
-    }
-
-    ncclIbDev* ibDev = ncclIbDevs + commDev->base.ibDevN;
-    remDevInfo->mtu = std::min(remDevInfo->mtu, ibDev->portAttr.active_mtu);
-    NCCLCHECKGOTO(ncclIbRtrQp(qp, &commDev->base.gidInfo, remQpInfo->qpn, remDevInfo, false, remMeta.tc, remMeta.sl), ret, fail);
-    NCCLCHECKGOTO(ncclIbRtsQp(qp), ret, fail);
-  }
-
-  comm->base.nDataQps = std::max(comm->base.vProps.ndevs, comm->base.nRemDevs);
-
-  comm->base.ready = 1;
-  stage->state = ncclIbCommStateConnected;
-  stage->offset = 0;
-
-ib_send_ready:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_SEND, &comm->base.sock, &comm->base.ready, sizeof(int), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(int)) return ncclSuccess;
-
-  *sendComm = comm;
-exit:
-  if (stage->buffer) free(stage->buffer);
-  stage->state = ncclIbCommStateStart;
-  return ret;
-fail:
-  free(comm);
-  goto exit;
-}
-
-NCCL_PARAM(IbWarnRailLocal, "IB_WARN_RAIL_LOCAL", 0);
-
-ncclResult_t ncclIbCheckVProps(ncclNetVDeviceProps_t* vProps1, ncclNetVDeviceProps_t* vProps2) {
-  ncclNetVDeviceProps_t  outVProps = {0};
-  ncclNetVDeviceProps_t* minVProps = vProps2;
-  ncclNetVDeviceProps_t* maxVProps = vProps1;
-  if (vProps2->ndevs > vProps1->ndevs) {
-    minVProps = vProps1;
-    maxVProps = vProps2;
-  }
-
-  // Find the intersection of devices
-  for (int i = 0; i < minVProps->ndevs; i++) {
-    int dev = minVProps->devs[i];
-    for (int j = 0; j < maxVProps->ndevs; j++) {
-      // Found
-      if (maxVProps->devs[j] == dev) {
-        outVProps.devs[outVProps.ndevs++] = dev;
-      }
-    }
-  }
-
-  // In the case that at least one side has a fused NIC but there are no matching physical NICs, we should check if the user wants this
-  if (ncclParamIbWarnRailLocal() && outVProps.ndevs < maxVProps->ndevs) {
-    char local[128];
-    int cursor = 1;
-    snprintf(local, sizeof(local), "%d", vProps1->devs[0]);
-    for (int i = 1; i < vProps1->ndevs; i++) {
-      snprintf(local+cursor, sizeof(local)-cursor, ",%d", vProps1->devs[i]);
-      cursor += 2;
-    }
-    char remote[128];
-    snprintf(remote, sizeof(remote), "%d", vProps2->devs[0]);
-    cursor = 1;
-    for (int i = 1; i < vProps2->ndevs; i++) {
-      snprintf(remote+cursor, sizeof(remote)-cursor, ",%d", vProps2->devs[i]);
-      cursor += 2;
-    }
-    INFO(NCCL_NET, "NET/IB : There are mismatched physical devices between local (%s) and remote (%s). To disable this warning, set NCCL_IB_WARN_RAIL_LOCAL=0", local, remote);
-  }
-
-  return ncclSuccess;
-}
-
-NCCL_PARAM(IbGdrFlushDisable, "GDR_FLUSH_DISABLE", 0);
-RCCL_PARAM(IbGdrFlushGpuMemNoRelaxedOrdering, "GDR_FLUSH_GPU_MEM_NO_RELAXED_ORDERING", 1);
-
-ncclResult_t ncclIbAccept(void* listenComm, void** recvComm, ncclNetDeviceHandle_t** /*recvDevComm*/) {
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbListenComm* lComm = (struct ncclIbListenComm*)listenComm;
-  struct ncclIbCommStage* stage = &lComm->stage;
-  struct ncclIbRecvComm* rComm = (struct ncclIbRecvComm*)stage->comm;
-  int ready;
-  int link_layer = IBV_LINK_LAYER_UNSPECIFIED;
-  // Pre-declare dmabuf variables because of goto
-  bool useDmaBuf = false;
-  bool peermemAvailable = false;
-  bool dmabufSupported = false;
-  bool dmabufEnabled = false;
-  bool dmabufAvailable = false;
-  bool forceDmaBuf = false;
-  *recvComm = NULL;
-
-  if (stage->state == ncclIbCommStateAccept)   goto ib_accept_check;
-  if (stage->state == ncclIbCommStateRecvDevList) goto ib_recv_dev_list;
-  if (stage->state == ncclIbCommStateSendDevList) goto ib_send_dev_list;
-  if (stage->state == ncclIbCommStateRecv) goto ib_recv;
-  if (stage->state == ncclIbCommStateSend) goto ib_send;
-  if (stage->state == ncclIbCommStatePendingReady) goto ib_recv_ready;
-  if (stage->state != ncclIbCommStateStart) {
-    WARN("Listencomm in unknown state %d", stage->state);
-    return ncclInternalError;
-  }
-
-  NCCLCHECK(ncclIbMalloc((void**)&rComm, sizeof(struct ncclIbRecvComm)));
-  NCCLCHECKGOTO(ncclIbStatsInit(&rComm->base.stats), ret, fail);
-  stage->comm = rComm;
-  stage->state = ncclIbCommStateAccept;
-  NCCLCHECKGOTO(ncclSocketInit(&rComm->base.sock), ret, fail);
-  NCCLCHECKGOTO(ncclSocketAccept(&rComm->base.sock, &lComm->sock), ret, fail);
-
-  // Alloc stage->buffer here to be used for all following steps
-  struct ncclIbConnectionMetadata remMeta;
-  stage->offset = 0;
-  NCCLCHECK(ncclIbMalloc((void**)&stage->buffer, sizeof(remMeta)));
-
-ib_accept_check:
-  NCCLCHECKGOTO(ncclSocketReady(&rComm->base.sock, &ready), ret, fail);
-  if (!ready) return ncclSuccess;
-  stage->state = ncclIbCommStateRecvDevList;
-  stage->offset = 0;
-
-// In the case of mismatched nDevs, we will make sure that both sides of a logical connection have the same number of RC qps
-ib_recv_dev_list:
-  NCCLCHECK(ncclSocketProgress(NCCL_SOCKET_RECV, &rComm->base.sock, stage->buffer, sizeof(ncclNetVDeviceProps_t), &stage->offset));
-  if (stage->offset != sizeof(ncclNetVDeviceProps_t)) return ncclSuccess;
-  ncclNetVDeviceProps_t remoteVProps;
-  memcpy(&remoteVProps, stage->buffer, sizeof(ncclNetVDeviceProps_t));
-  if (lComm->dev >= ncclNMergedIbDevs) {
-    WARN("NET/IB : Trying to use non-existent virtual device %d", lComm->dev);
-    return ncclInternalError;
-  }
-
-  // Reduce the physical device list and store in the connection base
-  struct ncclIbMergedDev* mergedDev;
-  mergedDev = ncclIbMergedDevs + lComm->dev;
-  NCCLCHECK(ncclIbCheckVProps(&mergedDev->vProps, &remoteVProps));
-  rComm->base.vProps = mergedDev->vProps;
-  memcpy(stage->buffer, &rComm->base.vProps, sizeof(ncclNetVDeviceProps_t));
-  rComm->base.isSend = false;
-  stage->offset = 0;
-  stage->state = ncclIbCommStateSendDevList;
-
-ib_send_dev_list:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_SEND, &rComm->base.sock, stage->buffer, sizeof(ncclNetVDeviceProps_t), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(ncclNetVDeviceProps_t)) return ncclSuccess;
-
-  stage->offset = 0;
-  stage->state = ncclIbCommStateRecv;
-
-ib_recv:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_RECV, &rComm->base.sock, stage->buffer, sizeof(remMeta), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(remMeta)) return ncclSuccess;
-
-  /* copy back the received info */
-  memcpy(&remMeta, stage->buffer, sizeof(struct ncclIbConnectionMetadata));
-
-  // IB setup
-  // Pre-declare variables because of goto
-  struct ncclIbDev* ibDev;
-  int ibDevN;
-  struct ncclIbRecvCommDev* rCommDev;
-  struct ncclIbDevInfo* remDevInfo;
-  struct ncclIbQp* qp;
-
-  mergedDev = ncclIbMergedDevs + lComm->dev;
-  rComm->base.nRemDevs = remMeta.ndevs;
-  rComm->base.nqps = ncclIbCalculateNqps(remMeta.isP2p, rComm->base.vProps.ndevs, 
-                                          remMeta.ndevs, __func__);
-  if (rComm->base.nRemDevs != rComm->base.vProps.ndevs) {
-    INFO(NCCL_NET, "NET/IB : Local mergedDev %s has a different number of devices=%d as remote %s %d",
-      mergedDev->devName, rComm->base.vProps.ndevs, remMeta.devName, rComm->base.nRemDevs);
-  }
-
-  // Metadata to send back to requestor (sender)
-  struct ncclIbConnectionMetadata meta;
-  memset(&meta, 0, sizeof(meta));
-  for (int i = 0; i < rComm->base.vProps.ndevs; i++) {
-    rCommDev = rComm->devs + i;
-    ibDevN = rComm->base.vProps.devs[i];
-    NCCLCHECKGOTO(ncclIbInitCommDevBase(ibDevN, &rCommDev->base, &rComm->base.stats), ret, fail);
-    ibDev = ncclIbDevs + ibDevN;
-    NCCLCHECKGOTO(ncclIbGetGidIndex(ibDev->context, ibDev->portNum, &ibDev->portAttr, &rCommDev->base.gidInfo.localGidIndex), ret, fail);
-    NCCLCHECKGOTO(wrap_ibv_query_gid(ibDev->context, ibDev->portNum, rCommDev->base.gidInfo.localGidIndex, &rCommDev->base.gidInfo.localGid), ret, fail);
-    if (link_layer == IBV_LINK_LAYER_UNSPECIFIED) link_layer = ibDev->portAttr.link_layer;
-    if (link_layer != ibDev->portAttr.link_layer) {
-      int ibDev0 = rComm->devs[0].base.ibDevN;
-      WARN("NET/IB : Attempted to connect incompatible devices: [%d]%s:%d/%s and [%d]%s:%d/%s. Try selecting NICs of only one link type using NCCL_IB_HCA",
-           ibDevN, ibDev->devName, ibDev->portNum, NCCL_IB_LLSTR(ibDev->portAttr.link_layer), ibDev0, ncclIbDevs[ibDev0].devName, ncclIbDevs[ibDev0].portNum, NCCL_IB_LLSTR(link_layer));
-      return ncclInternalError;
-    }
-  }
-
-  // Copy remDevInfo for things like remGidInfo, remFifoAddr, etc.
-  for (int i = 0; i < remMeta.ndevs; i++) {
-    rComm->base.remDevs[i] = remMeta.devs[i];
-    rComm->base.remDevs[i].remoteGid.global.interface_id  = rComm->base.remDevs[i].gid.global.interface_id;
-    rComm->base.remDevs[i].remoteGid.global.subnet_prefix = rComm->base.remDevs[i].gid.global.subnet_prefix;
-    if (remMeta.devs[i].link_layer != link_layer) {
-      int ibDev0 = rComm->devs[0].base.ibDevN;
-      WARN("NET/IB : Remote %s device is incompatible with the local [%d]%s:%d/%s. Try selecting NICs of only one link type using NCCL_IB_HCA",
-           NCCL_IB_LLSTR(remMeta.devs[i].link_layer), ibDev0, ncclIbDevs[ibDev0].devName, ncclIbDevs[ibDev0].portNum, NCCL_IB_LLSTR(link_layer));
-      return ncclInternalError;
-    }
-  }
-
-  // Stripe QP creation across merged devs
-  // Make sure to get correct remote peer dev and QP info
-  int remDevIndex;
-  int devIndex;
-  devIndex = 0;
-  for (int q = 0; q < rComm->base.nqps; q++) {
-    remDevIndex = remMeta.qpInfo[q].devIndex;
-    remDevInfo = remMeta.devs + remDevIndex;
-    qp = rComm->base.qps+q;
-    rCommDev = rComm->devs + devIndex;
-    qp->remDevIdx = remDevIndex;
-
-    // Local ibDevN
-    ibDevN = rComm->devs[devIndex].base.ibDevN;
-    ibDev = ncclIbDevs + ibDevN;
-    NCCLCHECKGOTO(ncclIbCreateQp(ibDev->portNum, &rCommDev->base, IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_ATOMIC, &rComm->base.stats, qp), ret, fail);
-    qp->devIndex = devIndex;
-    devIndex = (devIndex + 1) % rComm->base.vProps.ndevs;
-
-    // Set the ece (enhanced connection establishment) on this QP before RTR
-    if (remMeta.qpInfo[q].ece_supported) {
-      // Coverity suspects a copy-paste error below due to the use of remMeta in one argument and meta in another.
-      // However, this has been confirmed to be intentional.
-      // coverity[copy_paste_error]
-      NCCLCHECKGOTO(wrap_ibv_set_ece(qp->qp, &remMeta.qpInfo[q].ece, &meta.qpInfo[q].ece_supported), ret, fail);
-    } else {
-      meta.qpInfo[q].ece_supported = 0;
-    }
-
-    NCCLCHECKGOTO(ncclIbRtrQp(qp->qp, &rCommDev->base.gidInfo, remMeta.qpInfo[q].qpn, remDevInfo, true, remMeta.tc, remMeta.sl), ret, fail);
-    NCCLCHECKGOTO(ncclIbRtsQp(qp->qp), ret, fail);
-
-    // Query the reduced ece for this QP (matching enhancements between the requestor and the responder)
-    // Store this in our own qpInfo for returning to the requestor
-    if (remMeta.qpInfo[q].ece_supported && meta.qpInfo[q].ece_supported) {
-      NCCLCHECKGOTO(wrap_ibv_query_ece(qp->qp, &meta.qpInfo[q].ece, &meta.qpInfo[q].ece_supported), ret, fail);
-    }
-  }
-
-  // GDR mode selection logic
-  peermemAvailable = (ncclIbGdrSupport() == ncclSuccess);
-  dmabufSupported = (ncclIbDmaBufSupport(lComm->dev) == ncclSuccess);
-  dmabufEnabled = (ncclParamDmaBufEnable() != 0);
-  dmabufAvailable = dmabufSupported && dmabufEnabled;
-  forceDmaBuf = rcclParamForceEnableDMABUF();
-
-  if (forceDmaBuf) {
-    // RCCL_FORCE_ENABLE_DMABUF=1: always try DMAbuf, skip peermem
-    useDmaBuf = dmabufAvailable;
-    if (dmabufAvailable) {
-      INFO(NCCL_INIT|NCCL_NET, "NET/IB: Using GPU Direct RDMA (DMAbuf - forced) for device %d", lComm->dev);
-    } else {
-      WARN("NET/IB: RCCL_FORCE_ENABLE_DMABUF=1 but DMAbuf not available, GPU Direct RDMA disabled for device %d", lComm->dev);
-    }
-  } else {
-    // Normal path: prefer peermem over dmabuf
-    useDmaBuf = !peermemAvailable && dmabufAvailable;
-    if (peermemAvailable) {
-      INFO(NCCL_INIT|NCCL_NET, "NET/IB: Using GPU Direct RDMA (peermem) for device %d", lComm->dev);
-    } else if (useDmaBuf) {
-      WARN("NET/IB: Peermem not available, falling back to GPU Direct RDMA (DMAbuf) for device %d", lComm->dev);
-    } else if (!peermemAvailable && dmabufSupported && !dmabufEnabled) {
-      WARN("NET/IB: Peermem not available and DMAbuf is disabled (NCCL_DMABUF_ENABLE=0), GPU Direct RDMA disabled for device %d", lComm->dev);
-    } else {
-      WARN("NET/IB: GPU Direct RDMA not available for device %d", lComm->dev);
-    }
-  }
-
-  rComm->flushEnabled = ((peermemAvailable || useDmaBuf)
-                            && (ncclParamIbGdrFlushDisable() == 0)) ? 1 : 0;              
-  for (int i = 0; i < rComm->base.vProps.ndevs; i++) {
-    rCommDev = rComm->devs + i;
-    ibDev = ncclIbDevs + rCommDev->base.ibDevN;
-
-    // Retain remote fifo info and prepare my RDMA ops
-    rComm->remFifo.addr = remMeta.fifoAddr;
-    NCCLCHECKGOTO(wrap_ibv_reg_mr(&rCommDev->fifoMr, rCommDev->base.pd, &rComm->remFifo.elems, sizeof(struct ncclIbSendFifo)*MAX_REQUESTS*NCCL_NET_IB_MAX_RECVS, IBV_ACCESS_REMOTE_WRITE|IBV_ACCESS_LOCAL_WRITE|IBV_ACCESS_REMOTE_READ), ret, fail);
-    rCommDev->fifoSge.lkey = rCommDev->fifoMr->lkey;
-    if (ncclParamIbUseInline()) rComm->remFifo.flags = IBV_SEND_INLINE;
-
-    // Allocate Flush dummy buffer for GPU Direct RDMA
-    if (rComm->flushEnabled) {
-      bool gpuFlushRegistered = false;
-      rCommDev->gpuFlush.gpuFlushGpuMem = nullptr;
-      rCommDev->gpuFlush.gpuMr = nullptr;
-      rCommDev->gpuFlush.dmabuf_fd = -1;
-
-      if (rcclParamIbGdrFlushGpuMemNoRelaxedOrdering()) {
-        #if CUDA_VERSION >= 11070 || HIP_VERSION >= 71260540
-        if (ncclCuMemEnable()) {
-          NCCLCHECKGOTO(ncclMemAlloc((void**)&rCommDev->gpuFlush.gpuFlushGpuMem, sizeof(int)), ret, fail);
-          CUCHECKGOTO(cuMemGetHandleForAddressRange((void*)&rCommDev->gpuFlush.dmabuf_fd,
-                      (CUdeviceptr)rCommDev->gpuFlush.gpuFlushGpuMem, sizeof(int),
-                      CU_MEM_RANGE_HANDLE_TYPE_DMA_BUF_FD, 0), ret, cumem_flush_hsa);
-          NCCLCHECKGOTO(wrap_ibv_reg_dmabuf_mr(&rCommDev->gpuFlush.gpuMr, rCommDev->base.pd,
-                        0, sizeof(int), (uint64_t)rCommDev->gpuFlush.gpuFlushGpuMem,
-                        rCommDev->gpuFlush.dmabuf_fd,
-                        IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_READ), ret, cumem_flush_hsa);
-          gpuFlushRegistered = true;
-          goto flush_reg_done;
-cumem_flush_hsa:
-          if (rCommDev->gpuFlush.dmabuf_fd >= 0) {
-            close(rCommDev->gpuFlush.dmabuf_fd);
-            rCommDev->gpuFlush.dmabuf_fd = -1;
-          }
-        }
-#else
-#if defined(HIP_UNCACHED_MEMORY)
-        NCCLCHECKGOTO(ncclCudaCalloc(&rCommDev->gpuFlush.gpuFlushGpuMem, sizeof(int), /*manager=*/nullptr, ncclMemPersist, hipDeviceMallocUncached), ret, fail);
-#else
-        NCCLCHECKGOTO(ncclCudaCalloc(&rCommDev->gpuFlush.gpuFlushGpuMem, sizeof(int), /*manager=*/nullptr, ncclMemPersist, hipDeviceMallocFinegrained), ret, fail);
-#endif
-        if (useDmaBuf)
-        {
-          uint64_t export_offset = 0;
-          void *aligned_ptr = NULL;
-          size_t aligned_size = 0;
-          get_aligned_ptr_and_size(rCommDev->gpuFlush.gpuFlushGpuMem, sizeof(int), &aligned_ptr, &aligned_size);
-          HSACHECKGOTO(hsa_amd_portable_export_dmabuf(aligned_ptr, aligned_size, &rCommDev->gpuFlush.dmabuf_fd, &export_offset), ret, peermem_flush);
-          if (wrap_ibv_reg_dmabuf_mr(&rCommDev->gpuFlush.gpuMr, rCommDev->base.pd, export_offset, sizeof(int),
-                        (uint64_t)rCommDev->gpuFlush.gpuFlushGpuMem, rCommDev->gpuFlush.dmabuf_fd,
-                        IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_REMOTE_READ) != ncclSuccess) goto peermem_flush;
-          gpuFlushRegistered = true;
-          goto flush_reg_done;
-peermem_flush:
-          if (rCommDev->gpuFlush.dmabuf_fd >= 0) {
-            close(rCommDev->gpuFlush.dmabuf_fd);
-            rCommDev->gpuFlush.dmabuf_fd = -1;
-          }
-        }
-        #endif
-        flush_reg_done:
-                if (!gpuFlushRegistered) {
-                  if (rCommDev->gpuFlush.gpuFlushGpuMem) {
-                    ncclCudaFree(rCommDev->gpuFlush.gpuFlushGpuMem, /*manager=*/nullptr);
-                    rCommDev->gpuFlush.gpuFlushGpuMem = nullptr;
-                  }
-                  rCommDev->gpuFlush.gpuMr = nullptr;
-        }
-      }
-      NCCLCHECKGOTO(wrap_ibv_reg_mr(&rCommDev->gpuFlush.hostMr, rCommDev->base.pd, &rComm->gpuFlushHostMem, sizeof(int), IBV_ACCESS_LOCAL_WRITE), ret, fail);
-      rCommDev->gpuFlush.sge.addr = (uint64_t)&rComm->gpuFlushHostMem;
-      rCommDev->gpuFlush.sge.length = 1;
-      rCommDev->gpuFlush.sge.lkey = rCommDev->gpuFlush.hostMr->lkey;
-      NCCLCHECKGOTO(ncclIbCreateQp(ibDev->portNum, &rCommDev->base, IBV_ACCESS_LOCAL_WRITE | IBV_ACCESS_REMOTE_READ | IBV_ACCESS_REMOTE_WRITE, &rComm->base.stats, &rCommDev->gpuFlush.qp), ret, fail);
-      struct ncclIbDevInfo devInfo;
-      devInfo.lid         = ibDev->portAttr.lid;
-      devInfo.link_layer  = ibDev->portAttr.link_layer;
-      devInfo.ib_port     = ibDev->portNum;
-      devInfo.gid.global.subnet_prefix        = rCommDev->base.gidInfo.localGid.global.subnet_prefix;
-      devInfo.gid.global.interface_id         = rCommDev->base.gidInfo.localGid.global.interface_id;
-      devInfo.mtu         = ibDev->portAttr.active_mtu;
-      NCCLCHECKGOTO(ncclIbRtrQp(rCommDev->gpuFlush.qp.qp, &rCommDev->base.gidInfo, rCommDev->gpuFlush.qp.qp->qp_num, &devInfo, false, remMeta.tc, remMeta.sl), ret, fail);
-      NCCLCHECKGOTO(ncclIbRtsQp(rCommDev->gpuFlush.qp.qp), ret, fail);
-    }
-
-    // Fill Handle
-    meta.devs[i].lid                            = ibDev->portAttr.lid;
-    meta.devs[i].link_layer                     = rCommDev->base.gidInfo.link_layer = ibDev->portAttr.link_layer;
-    meta.devs[i].ib_port                        = ibDev->portNum;
-    meta.devs[i].gid.global.subnet_prefix       = rCommDev->base.gidInfo.localGid.global.subnet_prefix;
-    meta.devs[i].gid.global.interface_id        = rCommDev->base.gidInfo.localGid.global.interface_id;
-    meta.devs[i].mtu                            = ibDev->portAttr.active_mtu;
-    meta.devs[i].ibv_dev_index                  = rCommDev->base.ibDevN;
-
-    // Prepare sizes fifo
-    NCCLCHECKGOTO(wrap_ibv_reg_mr(&rComm->devs[i].sizesFifoMr, rComm->devs[i].base.pd, rComm->sizesFifo, sizeof(int)*MAX_REQUESTS*NCCL_NET_IB_MAX_RECVS, IBV_ACCESS_LOCAL_WRITE|IBV_ACCESS_REMOTE_WRITE|IBV_ACCESS_REMOTE_READ), ret, fail);
-    meta.devs[i].fifoRkey = rComm->devs[i].sizesFifoMr->rkey;
-  }
-  meta.fifoAddr = (uint64_t)rComm->sizesFifo;
-  meta.sl = remMeta.sl;
-  meta.tc = remMeta.tc;
-
-  for (int q = 0; q < rComm->base.nqps; q++) {
-    meta.qpInfo[q].qpn      = rComm->base.qps[q].qp->qp_num;
-    meta.qpInfo[q].devIndex = rComm->base.qps[q].devIndex;
-  }
-  meta.ndevs = rComm->base.vProps.ndevs;
-  meta.isP2p = remMeta.isP2p;
-  strncpy(meta.devName, mergedDev->devName, MAX_MERGED_DEV_NAME);
-  rComm->base.nDataQps = std::max(rComm->base.vProps.ndevs, rComm->base.nRemDevs);
-
-  stage->state = ncclIbCommStateSend;
-  stage->offset = 0;
-  if (stage->buffer) {
-    free(stage->buffer);
-    stage->buffer = NULL;
-  }
-  NCCLCHECKGOTO(ncclIbMalloc((void**)&stage->buffer, sizeof(struct ncclIbConnectionMetadata)), ret, fail);
-  memcpy(stage->buffer, &meta, sizeof(struct ncclIbConnectionMetadata));
-
-ib_send:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_SEND, &rComm->base.sock, stage->buffer, sizeof(struct ncclIbConnectionMetadata), &stage->offset), ret, fail);
-  if (stage->offset < sizeof(struct ncclIbConnectionMetadata)) return ncclSuccess;
-
-  stage->offset = 0;
-  stage->state = ncclIbCommStatePendingReady;
-
-ib_recv_ready:
-  NCCLCHECKGOTO(ncclSocketProgress(NCCL_SOCKET_RECV,  &rComm->base.sock, &rComm->base.ready, sizeof(int), &stage->offset), ret, fail);
-  if (stage->offset != sizeof(int)) return ncclSuccess;
-
-  *recvComm = rComm;
-exit:
-  /* reset lComm stage */
-  if (stage->buffer) free(stage->buffer);
-  stage->state = ncclIbCommStateStart;
-  stage->offset = 0;
-  stage->comm = NULL;
-  stage->buffer = NULL;
-  return ret;
-fail:
-  free(rComm);
-  goto exit;
-}
-
-ncclResult_t ncclIbGetRequest(struct ncclIbNetCommBase* base, struct ncclIbRequest** req) {
-  for (int i=0; i<MAX_REQUESTS; i++) {
-    struct ncclIbRequest* r = base->reqs+i;
-    if (r->type == NCCL_NET_IB_REQ_UNUSED) {
-      r->base = base;
-      r->sock = NULL;
-      memset(r->devBases, 0, sizeof(r->devBases));
-      memset(r->events, 0, sizeof(r->events));
-      *req = r;
-      return ncclSuccess;
-    }
-  }
-  WARN("NET/IB : unable to allocate requests");
-  *req = NULL;
-  return ncclInternalError;
-}
-
-ncclResult_t ncclIbFreeRequest(struct ncclIbRequest* r) {
-  r->type = NCCL_NET_IB_REQ_UNUSED;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbTest(void* request, int* done, int* size);
-
-ncclResult_t ncclIbRegMrDmaBufInternal(ncclIbNetCommDevBase* base, void* data, size_t size, int type, uint64_t offset, int fd, ibv_mr** mhandle) {
-  static __thread uintptr_t pageSize = 0;
-  if (pageSize == 0) pageSize = sysconf(_SC_PAGESIZE);
-  struct ncclIbMrCache* cache = &ncclIbDevs[base->ibDevN].mrCache;
-  uintptr_t addr = (uintptr_t)data & -pageSize;
-  size_t pages = ((uintptr_t)data + size - addr + pageSize-1)/pageSize;
-  std::lock_guard<std::mutex> lock(ncclIbDevs[base->ibDevN].mutex);
-  for (int slot=0; /*true*/; slot++) {
-    if (slot == cache->population || addr < cache->slots[slot].addr) { // didn't find in cache
-      if (cache->population == cache->capacity) { // must grow cache
-        cache->capacity = cache->capacity < 32 ? 32 : 2*cache->capacity;
-        NCCLCHECK(ncclRealloc(&cache->slots, cache->population, cache->capacity));
-      }
-      // Deregister / register
-      struct ibv_mr* mr;
-      // REMOTE_ATOMIC required for GIN proxy atomic fetch-add on signal MR;
-      // without it mlx5 returns WC_REM_ACCESS_ERR (vendor_err 0x88).
-      unsigned int flags = IBV_ACCESS_LOCAL_WRITE|IBV_ACCESS_REMOTE_WRITE|IBV_ACCESS_REMOTE_READ|IBV_ACCESS_REMOTE_ATOMIC;
-      if (ncclIbRelaxedOrderingEnabled) flags |= IBV_ACCESS_RELAXED_ORDERING;
-      if (fd != -1) {
-        /* DMA-BUF support */
-        if (!ncclIbDevs[base->ibDevN].capsProvider.mlx5.dataDirect) {
-          NCCLCHECK(wrap_ibv_reg_dmabuf_mr(&mr, base->pd, offset, pages*pageSize, addr, fd, flags));
-        } else {
-          NCCLCHECK(wrap_mlx5dv_reg_dmabuf_mr(&mr, base->pd, offset, pages*pageSize, addr, fd, flags, MLX5DV_REG_DMABUF_ACCESS_DATA_DIRECT));
-        }
-      } else {
-        if (ncclIbRelaxedOrderingEnabled) {
-          // Use IBVERBS_1.8 API - needed for IBV_ACCESS_RELAXED_ORDERING support
-          NCCLCHECK(wrap_ibv_reg_mr_iova2(&mr, base->pd, (void*)addr, pages*pageSize, addr, flags));
-        }
-        else {
-          NCCLCHECK(wrap_ibv_reg_mr(&mr, base->pd, (void*)addr, pages*pageSize, flags));
-        }
-      }
-      TRACE(NCCL_INIT|NCCL_NET,"regAddr=0x%lx size=%lld rkey=0x%x lkey=0x%x fd=%d", (unsigned long)addr, (long long)pages*pageSize, mr->rkey, mr->lkey, fd);
-      if (slot != cache->population) memmove(cache->slots+slot+1, cache->slots+slot, (cache->population-slot)*sizeof(struct ncclIbMr));
-      cache->slots[slot].addr = addr;
-      cache->slots[slot].pages = pages;
-      cache->slots[slot].refs = 1;
-      cache->slots[slot].mr = mr;
-      cache->population += 1;
-      *mhandle = mr;
-      return ncclSuccess;
-    } else if ((addr >= cache->slots[slot].addr) &&
-        ((addr-cache->slots[slot].addr)/pageSize+pages) <= cache->slots[slot].pages) {
-      cache->slots[slot].refs += 1;
-      *mhandle = cache->slots[slot].mr;
-      return ncclSuccess;
-    }
-  }
-  return ncclSuccess;
-}
-
-struct ncclIbNetCommDevBase* ncclIbGetNetCommDevBase(ncclIbNetCommBase* base, int devIndex) {
-  if (base->isSend) {
-    struct ncclIbSendComm* sComm = (struct ncclIbSendComm*) base;
-    return &sComm->devs[devIndex].base;
-  } else {
-    struct ncclIbRecvComm* rComm = (struct ncclIbRecvComm*) base;
-    return &rComm->devs[devIndex].base;
-  }
-}
-
-/* DMA-BUF support */
-ncclResult_t ncclIbRegMrDmaBuf(void* comm, void* data, size_t size, int type, uint64_t offset, int fd, void** mhandle) {
-  ncclResult_t ret = ncclSuccess;
-  assert(size > 0);
-  struct ncclIbNetCommBase* base = (struct ncclIbNetCommBase*) comm;
-  struct ncclIbMrHandle* mhandleWrapper = (struct ncclIbMrHandle*) malloc(sizeof(struct ncclIbMrHandle));
-  if (mhandleWrapper == nullptr) {
-    WARN("Failed to allocate IB MR handle wrapper");
-    return ncclSystemError;
-  }
-  for (int i = 0; i < base->vProps.ndevs; i++) {
-    // Each ncclIbNetCommDevBase is at different offset in send and recv netComms
-    struct ncclIbNetCommDevBase* devComm = ncclIbGetNetCommDevBase(base, i);
-    NCCLCHECKGOTO(ncclIbRegMrDmaBufInternal(devComm, data, size, type, offset, fd, mhandleWrapper->mrs + i), ret, fail);
-  }
-  *mhandle = (void*) mhandleWrapper;
-exit:
-  return ret;
-fail:
-  free(mhandleWrapper);
-  goto exit;
-}
-
-ncclResult_t ncclIbRegMr(void* comm, void* data, size_t size, int type, void** mhandle) {
-  return ncclIbRegMrDmaBuf(comm, data, size, type, 0ULL, -1, mhandle);
-}
-
-ncclResult_t ncclIbDeregMrInternal(ncclIbNetCommDevBase* base, ibv_mr* mhandle) {
-  struct ncclIbMrCache* cache = &ncclIbDevs[base->ibDevN].mrCache;
-  std::lock_guard<std::mutex> lock(ncclIbDevs[base->ibDevN].mutex);
-  for (int i=0; i < cache->population; i++) {
-    if (mhandle == cache->slots[i].mr) {
-      if (0 == --cache->slots[i].refs) {
-        memmove(&cache->slots[i], &cache->slots[--cache->population], sizeof(struct ncclIbMr));
-        if (cache->population == 0) {
-          free(cache->slots);
-          cache->slots = NULL;
-          cache->capacity = 0;
-        }
-        NCCLCHECK(wrap_ibv_dereg_mr(mhandle));
-      }
-      return ncclSuccess;
-    }
-  }
-  WARN("NET/IB: could not find mr %p inside cache of %d entries", mhandle, cache->population);
-  return ncclInternalError;
-}
-
-ncclResult_t ncclIbDeregMr(void* comm, void* mhandle) {
-  if (mhandle == NULL) return ncclSuccess;
-
-  struct ncclIbMrHandle* mhandleWrapper = (struct ncclIbMrHandle*) mhandle;
-  struct ncclIbNetCommBase* base = (struct ncclIbNetCommBase*) comm;
-  for (int i = 0; i < base->vProps.ndevs; i++) {
-    // Each ncclIbNetCommDevBase is at different offset in send and recv netComms
-    struct ncclIbNetCommDevBase* devComm = ncclIbGetNetCommDevBase(base, i);
-    NCCLCHECK(ncclIbDeregMrInternal(devComm, mhandleWrapper->mrs[i]));
-  }
-  free(mhandleWrapper);
-  return ncclSuccess;
-}
-
-NCCL_PARAM(IbSplitDataOnQps, "IB_SPLIT_DATA_ON_QPS", 0);
-RCCL_PARAM(IbSplitDataThreshold, "IB_SPLIT_DATA_THRESHOLD", 128);
-
-ncclResult_t ncclIbMultiSend(struct ncclIbSendComm* comm, int slot) {
-  struct ncclIbRequest** reqs = comm->fifoReqs[slot];
-  volatile struct ncclIbSendFifo* slots = comm->fifo[slot];
-  ncclResult_t ret = ncclSuccess;
-  int nreqs = slots[0].nreqs;
-  if (nreqs > NCCL_NET_IB_MAX_RECVS) return ncclInternalError;
-
-  uint64_t wr_id = 0ULL;
-  for (int r=0; r<nreqs; r++) {
-    struct ibv_send_wr* wr = comm->wrs+r;
-    memset(wr, 0, sizeof(struct ibv_send_wr));
-
-    struct ibv_sge* sge = comm->sges+r;
-    sge->addr=(uintptr_t)reqs[r]->send.data;
-    wr->opcode = IBV_WR_RDMA_WRITE;
-    wr->send_flags = 0;
-    wr->wr.rdma.remote_addr = slots[r].addr;
-    wr->next = wr + 1;
-    wr_id += (reqs[r] - comm->base.reqs) << (r*8);
-#ifdef NCCL_ENABLE_NET_PROFILING
-    reqs[r]->pInfo[0].nEventHandles = 0;
-#endif
-  }
-
-  // Write size as immediate data. In the case of multi-send, only write
-  // 0 or 1 as size to indicate whether there was data sent or received.
-  uint32_t immData = 0;
-  if (nreqs == 1) {
-    immData = reqs[0]->send.size;
-  } else {
-    int* sizes = comm->remSizesFifo.elems[slot];
-    for (int r=0; r<nreqs; r++) sizes[r] = reqs[r]->send.size;
-    comm->remSizesFifo.sge.addr = (uint64_t)sizes;
-    comm->remSizesFifo.sge.length = nreqs*sizeof(int);
-  }
-
-  struct ibv_send_wr* lastWr = comm->wrs+nreqs-1;
-  if (nreqs > 1 || (comm->ar && reqs[0]->send.size > ncclParamIbArThreshold())) {
-    // When using ADAPTIVE_ROUTING, send the bulk of the data first as an
-    // RDMA_WRITE, then a 0-byte RDMA_WRITE_WITH_IMM to trigger a remote
-    // completion.
-    lastWr++;
-    memset(lastWr, 0, sizeof(struct ibv_send_wr));
-    if (nreqs > 1) {
-      // Write remote sizes Fifo
-      lastWr->wr.rdma.remote_addr = comm->remSizesFifo.addr + slot*NCCL_NET_IB_MAX_RECVS*sizeof(int);
-      lastWr->num_sge = 1;
-      lastWr->sg_list = &comm->remSizesFifo.sge;
-    }
-  }
-  lastWr->wr_id = wr_id;
-  lastWr->opcode = IBV_WR_RDMA_WRITE_WITH_IMM;
-  lastWr->imm_data = immData;
-  lastWr->next = NULL;
-  lastWr->send_flags = IBV_SEND_SIGNALED;
-
-  // Multi-QP: make sure IB writes are multiples of 128B so that LL and LL128 protocols still work
-  const int align = 128;
-  int nqps = ncclParamIbSplitDataOnQps() ? comm->base.nqps : comm->base.nDataQps;
-
-  // For small messages (< splitDataThreshold) with multiple QPs, avoid splitting:
-  // send the entire payload through one QP selected via round-robin, post zero-length WRs on the rest.
-  int smallMsgActiveQp = comm->fifoHead % nqps;
-  int64_t splitDataThreshold = rcclParamIbSplitDataThreshold();
-
-  for (int i = 0; i < nqps; i++) {
-    int qpIndex = comm->base.qpIndex;
-    ncclIbQp* qp = comm->base.qps + qpIndex;
-    int devIndex = qp->devIndex;
-    for (int r=0; r<nreqs; r++) {
-      // Track this event for completion
-      //ncclIbAddEvent(reqs[r], devIndex, &comm->devs[devIndex].base);
-
-      // Select proper rkey (needed even for 0-size send)
-      comm->wrs[r].wr.rdma.rkey = slots[r].rkeys[qp->remDevIdx];
-
-      int chunkSize;
-      if (reqs[r]->send.size < splitDataThreshold) {
-        chunkSize = (i == smallMsgActiveQp) ? reqs[r]->send.size : 0;
-      } else {
-        chunkSize = DIVUP(DIVUP(reqs[r]->send.size, nqps), align) * align;
-      }
-      int length = std::min(reqs[r]->send.size-reqs[r]->send.offset, chunkSize);
-      if (length <= 0) {
-        comm->wrs[r].sg_list = NULL;
-        comm->wrs[r].num_sge = 0;
-      } else {
-        // Select proper lkey
-        comm->sges[r].lkey = reqs[r]->send.lkeys[devIndex];
-        comm->sges[r].length = length;
-        comm->wrs[r].sg_list = comm->sges+r;
-        comm->wrs[r].num_sge = 1;
-      }
-    }
-
-    if (nreqs > 1) {
-      // Also make sure lastWr writes remote sizes using the right lkey
-      comm->remSizesFifo.sge.lkey = comm->remSizesFifo.mrs[devIndex]->lkey;
-      lastWr->wr.rdma.rkey = comm->remSizesFifo.rkeys[devIndex];
-    }
-
-    struct ibv_send_wr* bad_wr;
-#ifdef NCCL_ENABLE_NET_PROFILING
-    // QP profiling loop
-    for (int r=0; r<nreqs; r++) {
-      // Store comm qpIndex for this request
-      int nEventHandles = reqs[r]->pInfo[0].nEventHandles;
-      assert(nEventHandles < MAX_QPS_PER_REQ);
-      reqs[r]->pInfo[0].qpIndex[nEventHandles] = qpIndex;
-      // Store info for profiler
-      int64_t pluginId = NCCL_PROFILER_NET_TYPE_IB | NCCL_PROFILER_NET_IB_VER;
-      reqs[r]->pInfo[0].data.type = ncclProfileQp;
-      reqs[r]->pInfo[0].data.qp.device = devIndex;
-      reqs[r]->pInfo[0].data.qp.wr_id = comm->wrs[r].wr_id;
-      reqs[r]->pInfo[0].data.qp.opcode = comm->wrs[r].opcode;
-      reqs[r]->pInfo[0].data.qp.qpNum = qp->qp->qp_num;
-      reqs[r]->pInfo[0].data.qp.length = comm->sges[r].length;
-      void* pHandle = reqs[r]->pInfo[0].pHandle;
-      NCCLCHECK(ncclProfilerFunction(&reqs[r]->pInfo[0].qpEventHandles[nEventHandles], ncclProfilerNetEventStart, pHandle, pluginId, &reqs[r]->pInfo[0].data));
-      reqs[r]->pInfo[0].nEventHandles++;
-    }
-#endif
-    ret = wrap_ibv_post_send(qp->qp, comm->wrs, &bad_wr);
-    if (ret != ncclSuccess) {
-      // Mark connection as fatal. DO NOT free requests here - the caller
-      // (ncclIbIsend) owns the requests and will handle cleanup.
-      // Some requests may have events from earlier QP iterations.
-      ncclIbStatsFatalError(&comm->base.stats);
-      return ret;
-    }
-
-    for (int r=0; r<nreqs; r++) {
-      int chunkSize;
-      if (reqs[r]->send.size < splitDataThreshold) {
-        chunkSize = (i == smallMsgActiveQp) ? reqs[r]->send.size : 0;
-      } else {
-        chunkSize = DIVUP(DIVUP(reqs[r]->send.size, nqps), align) * align;
-      }
-      reqs[r]->send.offset += chunkSize;
-      comm->sges[r].addr += chunkSize;
-      comm->wrs[r].wr.rdma.remote_addr += chunkSize;
-
-      TRACE(NCCL_VERBS, "Posted send wr_id=%lu, wr_indx=%d, qp_num=%d, src_nic=%d, dst_nic=%d, dlid=%d, opcode=%d, send_flags=%d, imm_data=%d, remote_addr=%lx, rkey=%x, length=%d, lkey=%x",
-        comm->wrs[r].wr_id, r, qp->qp->qp_num, comm->devs[qp->devIndex].base.ibDevN , comm->base.remDevs[qp->remDevIdx].ibv_dev_index, comm->base.remDevs[qp->remDevIdx].lid,
-        comm->wrs[r].opcode, comm->wrs[r].send_flags, comm->wrs[r].imm_data, comm->wrs[r].wr.rdma.remote_addr,
-        comm->wrs[r].wr.rdma.rkey,comm->wrs[r].sg_list ? comm->wrs[r].sg_list->length : 0, comm->wrs[r].sg_list ? comm->wrs[r].sg_list->lkey : 0);
-    }
-
-    // Select the next qpIndex
-    comm->base.qpIndex = (comm->base.qpIndex+1) % comm->base.nqps;
-  }
-
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbIsend(void* sendComm, void* data, size_t size, int tag, void* mhandle, void* phandle, void** request) {
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)sendComm;
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbRequest* req = NULL;
-  struct ncclIbMrHandle* mhandleWrapper = (struct ncclIbMrHandle*) mhandle;
-  int nreqs = 0;
-  volatile struct ncclIbSendFifo* slots = NULL;
-  int slot = (comm->fifoHead) % MAX_REQUESTS;
-  struct ncclIbRequest** reqs = comm->fifoReqs[slot];
-  uint64_t idx = 0;
-
-  if (comm->base.ready == 0) {
-    WARN("NET/IB: ncclIbIsend() called when comm->base.ready == 0");
-    *request = NULL;
-    return ncclInternalError;
-  }
-  NCCLCHECKGOTO(ncclIbStatsCheckFatalCount(&comm->base.stats,__func__), ret, fail);
-
-  // Wait for the receiver to have posted the corresponding receive
-  slots = comm->fifo[slot];
-  idx = comm->fifoHead+1;
-  if (slots[0].idx != idx) { *request = NULL; return ncclSuccess; }
-  nreqs = slots[0].nreqs;
-  // Wait until all data has arrived
-  for (int r=1; r<nreqs; r++) while(slots[r].idx != idx);
-  __sync_synchronize(); // order the nreqsPtr load against tag/rkey/addr loads below
-  for (int r=0; r<nreqs; r++) {
-    if (reqs[r] != NULL || slots[r].tag != tag) continue;
-
-    if (size > slots[r].size) size = slots[r].size;
-    // Sanity checks
-    if (slots[r].size < 0 || slots[r].addr == 0 || slots[r].rkeys[0] == 0) {
-      char line[SOCKET_NAME_MAXLEN + 1];
-      union ncclSocketAddress addr;
-      ncclSocketGetAddr(&comm->base.sock, &addr);
-      WARN("NET/IB : req %d/%d tag %x peer %s posted incorrect receive info: size %ld addr %lx rkeys[0]=%x",
-        r, nreqs, tag, ncclSocketToString(&addr, line), slots[r].size, slots[r].addr, slots[r].rkeys[0]);
-      return ncclInternalError;
-    }
-
-    NCCLCHECKGOTO(ncclIbGetRequest(&comm->base, &req), ret, fail);
-    req->type = NCCL_NET_IB_REQ_SEND;
-    req->sock = &comm->base.sock;
-    req->base = &comm->base;
-    req->nreqs = nreqs;
-    req->send.size = size;
-    req->send.data = data;
-    req->send.offset = 0;
-#ifdef NCCL_ENABLE_NET_PROFILING
-    req->pInfo[0].pHandle = phandle;
-#endif
-
-    // Populate events
-    int nEvents = ncclParamIbSplitDataOnQps() ? comm->base.nqps : comm->base.nDataQps;
-    int qpIndex = comm->base.qpIndex;
-    // Count down
-    while (nEvents > 0) {
-      ncclIbQp* qp = comm->base.qps + qpIndex;
-      int devIndex = qp->devIndex;
-      ncclIbAddEvent(req, devIndex, &comm->devs[devIndex].base);
-      // Track the valid lkey for this RDMA_Write
-      req->send.lkeys[devIndex] = mhandleWrapper->mrs[devIndex]->lkey;
-      nEvents--;
-      // Don't update comm->base.qpIndex yet, we need to run through this same set of QPs inside ncclIbMultiSend()
-      qpIndex = (qpIndex+1)%comm->base.nqps;
-    }
-
-    // Store all lkeys
-    for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-      req->send.lkeys[i] = mhandleWrapper->mrs[i]->lkey;
-    }
-
-    *request = reqs[r] = req;
-
-    // If this is a multi-recv, send only when all requests have matched.
-    for (int r=0; r<nreqs; r++) {
-      if (reqs[r] == NULL) return ncclSuccess;
-    }
-
-    TIME_START(0);
-    NCCLCHECKGOTO(ncclIbMultiSend(comm, slot), ret, fail);
-
-    // Clear slots[0]->nreqs, as well as other fields to help debugging and sanity checks
-    memset((void*)slots, 0, sizeof(struct ncclIbSendFifo));
-    memset(reqs, 0, NCCL_NET_IB_MAX_RECVS*sizeof(struct ncclIbRequest*));
-    comm->fifoHead++;
-    TIME_STOP(0);
-    return ncclSuccess;
-  }
-
-  *request = NULL;
-  return ncclSuccess;
-
-fail:
-  ncclIbStatsFatalError(&comm->base.stats);
-  if (req) {
-    // If events were added (IBV ops posted), we can't free immediately -
-    // completions may still arrive. Mark as failed and return it so
-    // caller can call Test to drain completions.
-    if (ncclIbRequestHasEvents(req)) {
-      req->type = NCCL_NET_IB_REQ_FAILED;
-      *request = req;
-      // Return success so caller proceeds to call Test() which will drain
-      // completions. The fatal error is recorded in stats and will be
-      // caught on the next operation.
-      return ncclSuccess;
-    } else {
-      ncclIbFreeRequest(req);
-      *request = NULL;
-    }
-  } else {
-    *request = NULL;
-  }
-  return ret;
-}
-
-ncclResult_t ncclIbPostFifo(struct ncclIbRecvComm* comm, int n, void** data, size_t* sizes, int* tags, void** mhandles, struct ncclIbRequest* req) {
-  struct ibv_send_wr wr;
-  memset(&wr, 0, sizeof(wr));
-
-  int slot = comm->remFifo.fifoTail%MAX_REQUESTS;
-  req->recv.sizes = comm->sizesFifo[slot];
-  for (int i=0; i<n; i++) req->recv.sizes[i] = 0;
-  struct ncclIbSendFifo* localElem = comm->remFifo.elems[slot];
-
-  // Select the next devIndex (local) and QP to use for posting this CTS message
-  // Since QPs are initialized by striping across devIndex, we can simply assign this to the same value
-  ncclIbQp* ctsQp = comm->base.qps + comm->base.devIndex;
-  comm->base.devIndex = (comm->base.devIndex + 1) % comm->base.vProps.ndevs;
-
-  for (int i=0; i<n; i++) {
-    localElem[i].addr = (uint64_t)data[i];
-    struct ncclIbMrHandle* mhandleWrapper = (struct ncclIbMrHandle*) mhandles[i];
-
-    // Send all applicable rkeys
-    for (int j = 0; j < comm->base.vProps.ndevs; j++)
-      localElem[i].rkeys[j] = mhandleWrapper->mrs[j]->rkey;
-
-    localElem[i].nreqs = n;
-    localElem[i].size = sizes[i]; // Sanity/Debugging
-    localElem[i].tag = tags[i];
-    localElem[i].idx = comm->remFifo.fifoTail+1;
-  }
-  wr.wr.rdma.remote_addr = comm->remFifo.addr + slot*NCCL_NET_IB_MAX_RECVS*sizeof(struct ncclIbSendFifo);
-
-  // Lookup the correct fifoRkey
-  wr.wr.rdma.rkey = comm->base.remDevs[ctsQp->remDevIdx].fifoRkey;
-
-  // Set the correct sge properties
-  comm->devs[ctsQp->devIndex].fifoSge.addr   = (uint64_t)localElem;
-  comm->devs[ctsQp->devIndex].fifoSge.length = n*sizeof(struct ncclIbSendFifo);
-  wr.sg_list = &comm->devs[ctsQp->devIndex].fifoSge;
-  wr.num_sge = 1;
-
-  wr.opcode = IBV_WR_RDMA_WRITE;
-  wr.send_flags = comm->remFifo.flags; // IBV_SEND_INLINE
-
-  // We need to occasionally post a request with the IBV_SEND_SIGNALED flag, otherwise
-  // the send queue will never empty.
-  //
-  // From https://www.rdmamojo.com/2014/06/30/working-unsignaled-completions/
-  // "How to use Unsignaled Completion?" / "Gotchas and Pitfalls"
-  // All posted Send Requested, Signaled and Unsignaled, are considered outstanding until
-  // a Work Completion that they, or Send Requests that were posted after them, was polled
-  // from the Completion Queue associated with the Send Queue. This means if one works with
-  // a Queue Pair that was configured to work with Unsignaled Completions, he must make
-  // sure that occasionally (before the Send Queue is full with outstanding Send Requests)
-  // a Send Request that generate Work Completion will be posted.
-  //
-  // Not following this rule may lead to a case that the Send Queue is full with Send
-  // Requests that won't generate Work Completion:
-  //
-  //  - The Send Queue is full, so no new Send Requests can be posted to it
-  //  - The Send Queue can't be emptied, since no Work Completion can be generated anymore
-  //    (the reason is that no Work Completion, that can generate Work Completion that
-  //    polling it will empty the Send Queue, can be posted)
-  //  - The status of all posted Send Request is considered unknown
-  //
-  // slot == devIndex - When writing to fifo slot N, and this QP lives on device index N, it should send signalled.
-  // This works out that each fifo posting QP gets drained
-  if (slot == ctsQp->devIndex) {
-    wr.send_flags |= IBV_SEND_SIGNALED;
-    wr.wr_id = req - comm->base.reqs;
-    ncclIbAddEvent(req, ctsQp->devIndex, &comm->devs[ctsQp->devIndex].base);
-  }
-
-  struct ibv_send_wr* bad_wr;
-  ncclResult_t ret = wrap_ibv_post_send(ctsQp->qp, &wr, &bad_wr);
-  if (ret != ncclSuccess) {
-    // Mark connection as fatal. Don't free request here - caller owns it
-    // and will handle cleanup based on whether events were posted.
-    ncclIbStatsFatalError(&comm->base.stats);
-    return ret;
-  }
-
-  TRACE(NCCL_VERBS, "Posted send wr_id=%lu, wr_indx=%d, qp_num=%d, src_nic=%d, dst_nic=%d, dlid=%u, opcode=%d, send_flags=%d, imm_data=%d, remote_addr=%lx, rkey=%x, length=%d, lkey=%x",
-        wr.wr_id, 0, ctsQp->qp->qp_num, comm->devs[ctsQp->devIndex].base.ibDevN, comm->base.remDevs[ctsQp->remDevIdx].ibv_dev_index, comm->base.remDevs[ctsQp->remDevIdx].lid,
-        wr.opcode, wr.send_flags, wr.imm_data, wr.wr.rdma.remote_addr, wr.wr.rdma.rkey, wr.sg_list ? wr.sg_list->length : 0, wr.sg_list ? wr.sg_list->lkey : 0);
-
-  comm->remFifo.fifoTail++;
-
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbIrecv(void* recvComm, int n, void** data, size_t* sizes, int* tags, void** mhandles, void** phandles, void** request) {
-  struct ncclIbRecvComm* comm = (struct ncclIbRecvComm*)recvComm;
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbRequest* req = NULL;
-  struct ibv_recv_wr wr;
-  struct ibv_recv_wr* bad_wr = NULL;
-  int nqps = 0;
-  if (comm->base.ready == 0) {
-    WARN("NET/IB: ncclIbIrecv() called when comm->base.ready == 0");
-    *request = NULL;
-    return ncclInternalError;
-  }
-  if (n > NCCL_NET_IB_MAX_RECVS) return ncclInternalError;
-  NCCLCHECKGOTO(ncclIbStatsCheckFatalCount(&comm->base.stats,__func__), ret, fail);
-
-  NCCLCHECKGOTO(ncclIbGetRequest(&comm->base, &req), ret, fail);
-  req->type = NCCL_NET_IB_REQ_RECV;
-  req->sock = &comm->base.sock;
-  req->nreqs = n;
-#ifdef NCCL_ENABLE_NET_PROFILING
-  for (int r = 0; r < n && phandles; r++) req->pInfo[r].nEventHandles = 0;
-#endif
-
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    req->devBases[i] = &comm->devs[i].base;
-  }
-
-  memset(&wr, 0, sizeof(wr));
-  wr.wr_id = req - comm->base.reqs;
-  wr.sg_list = NULL;
-  wr.num_sge = 0;
-
-  TIME_START(1);
-  // Select either all QPs, or one qp per-device
-  nqps = ncclParamIbSplitDataOnQps() ? comm->base.nqps : comm->base.nDataQps;
-
-  // Post recvs
-  for (int i = 0; i < nqps; i++) {
-    struct ncclIbQp* qp = comm->base.qps + comm->base.qpIndex;
-    ncclIbAddEvent(req, qp->devIndex, &comm->devs[qp->devIndex].base);
-#ifdef NCCL_ENABLE_NET_PROFILING
-    // Start a QP event for every request in the multirecv and every qp
-    for (int r = 0; r < n; r++) {
-      int nEventHandles = req->pInfo[r].nEventHandles;
-      assert(nEventHandles < MAX_QPS_PER_REQ);
-      req->pInfo[r].qpIndex[nEventHandles] = comm->base.qpIndex;
-      // Store info for profiler
-      int64_t pluginId = NCCL_PROFILER_NET_TYPE_IB | NCCL_PROFILER_NET_IB_VER;
-      req->pInfo[r].data.type = ncclProfileQp;
-      req->pInfo[r].data.qp.device = qp->devIndex;
-      req->pInfo[r].data.qp.wr_id = wr.wr_id;
-      req->pInfo[r].data.qp.qpNum = qp->qp->qp_num;
-      NCCLCHECK(ncclProfilerFunction(&req->pInfo[r].qpEventHandles[nEventHandles], ncclProfilerNetEventStart, phandles[r], pluginId, &req->pInfo[r].data));
-      req->pInfo[r].nEventHandles++;
-    }
-#endif
-    NCCLCHECKGOTO(wrap_ibv_post_recv(qp->qp, &wr, &bad_wr), ret, fail);
-    comm->base.qpIndex = (comm->base.qpIndex+1)%comm->base.nqps;
-  }
-
-  TIME_STOP(1);
-
-  // Post to FIFO to notify sender
-  TIME_START(2);
-  NCCLCHECKGOTO(ncclIbPostFifo(comm, n, data, sizes, tags, mhandles, req), ret, fail);
-  TIME_STOP(2);
-
-  *request = req;
-  return ncclSuccess;
-
-fail:
-ncclIbStatsFatalError(&comm->base.stats);
-if (req) {
-  // If events were added (IBV ops posted), we can't free immediately -
-  // completions may still arrive. Mark as failed and return it so
-  // caller can call Test to drain completions.
-  if (ncclIbRequestHasEvents(req)) {
-    req->type = NCCL_NET_IB_REQ_FAILED;
-    *request = req;
-    // Return success so caller proceeds to call Test() which will drain
-    // completions. The fatal error is recorded in stats and will be
-    // caught on the next operation.
-    return ncclSuccess;
-  } else {
-    ncclIbFreeRequest(req);
-    *request = NULL;
-  }
-} else {
-  *request = NULL;
-}
-return ret;
-}
-
-ncclResult_t ncclIbIflush(void* recvComm, int n, void** data, int* sizes, void** mhandles, void** request) {
-  struct ncclIbRecvComm* comm = (struct ncclIbRecvComm*)recvComm;
-  ncclResult_t ret = ncclSuccess;
-  struct ncclIbRequest* req = NULL;
-  struct ncclIbMrHandle* mhandle = NULL;
-  int last = -1;
-  for (int i=0; i<n; i++) if (sizes[i]) last = i;
-  if (comm->flushEnabled == 0 || last == -1) return ncclSuccess;
-
-  // Only flush once using the last non-zero receive
-  mhandle = (struct ncclIbMrHandle*) mhandles[last];
-  NCCLCHECKGOTO(ncclIbGetRequest(&comm->base, &req), ret, fail);
-  req->type = NCCL_NET_IB_REQ_FLUSH;
-  req->sock = &comm->base.sock;
-
-  // We don't know which devIndex the recv was on, so we flush on all devices
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    struct ibv_send_wr wr;
-    // Check if GPU flush buffer was successfully registered
-    // If not, fall back to using user data buffer for flush
-    bool useGpuFlushMem = rcclParamIbGdrFlushGpuMemNoRelaxedOrdering() &&
-                          comm->devs[i].gpuFlush.gpuFlushGpuMem != nullptr &&
-                          comm->devs[i].gpuFlush.gpuMr != nullptr;
-    memset(&wr, 0, sizeof(wr));
-    wr.wr_id = req - comm->base.reqs;
-    if (useGpuFlushMem) {
-      wr.wr.rdma.remote_addr = (uint64_t)(comm->devs[i].gpuFlush.gpuFlushGpuMem);
-      wr.wr.rdma.rkey = comm->devs[i].gpuFlush.gpuMr->rkey;
-      wr.sg_list = &comm->devs[i].gpuFlush.sge;
-      wr.num_sge = 1;
-      wr.opcode = IBV_WR_RDMA_WRITE;
-      wr.send_flags = 0;
-      struct ibv_send_wr* bad_wr;
-      NCCLCHECKGOTO(wrap_ibv_post_send(comm->devs[i].gpuFlush.qp.qp, &wr, &bad_wr), ret, fail);
-    }
-    memset(&wr, 0, sizeof(wr));
-    wr.wr_id = req - comm->base.reqs;
-    if (useGpuFlushMem) {
-      wr.wr.rdma.remote_addr = (uint64_t)(comm->devs[i].gpuFlush.gpuFlushGpuMem);
-      wr.wr.rdma.rkey = comm->devs[i].gpuFlush.gpuMr->rkey;
-    } else {
-      wr.wr.rdma.remote_addr = (uint64_t)data[last];
-      wr.wr.rdma.rkey = mhandle->mrs[i]->rkey;
-    }
-    wr.sg_list = &comm->devs[i].gpuFlush.sge;
-    wr.num_sge = 1;
-    wr.opcode = IBV_WR_RDMA_READ;
-    wr.send_flags = IBV_SEND_SIGNALED;
-
-    TIME_START(4);
-    struct ibv_send_wr* bad_wr;
-    NCCLCHECKGOTO(wrap_ibv_post_send(comm->devs[i].gpuFlush.qp.qp, &wr, &bad_wr), ret, fail);
-    TIME_STOP(4);
-
-    ncclIbAddEvent(req, i, &comm->devs[i].base);
-  }
-
-  *request = req;
-  return ncclSuccess;
-
-fail:
-ncclIbStatsFatalError(&comm->base.stats);
-if (req) {
-  // If events were added (IBV ops posted), we can't free immediately -
-  // completions may still arrive. Mark as failed and return it so
-  // caller can call Test to drain completions.
-  if (ncclIbRequestHasEvents(req)) {
-    req->type = NCCL_NET_IB_REQ_FAILED;
-    *request = req;
-    // Return success so caller proceeds to call Test() which will drain
-    // completions. The fatal error is recorded in stats and will be
-    // caught on the next operation.
-    return ncclSuccess;
-  } else {
-    ncclIbFreeRequest(req);
-    *request = NULL;
-  }
-} else {
-  *request = NULL;
-}
-return ret;
-}
-
-#define HCA_NAME(req, index) ((req)->devBases[(index)]->pd->context->device->name)
-
-#ifdef NCCL_ENABLE_NET_PROFILING
-static int getReqQpIndex(struct ncclIbRequest* req, int request, int qpNumber) {
-  for (int i = 0; i < MAX_QPS_PER_REQ; i++) {
-    int qpIndex = req->pInfo[request].qpIndex[i];
-    if (req->base->qps[qpIndex].qp->qp_num == qpNumber) return i;
-  }
-  return 0;
-}
-#endif
-
-ncclResult_t ncclIbTest(void* request, int* done, int* sizes) {
-  struct ncclIbRequest *r = (struct ncclIbRequest*)request;
-  *done = 0;
-  ncclResult_t ret = ncclSuccess;
-  int failDevIdx = -1;
-  while (1) {
-    NCCLCHECKGOTO(ncclIbStatsCheckFatalCount(&r->base->stats,__func__), ret, fail);
-    if (r->events[0] == 0 && r->events[1] == 0 && r->events[2] == 0 && r->events[3] == 0) {
-      TRACE(NCCL_NET, "r=%p done", r);
-      *done = 1;
-            // If this was a failed request, we were just draining completions.
-      // Now that all events are done, free and return success.
-      if (r->type == NCCL_NET_IB_REQ_FAILED) {
-        NCCLCHECK(ncclIbFreeRequest(r));
-        // Return success since we successfully drained. The fatal error
-        // is already recorded in stats and will be caught on next operation.
-        return ncclSuccess;
-      }
-      if (sizes && r->type == NCCL_NET_IB_REQ_RECV) {
-        for (int i=0; i<r->nreqs; i++) {
-          sizes[i] = r->recv.sizes[i];
-#ifdef NCCL_ENABLE_NET_PROFILING
-          for (int j = 0; j < r->pInfo[i].nEventHandles; j++) {
-            NCCLCHECK(ncclProfilerFunction(&r->pInfo[i].qpEventHandles[j], ncclProfilerNetEventStop, NULL, 0, NULL));
-          }
-#endif
-        }
-      }
-      if (sizes && r->type == NCCL_NET_IB_REQ_SEND) {
-        sizes[0] = r->send.size;
-#ifdef NCCL_ENABLE_NET_PROFILING
-        for (int j = 0; j < r->pInfo[0].nEventHandles; j++) {
-          NCCLCHECK(ncclProfilerFunction(&r->pInfo[0].qpEventHandles[j], ncclProfilerNetEventStop, NULL, 0, NULL));
-        }
-#endif
-      }
-      // Stop all remaining Qp events for this event
-      NCCLCHECK(ncclIbFreeRequest(r));
-      return ncclSuccess;
-    }
-
-    int totalWrDone = 0;
-    int wrDone = 0;
-    struct ibv_wc wcs[4];
-
-    for (int i = 0; i < NCCL_IB_MAX_DEVS_PER_NIC; i++) {
-      TIME_START(3);
-      // If we expect any completions from this device's CQ
-      if (r->events[i]) {
-        ret = wrap_ibv_poll_cq(r->devBases[i]->cq, 4, wcs, &wrDone);
-        if (ret != ncclSuccess) {
-          failDevIdx = i;
-          goto fail;
-        }
-        totalWrDone += wrDone;
-        if (wrDone == 0) { TIME_CANCEL(3); } else { TIME_STOP(3); }
-        if (wrDone == 0) continue;
-        for (int w=0; w<wrDone; w++) {
-          struct ibv_wc *wc = wcs+w;
-          if (wc->status != IBV_WC_SUCCESS) {
-            union ncclSocketAddress addr;
-            ncclSocketGetAddr(r->sock, &addr);
-            char localGidString[INET6_ADDRSTRLEN] = "";
-            char remoteGidString[INET6_ADDRSTRLEN] = "";
-            const char* localGidStr = NULL, *remoteGidStr = NULL;
-            if (r->devBases[i]->gidInfo.link_layer == IBV_LINK_LAYER_ETHERNET) {
-              localGidStr = ibvGetGidStr(&r->devBases[i]->gidInfo.localGid, localGidString, sizeof(localGidString));
-              remoteGidStr = ibvGetGidStr(&r->base->remDevs[i].remoteGid, remoteGidString, sizeof(remoteGidString));
-            }
-
-            char line[SOCKET_NAME_MAXLEN+1];
-            char *hcaName = r->devBases[i]->pd->context->device->name;
-            int reqSize = wc->byte_len;
-            struct ncclIbRequest* req = r->base->reqs+(wc->wr_id & 0xff);
-            if (req && req->type == NCCL_NET_IB_REQ_SEND) {
-              // For Send use the request size as WC byte_len is not reliable
-              reqSize = req->send.size;
-            }
-            WARN("NET/IB: Got completion from peer %s with status=%s(%d) opcode=%s(%d) reqSize=%d vendor_err=%u req_type=%s%s%s%s%s hca %s",
-                ncclSocketToString(&addr, line), ibvWcStatusStr(wc->status), wc->status,
-                ibvWcOpcodeStr(wc->opcode), wc->opcode, reqSize, wc->vendor_err, reqTypeStr[r->type],
-                localGidStr ?  " localGid ":"", localGidString, remoteGidStr ? " remoteGids":"", remoteGidString, hcaName);
-            return ncclRemoteError;
-          }
-
-          union ncclSocketAddress addr;
-          ncclSocketGetAddr(r->sock, &addr);
-          struct ncclIbRequest* req = r->base->reqs+(wc->wr_id & 0xff);
-
-          #ifdef ENABLE_TRACE
-          char line[SOCKET_NAME_MAXLEN+1];
-          TRACE(NCCL_NET, "Got completion from peer %s with status=%d opcode=%d len=%u wr_id=%lu r=%p type=%d events={%d,%d,%d,%d}, i=%d",
-            ncclSocketToString(&addr, line), wc->status, wc->opcode,wc->byte_len, wc->wr_id, req, req->type, req->events[0], req->events[1], req->events[2], req->events[3], i);
-          #endif
-          if (req && req->type == NCCL_NET_IB_REQ_SEND) {
-            for (int j = 0; j < req->nreqs; j++) {
-              struct ncclIbRequest* sendReq = r->base->reqs+((wc->wr_id >> (j*8)) & 0xff);
-              if ((sendReq->events[i] <= 0)) {
-                WARN("NET/IB: sendReq(%p)->events={%d,%d,%d,%d}, i=%d, j=%d <= 0", sendReq, sendReq->events[0], sendReq->events[1], sendReq->events[2], sendReq->events[3], i, j);
-                return ncclInternalError;
-              }
-              sendReq->events[i]--;
-#ifdef NCCL_ENABLE_NET_PROFILING
-              // Stop Qp event for sendReq
-              int qpIndex = getReqQpIndex(sendReq, j, wc->qp_num);
-              NCCLCHECK(ncclProfilerFunction(&sendReq->pInfo[j].qpEventHandles[qpIndex], ncclProfilerNetEventStop, NULL, 0, NULL));
-#endif
-            }
-          } else {
-            if (req && wc->opcode == IBV_WC_RECV_RDMA_WITH_IMM) {
-              if (req->type != NCCL_NET_IB_REQ_RECV) {
-                WARN("NET/IB: wc->opcode == IBV_WC_RECV_RDMA_WITH_IMM and req->type=%d", req->type);
-                return ncclInternalError;
-              }
-              if (req->nreqs == 1) {
-                req->recv.sizes[0] = wc->imm_data;
-              }
-            }
-            req->events[i]--;
-#ifdef NCCL_ENABLE_NET_PROFILING
-            // Stop Qp event for workFifo
-            for (int j = 0; j < req->nreqs; j++) {
-              int qpIndex = getReqQpIndex(req, j, wc->qp_num);
-              NCCLCHECK(ncclProfilerFunction(&req->pInfo[j].qpEventHandles[qpIndex], ncclProfilerNetEventStop, NULL, 0, NULL));
-            }
-#endif
-          }
-        }
-        // Once the IB fatal event is reported in the async thread, we want to propagate this error
-        // to communicator and prevent further polling to reduce error pollution.
-        ret = ncclIbStatsCheckFatalCount(&ncclIbDevs[r->devBases[i]->ibDevN].stats,__func__);
-        if (ret != ncclSuccess) {
-          failDevIdx = i;
-          goto fail;
-        }
-      }
-    }
-
-    // If no CQEs found on any device, return and come back later
-    if (totalWrDone == 0) return ncclSuccess;
-  }
-  fail:
-  // Mark connection and device as fatal
-  ncclIbStatsFatalError(&r->base->stats);
-  if (failDevIdx >= 0 && r->devBases[failDevIdx] != NULL) {
-    ncclIbStatsFatalError(&ncclIbDevs[r->devBases[failDevIdx]->ibDevN].stats);
-  }
-  // Mark request as failed so subsequent Test calls drain completions
-  r->type = NCCL_NET_IB_REQ_FAILED;
-  *done = 1;
-  return ret;
-}
-
-ncclResult_t ncclIbCloseSend(void* sendComm) {
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)sendComm;
-  if (comm) {
-    NCCLCHECK(ncclSocketClose(&comm->base.sock));
-
-    for (int q = 0; q < comm->base.nqps; q++)
-      if (comm->base.qps[q].qp != NULL) NCCLCHECK(wrap_ibv_destroy_qp(comm->base.qps[q].qp));
-
-    for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-      struct ncclIbSendCommDev* commDev = comm->devs + i;
-      if (commDev->fifoMr != NULL) NCCLCHECK(wrap_ibv_dereg_mr(commDev->fifoMr));
-      if (comm->remSizesFifo.mrs[i] != NULL) NCCLCHECK(wrap_ibv_dereg_mr(comm->remSizesFifo.mrs[i]));
-      if (commDev->putSignalScratchpadMr != NULL)
-        NCCLCHECK(wrap_ibv_dereg_mr(commDev->putSignalScratchpadMr));
-      NCCLCHECK(ncclIbDestroyBase(&commDev->base));
-    }
-
-    free(comm);
-  }
-  TIME_PRINT("IB");
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbCloseRecv(void* recvComm) {
-  struct ncclIbRecvComm* comm = (struct ncclIbRecvComm*)recvComm;
-  if (comm) {
-    NCCLCHECK(ncclSocketClose(&comm->base.sock));
-
-    for (int q = 0; q < comm->base.nqps; q++)
-      if (comm->base.qps[q].qp != NULL) NCCLCHECK(wrap_ibv_destroy_qp(comm->base.qps[q].qp));
-
-    for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-      struct ncclIbRecvCommDev* commDev = comm->devs + i;
-      if (comm->flushEnabled) {
-        if (commDev->gpuFlush.gpuFlushGpuMem != nullptr) {
-          NCCLCHECK(ncclCudaFree(commDev->gpuFlush.gpuFlushGpuMem, /*manager=*/nullptr));
-          commDev->gpuFlush.gpuFlushGpuMem = nullptr;
-          if (commDev->gpuFlush.gpuMr != nullptr) NCCLCHECK(wrap_ibv_dereg_mr(commDev->gpuFlush.gpuMr));
-          commDev->gpuFlush.gpuMr = nullptr;
-          if(commDev->gpuFlush.dmabuf_fd > 0) { close(commDev->gpuFlush.dmabuf_fd);}
-        }
-        if (commDev->gpuFlush.qp.qp != NULL) NCCLCHECK(wrap_ibv_destroy_qp(commDev->gpuFlush.qp.qp));
-        if (commDev->gpuFlush.hostMr != NULL) NCCLCHECK(wrap_ibv_dereg_mr(commDev->gpuFlush.hostMr));
-      }
-      if (commDev->fifoMr != NULL) NCCLCHECK(wrap_ibv_dereg_mr(commDev->fifoMr));
-      if (commDev->sizesFifoMr != NULL) NCCLCHECK(wrap_ibv_dereg_mr(commDev->sizesFifoMr));
-      NCCLCHECK(ncclIbDestroyBase(&commDev->base));
-    }
-    free(comm);
-  }
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbCloseListen(void* listenComm) {
-  struct ncclIbListenComm* comm = (struct ncclIbListenComm*)listenComm;
-  if (comm) {
-    NCCLCHECK(ncclSocketClose(&comm->sock));
-    free(comm);
-  }
-  return ncclSuccess;
-}
-
-ncclResult_t ncclIbFinalize(void* ctx) {
-  free(ctx);
-  return ncclIbFinalizeDevices();
-}
-
-ncclResult_t rcclNetP2pPolicy(void* handle, int isP2p) {
-  if (!handle) return ncclInvalidArgument;
-  struct ncclIbHandle* ibHandle = (struct ncclIbHandle*)handle;
-  if (ibHandle->magic != NCCL_SOCKET_MAGIC) return ncclInvalidArgument;
-  ibHandle->isP2p = isP2p;
-  return ncclSuccess;
-}
-
-ncclNet_t ncclNetIb = {
-  "IB",
-  ncclIbInit,
-  ncclIbDevices,
-  ncclIbGetProperties,
-  ncclIbListen,
-  ncclIbConnect,
-  ncclIbAccept,
-  ncclIbRegMr,
-  ncclIbRegMrDmaBuf,
-  ncclIbDeregMr,
-  ncclIbIsend,
-  ncclIbIrecv,
-  ncclIbIflush,
-  ncclIbTest,
-  ncclIbCloseSend,
-  ncclIbCloseRecv,
-  ncclIbCloseListen,
-  NULL /* getDeviceMr */,
-  NULL /* irecvConsumed */,
-  ncclIbMakeVDevice,
-  ncclIbFinalize,
-  ncclIbSetNetAttr,
-};
-
-
-
-/// GIN IB Plugin
-
-#include "gin/gin_host.h"
-// [RCCL] hipify add_file_unique() renames net_ib/gin.h to net_ib/gin_tmp.h
-// to avoid a basename collision with include/nccl_device/gin.h.
-#include "net_ib/gin_tmp.h"
-
-const int NCCL_GIN_IB_ALLGATHER_TAG = 0xa0;
-const int NCCL_GIN_IB_ALLTOALL_TAG = 0xa1;
-
-ncclResult_t ncclGinIbInit(void** ctx, uint64_t commId, ncclDebugLogger_t logFunction) {
-  ncclNetCommConfig_t* netCommConfig = nullptr;
-  NCCLCHECK(ncclIbInitDevices(logFunction, nullptr));
-  NCCLCHECK(ncclCalloc(&netCommConfig, 1));
-  *ctx = netCommConfig;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbFinalize(void *ctx) {
-  if (ctx) free(ctx);
-  return ncclIbFinalizeDevices();
-}
-
-static ncclResult_t ncclGinIbAllGather(struct ncclGinIbCollComm *cComm, void *srcBuf, void *recvBuf, size_t len) {
-  ncclResult_t status = ncclSuccess;
-  void *rMhandle = NULL, *sMhandle = NULL;
-  void *srequest = NULL, *rrequest = NULL;
-  int speer;
-  int rpeer;
-  void *rbuf;
-  int tag;
-  int done;
-
-  NCCLCHECKGOTO(ncclNetIb.regMr(cComm->recvComm, recvBuf,
-                                cComm->nranks * len, NCCL_PTR_HOST,
-                                &rMhandle),
-                status, out);
-  NCCLCHECKGOTO(ncclNetIb.regMr(cComm->sendComm, recvBuf,
-                                cComm->nranks * len, NCCL_PTR_HOST,
-                                &sMhandle),
-                status, out);
-
-  speer = cComm->rank;
-  memcpy((void *)((uintptr_t)recvBuf + speer * len), srcBuf, len);
-  for (int i = 0; i < cComm->nranks - 1; i++) {
-    rpeer = (speer - 1 + cComm->nranks) % cComm->nranks;
-    while (srequest == NULL || rrequest == NULL) {
-      rbuf = (void *)((uintptr_t)recvBuf + rpeer * len);
-      tag = NCCL_GIN_IB_ALLGATHER_TAG;
-      if (srequest == NULL)
-        NCCLCHECKGOTO(ncclNetIb.isend(cComm->sendComm,
-                                      (void *)((uintptr_t)recvBuf + speer * len),
-                                      len, tag, sMhandle, NULL, &srequest),
-                      status, out);
-      if (rrequest == NULL)
-        NCCLCHECKGOTO(ncclNetIb.irecv(cComm->recvComm, 1, &rbuf, &len,
-                                      &tag, &rMhandle, NULL, &rrequest),
-                      status, out);
-    }
-    while (srequest || rrequest) {
-      if (rrequest)
-        NCCLCHECKGOTO(ncclNetIb.test(rrequest, &done, NULL),
-                      status, out);
-      if (done)
-        rrequest = NULL;
-      if (srequest)
-        NCCLCHECKGOTO(ncclNetIb.test(srequest, &done, NULL),
-                      status, out);
-      if (done)
-        srequest = NULL;
-    }
-    speer = rpeer;
-  }
-
-out:
-  if (rMhandle)
-    ncclNetIb.deregMr(cComm->recvComm, rMhandle);
-
-  if (sMhandle)
-    ncclNetIb.deregMr(cComm->sendComm, sMhandle);
-
-  return status;
-}
-
-static ncclResult_t ncclGinIbAllToAll(struct ncclGinIbCollComm *cComm, void *src_buf, void *recv_buf, size_t len) {
-  ncclResult_t status = ncclSuccess;
-
-  void *tmp_buf = nullptr;
-  NCCLCHECK(ncclIbMalloc((void **)&tmp_buf, cComm->nranks * cComm->nranks * len));
-  NCCLCHECKGOTO(cComm->allGather(cComm, src_buf, tmp_buf, cComm->nranks * len), status, out);
-
-  for (int i = 0; i < cComm->nranks; i++) {
-    memcpy((void *)((uintptr_t)recv_buf + i * len), (void *)((uintptr_t)tmp_buf + i * cComm->nranks * len + cComm->rank * len), len);
-  }
-
-out:
-  if (tmp_buf)
-    free(tmp_buf);
-
-  return status;
-}
-
-ncclResult_t ncclGinIbP2PBarrier(struct ncclGinIbCollComm *cComm) {
-  // TODO: move allocation to init or use zero-byte allgather
-  int *dummy;
-  NCCLCHECK(ncclIbMalloc((void **)&dummy, cComm->nranks * sizeof(int)));
-  NCCLCHECK(ncclGinIbAllGather(cComm, dummy + cComm->rank * sizeof(int),
-                               dummy, sizeof(int)));
-  free(dummy);
-  return ncclSuccess;
-}
-
-// [RCCL] NCCL 2.29.7 added nConnections + queueDepth args to ncclGin_t::connect.
-// We don't yet thread connection multiplexing through this transport so we
-// accept the new args and ignore them; behaviour matches the prior
-// nConnections == 1 single-QP setup.
-ncclResult_t ncclGinIbConnect(void* ctx, void* handles[], int nranks, int rank, int nConnections, int queueDepth, void* listenComm, void** collComm) {
-  (void)nConnections; (void)queueDepth;
-  struct ncclIbListenComm *lComm = (struct ncclIbListenComm *)listenComm;
-  struct ncclGinIbCollComm *cComm = nullptr;
-  int next;
-
-  NCCLCHECK(ncclIbMalloc((void **)&cComm, sizeof(*cComm)));
-  NCCLCHECK(ncclIbMalloc((void**)&cComm->fullSendComm, sizeof(void *) * nranks));
-  NCCLCHECK(ncclIbMalloc((void**)&cComm->fullRecvComm, sizeof(void *) * nranks));
-
-  cComm->nranks = nranks;
-  cComm->rank = rank;
-
-  next = (cComm->rank + 1) % nranks;
-  do
-  {
-    if (cComm->sendComm == NULL) {
-      NCCLCHECK(ncclNetIb.connect(ctx, lComm->dev, handles[next], &cComm->sendComm, NULL));
-    }
-    if (cComm->recvComm == NULL)
-      NCCLCHECK(ncclNetIb.accept(lComm, &cComm->recvComm, NULL));
-  } while (cComm->sendComm == NULL || cComm->recvComm == NULL);
-
-  cComm->getProperties = (ncclResult_t(*)(int dev, void *props))ncclIbGetProperties;
-  cComm->allGather = ncclGinIbAllGather;
-  cComm->allToAll = ncclGinIbAllToAll;
-  cComm->getGidIndex = ncclIbGetGidIndex;
-  cComm->dev = lComm->dev;
-
-  for (int i = 0; i < nranks; i++)
-  {
-    int connectPeer = (cComm->rank + i) % nranks;
-    int acceptPeer = (cComm->rank - i + nranks) % nranks;
-    do
-    {
-      if (cComm->fullSendComm[connectPeer] == NULL)
-        NCCLCHECK(ncclNetIb.connect(ctx, lComm->dev, handles[connectPeer], &cComm->fullSendComm[connectPeer], NULL));
-      if (cComm->fullRecvComm[acceptPeer] == NULL)
-        NCCLCHECK(ncclNetIb.accept(lComm, &cComm->fullRecvComm[acceptPeer], NULL));
-    } while ((cComm->fullSendComm[connectPeer] == NULL) || (cComm->fullRecvComm[acceptPeer] == NULL));
-    NCCLCHECK(ncclGinIbP2PBarrier(cComm));
-  }
-
-  *collComm = cComm;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbCloseColl(void* collComm) {
-  struct ncclGinIbCollComm* cComm = (struct ncclGinIbCollComm*)collComm;
-  if (!cComm) return ncclSuccess;
-
-  if (cComm->fullRecvComm) {
-    for (int i=0; i<cComm->nranks; i++) {
-      NCCLCHECK(ncclNetIb.closeRecv(cComm->fullRecvComm[i]));
-    }
-    free(cComm->fullRecvComm);
-    cComm->fullRecvComm = NULL;
-  }
-
-  if (cComm->fullSendComm) {
-    for (int i=0; i<cComm->nranks; i++) {
-      NCCLCHECK(ncclNetIb.closeSend(cComm->fullSendComm[i]));
-    }
-    free(cComm->fullSendComm);
-    cComm->fullSendComm = NULL;
-  }
-
-  if (cComm->recvComm) {
-    NCCLCHECK(ncclNetIb.closeRecv(cComm->recvComm));
-    cComm->recvComm = NULL;
-  }
-
-  if (cComm->sendComm) {
-    NCCLCHECK(ncclNetIb.closeSend(cComm->sendComm));
-    cComm->sendComm = NULL;
-  }
-
-  memset(cComm, 0, sizeof(*cComm));
-
-  free(cComm);
-  return ncclSuccess;
-}
-
-#if !defined(__HIP_PLATFORM_AMD__)
-#include "net_ib/gdaki/gin_host_gdaki.h"
-
-static std::mutex ncclGinIbGdakiLockMutex;
-static int ncclGinIbGdakiNDevs = -1;
-int ncclGinIbGdakiDevIndexes[MAX_IB_DEVS];
-
-ncclResult_t ncclGinIbGdakiInit(void** ctx, uint64_t commId, ncclDebugLogger_t logFunction) {
-  NCCLCHECK(ncclGinIbInit(ctx, commId, logFunction));
-  std::lock_guard<std::mutex> lock(ncclGinIbGdakiLockMutex);
-  if (ncclGinIbGdakiNDevs == -1) {
-    int ndevs = 0;
-    for (int i = 0; i < ncclNIbDevs; i++) {
-      if (ncclIbDevs[i].ibProvider == IB_PROVIDER_MLX5) {
-        ncclGinIbGdakiDevIndexes[ndevs] = i;
-        ++ndevs;
-      }
-    }
-    ncclGinIbGdakiNDevs = ndevs;
-  }
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbGdakiDevices(int* ndev) {
-  std::lock_guard<std::mutex> lock(ncclGinIbGdakiLockMutex);
-  *ndev = ncclGinIbGdakiNDevs;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbGdakiGetProperties(int dev, ncclNetProperties_t* props) {
-  std::lock_guard<std::mutex> lock(ncclGinIbGdakiLockMutex);
-  NCCLCHECK(ncclNetIb.getProperties(ncclGinIbGdakiDevIndexes[dev], props));
-  props->netDeviceType = NCCL_NET_DEVICE_GIN_GDAKI;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbGdakiListen(void* ctx, int dev, void* opaqueHandle, void** listenComm) {
-  std::lock_guard<std::mutex> lock(ncclGinIbGdakiLockMutex);
-  return ncclNetIb.listen(ctx, ncclGinIbGdakiDevIndexes[dev], opaqueHandle, listenComm);
-}
-
-// [RCCL] 2.29.7: createContext gained nContexts; regMrSym gained mr_flags.
-ncclResult_t ncclGinIbGdakiCreateContext(void* collComm, int nSignals, int nCounters, int nContexts, void **ginCtx, ncclNetDeviceHandle_v11_t** devHandle) {
-  struct ncclGinIbCollComm* cComm = (struct ncclGinIbCollComm*)collComm;
-
-  NCCLCHECK(ncclGinGdakiCreateContext(cComm, nSignals, nCounters, nContexts, ginCtx, devHandle));
-
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbGdakiRegMrSym(void* collComm, void* data, size_t size, int type, uint64_t mr_flags, void** mhandle, void **ginHandle) {
-  return ncclGinGdakiRegMrSym((struct ncclGinIbCollComm *)collComm, data, size, type, mr_flags, mhandle, ginHandle);
-}
-
-ncclResult_t ncclGinIbGdakiDeregMrSym(void* collComm, void* mhandle) {
-  return ncclGinGdakiDeregMrSym((struct ncclGinIbCollComm *)collComm, mhandle);
-}
-
-ncclResult_t ncclGinIbGdakiDestroyContext(void* ginCtx) {
-  return ncclGinGdakiDestroyContext(ginCtx);
-}
-
-ncclResult_t ncclGinIbGdakiProgress(void *collComm)
-{
-  return ncclGinGdakiProgress(collComm);
-}
-
-ncclResult_t ncclGinIbGdakiQueryLastError(void *ginCtx, bool *hasError) {
-  return ncclGinGdakiQueryLastError(ginCtx, hasError);
-}
-
-ncclGin_t ncclGinIbGdaki = {
-  "GIN_IB_GDAKI",
-  ncclGinIbGdakiInit,
-  ncclGinIbGdakiDevices,
-  ncclGinIbGdakiGetProperties,
-  ncclGinIbGdakiListen,
-  ncclGinIbConnect,
-  ncclGinIbGdakiCreateContext,
-  ncclGinIbGdakiRegMrSym,
-  NULL, // regMrSymDmaBuf
-  ncclGinIbGdakiDeregMrSym,
-  ncclGinIbGdakiDestroyContext,
-  ncclGinIbCloseColl,
-  ncclIbCloseListen,
-  NULL,
-  NULL,
-  NULL,
-  ncclGinIbGdakiProgress,
-  ncclGinIbGdakiQueryLastError,
-  ncclGinIbFinalize
-};
-
-#endif // !defined(__HIP_PLATFORM_AMD__)
-
-struct ncclIbGinProxyMrHandle {
-  struct ncclIbMrHandle *mrHandle;
-  uintptr_t *base_vas;
-  uint32_t *rkeys;
-};
-
-ncclResult_t ncclGinIbProxyGetProperties(int dev, ncclNetProperties_t* props) {
-  NCCLCHECK(ncclNetIb.getProperties(dev, props));
-  props->netDeviceType = NCCL_NET_DEVICE_GIN_PROXY;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbProxyRegMrSymDmaBuf(void* collComm, void* data, size_t size, int type, uint64_t offset, int fd, uint64_t mr_flags, void** mhandle, void **ginHandle) {
-  struct ncclGinIbCollComm *cComm = (struct ncclGinIbCollComm *)collComm;
-  struct ncclIbGinProxyMrHandle *ginMrHandle;
-  NCCLCHECK(ncclCalloc(&ginMrHandle, 1));
-
-  NCCLCHECKNOWARN(ncclIbRegMrDmaBuf(cComm->recvComm, data, size, type, offset, fd, (void**)&ginMrHandle->mrHandle), NCCL_NET);
-
-  NCCLCHECK(ncclCalloc(&ginMrHandle->base_vas, cComm->nranks));
-  NCCLCHECK(ncclCalloc(&ginMrHandle->rkeys, cComm->nranks));
-
-  NCCLCHECK(cComm->allGather(cComm, &data, ginMrHandle->base_vas, sizeof(uintptr_t)));
-  NCCLCHECK(cComm->allGather(cComm, &ginMrHandle->mrHandle->mrs[0]->rkey, ginMrHandle->rkeys, sizeof(uint32_t)));
-
-  *mhandle = ginMrHandle;
-  *ginHandle = ginMrHandle;
-
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbProxyRegMrSym(void* collComm, void* data, size_t size, int type, uint64_t mr_flags, void** mhandle, void **ginHandle) {
-  return ncclGinIbProxyRegMrSymDmaBuf(collComm, data, size, type, 0, -1, mr_flags, mhandle, ginHandle);
-}
-
-ncclResult_t ncclGinIbProxyDeregMrSym(void* collComm, void* mhandle) {
-  struct ncclGinIbCollComm *cComm = (struct ncclGinIbCollComm *)collComm;
-  struct ncclIbGinProxyMrHandle *ginMrHandle = (struct ncclIbGinProxyMrHandle *)mhandle;
-
-  NCCLCHECK(ncclNetIb.deregMr(cComm->recvComm, ginMrHandle->mrHandle));
-  free(ginMrHandle->base_vas);
-  free(ginMrHandle->rkeys);
-  free(ginMrHandle);
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbProxyCloseColl(void* collComm) {
-  free(collComm);
-  return ncclSuccess;
-}
-
-// [RCCL] 2.29.7 added an int connectionId parameter -- we don't yet model
-// per-connection multiplexing here, so just ignore it (matches connectionId==0).
-ncclResult_t ncclGinIbProxyIPut(void *collComm, uint64_t srcOff, void *srcMhandle, size_t size,
-                                uint64_t dstOff, void *dstMhandle, uint32_t rank, int connectionId, void **request)
-{
-  (void)connectionId;
-  struct ncclGinIbCollComm* cComm = (struct ncclGinIbCollComm*)collComm;
-
-  struct ncclIbGinProxyMrHandle *srcMrHandle = (struct ncclIbGinProxyMrHandle *)srcMhandle;
-  struct ncclIbGinProxyMrHandle *dstMrHandle = (struct ncclIbGinProxyMrHandle *)dstMhandle;
-
-  void *srcPtr = (void *)(srcMrHandle->base_vas[cComm->rank] + srcOff);
-  void *dstPtr = (void *)(dstMrHandle->base_vas[rank] + dstOff);
-  uint32_t lkey = srcMrHandle->mrHandle->mrs[0]->lkey;
-  uint32_t rkey = dstMrHandle->rkeys[rank];
-
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)cComm->fullSendComm[rank];
-  struct ncclIbQp *qp = &comm->base.qps[0];
-
-  struct ncclIbRequest* req;
-  NCCLCHECK(ncclIbGetRequest(&comm->base, &req));
-  req->type = NCCL_NET_IB_REQ_GIN_IPUT;
-  req->sock = &comm->base.sock;
-  req->iput.rank = rank;
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    req->devBases[i] = &comm->devs[i].base;
-  }
-
-  struct ibv_send_wr wr;
-  memset(&wr, 0, sizeof(wr));
-  struct ibv_sge sge;
-  memset(&sge, 0, sizeof(sge));
-
-  wr.opcode                  = IBV_WR_RDMA_WRITE;
-  wr.send_flags              = IBV_SEND_SIGNALED;
-  wr.wr_id                   = req - comm->base.reqs;
-  wr.next                    = NULL;
-  wr.wr.rdma.remote_addr     = (uint64_t)dstPtr;
-  wr.wr.rdma.rkey            = rkey;
-  wr.sg_list = &sge;
-  wr.num_sge = 1;
-
-  sge.addr = (uintptr_t)srcPtr;  // Local buffer address
-  sge.length = size;  // Size of the transfer
-  sge.lkey = lkey;  // Local key
-
-  struct ibv_send_wr* bad_wr;
-  NCCLCHECK(wrap_ibv_post_send(qp->qp, &wr, &bad_wr));
-  ncclIbAddEvent(req, qp->devIndex, &comm->devs[qp->devIndex].base);
-
-  *request = req;
-  return ncclSuccess;
-}
-
-// [RCCL] 2.29.7 added an int connectionId parameter -- ignored as above.
-ncclResult_t ncclGinIbProxyIPutSignal(void *collComm, uint64_t srcOff, void *srcMhandle,
-                                      size_t size, uint64_t dstOff, void *dstMhandle,
-                                      uint32_t rank, uint64_t signalOff, void *signalMhandle,
-                                      uint64_t signalValue, uint32_t signalOp, int connectionId, void **request)
-{
-  (void)connectionId;
-  if (signalOp != NCCL_NET_SIGNAL_OP_INC && signalOp != NCCL_NET_SIGNAL_OP_ADD) {
-    WARN("ncclGinIbProxyIPutSignal: Unsupported signalOp %u", signalOp);
-    return ncclInvalidArgument;
-  }
-
-  struct ncclGinIbCollComm* cComm = (struct ncclGinIbCollComm*)collComm;
-
-  struct ncclIbGinProxyMrHandle *srcMrHandle = (struct ncclIbGinProxyMrHandle *)srcMhandle;
-  struct ncclIbGinProxyMrHandle *dstMrHandle = (struct ncclIbGinProxyMrHandle *)dstMhandle;
-  struct ncclIbGinProxyMrHandle *signalMrHandle = (struct ncclIbGinProxyMrHandle *)signalMhandle;
-
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)cComm->fullSendComm[rank];
-  struct ncclIbQp *qp = &comm->base.qps[0];
-  int devIndex = qp->devIndex;
-
-  struct ncclIbRequest* req;
-  NCCLCHECK(ncclIbGetRequest(&comm->base, &req));
-  req->type = NCCL_NET_IB_REQ_GIN_IPUT;
-  req->sock = &comm->base.sock;
-  req->iput.rank = rank;
-  for (int i = 0; i < comm->base.vProps.ndevs; i++) {
-    req->devBases[i] = &comm->devs[i].base;
-  }
-
-  struct ibv_send_wr wr[2];
-  memset(&wr, 0, sizeof(wr));
-  struct ibv_sge sge[2];
-  memset(&sge, 0, sizeof(sge));
-
-  // If size is 0, we only need to send the signal. srcMrHandle must be non-NULL
-  if (size > 0 && dstMrHandle) {
-    void *srcPtr = (void *)(srcMrHandle->base_vas[cComm->rank] + srcOff);
-    void *dstPtr = (void *)(dstMrHandle->base_vas[rank] + dstOff);
-    uint32_t lkey = srcMrHandle->mrHandle->mrs[0]->lkey;
-    uint32_t rkey = dstMrHandle->rkeys[rank];
-
-    // PUT
-    wr[0].opcode                  = IBV_WR_RDMA_WRITE;
-    wr[0].send_flags              = 0; // We only need the CQE from the signal
-    wr[0].wr_id                   = req - comm->base.reqs;
-    wr[0].next                    = &wr[1];
-    wr[0].wr.rdma.remote_addr     = (uint64_t)dstPtr;
-    wr[0].wr.rdma.rkey            = rkey;
-    wr[0].sg_list = &sge[0];
-    wr[0].num_sge = 1;
-
-    sge[0].addr = (uintptr_t)srcPtr;  // Local buffer address
-    sge[0].length = size;  // Size of the transfer
-    sge[0].lkey = lkey;  // Local key
-  }
-
-  void *signalPtr = (void *)(signalMrHandle->base_vas[rank] + signalOff);
-  uint32_t signalRkey = signalMrHandle->rkeys[rank];
-
-  // SIGNAL
-  wr[1].opcode                  = IBV_WR_ATOMIC_FETCH_AND_ADD;
-  wr[1].send_flags              = IBV_SEND_SIGNALED;
-  wr[1].wr_id                   = req - comm->base.reqs;  // used for matching completions with request
-  wr[1].next                    = NULL;
-  wr[1].wr.atomic.remote_addr   = (uint64_t)signalPtr;
-  wr[1].wr.atomic.compare_add   = signalOp == NCCL_NET_SIGNAL_OP_INC ? 1 : signalValue;
-  wr[1].wr.atomic.rkey          = signalRkey;
-  wr[1].sg_list = &sge[1];
-  wr[1].num_sge = 1;
-
-  sge[1].addr = (uintptr_t)&comm->putSignalScratchpad;
-  sge[1].length = sizeof(comm->putSignalScratchpad);
-  sge[1].lkey = comm->devs[devIndex].putSignalScratchpadMr->lkey;
-
-  // Send the put and the signal in one go
-  struct ibv_send_wr* bad_wr;
-  NCCLCHECK(wrap_ibv_post_send(qp->qp, size > 0 ? &wr[0] : &wr[1], &bad_wr));
-  ncclIbAddEvent(req, qp->devIndex, &comm->devs[qp->devIndex].base);
-  *request = req;
-  return ncclSuccess;
-}
-
-ncclResult_t ncclGinIbProxyTest(void *collComm, void *request, int *done) {
-  struct ncclGinIbCollComm* cComm = (struct ncclGinIbCollComm*)collComm;
-  struct ncclIbRequest* req = (struct ncclIbRequest*)request;
-  int rank = req->iput.rank;
-  *done = 0;
-
-  if (req->events[0] == 0) {
-    *done = 1;
-    NCCLCHECK(ncclIbFreeRequest(req));
-    return ncclSuccess;
-  }
-  int wrDone = 0;
-  struct ibv_wc wc[4];
-
-  struct ncclIbSendComm* comm = (struct ncclIbSendComm*)cComm->fullSendComm[rank];
-  NCCLCHECK(wrap_ibv_poll_cq(comm->devs[0].base.cq, 4, wc, &wrDone));
-  for (int i = 0; i < wrDone; i++) {
-    if (wc[i].status != IBV_WC_SUCCESS) {
-      union ncclSocketAddress addr;
-      ncclSocketGetAddr(req->sock, &addr);
-      char localGidString[INET6_ADDRSTRLEN] = "";
-      char remoteGidString[INET6_ADDRSTRLEN] = "";
-      const char* localGidStr = NULL, *remoteGidStr = NULL;
-      if (req->devBases[i]->gidInfo.link_layer == IBV_LINK_LAYER_ETHERNET) {
-        localGidStr = ibvGetGidStr(&req->devBases[i]->gidInfo.localGid, localGidString, sizeof(localGidString));
-        remoteGidStr = ibvGetGidStr(&req->base->remDevs[i].remoteGid, remoteGidString, sizeof(remoteGidString));
-      }
-
-      char line[SOCKET_NAME_MAXLEN+1];
-      char *hcaName = req->devBases[i]->pd->context->device->name;
-      WARN("NET/IB/GIN: Got completion from peer %s with status=%d opcode=%d len=%u vendor err %u (%s)%s%s%s%s hca %s",
-          ncclSocketToString(&addr, line), wc[i].status, wc[i].opcode, wc[i].byte_len, wc[i].vendor_err, reqTypeStr[req->type],
-          localGidStr ?  " localGid ":"", localGidString, remoteGidStr ? " remoteGids":"", remoteGidString, hcaName);
-      return ncclRemoteError;
-    }
-
-    struct ncclIbRequest* wcReq = comm->base.reqs + wc[i].wr_id;
-
-    wcReq->events[0]--;
-    if (wcReq == req && wcReq->events[0] == 0) {
-      *done = 1;
-      NCCLCHECK(ncclIbFreeRequest(wcReq));
-    }
-  }
-  return ncclSuccess;
-}
-
-// No support for NCCL_IB_SPLIT_DATA_ON_QPS or NCCL_IB_MERGE_NICS
-ncclGin_t ncclGinIbProxy = {
-  "GIN_IB_PROXY",
-  ncclGinIbInit,
-  ncclIbDevices,
-  ncclGinIbProxyGetProperties,
-  ncclIbListen,
-  ncclGinIbConnect,
-  NULL,
-  ncclGinIbProxyRegMrSym,
-  ncclGinIbProxyRegMrSymDmaBuf,
-  ncclGinIbProxyDeregMrSym,
-  NULL,
-  ncclGinIbCloseColl,
-  ncclIbCloseListen,
-  ncclGinIbProxyIPut,
-  ncclGinIbProxyIPutSignal,
-  ncclGinIbProxyTest,
-  NULL,
-  NULL,
-  ncclGinIbFinalize
-};
-
-// [RCCL] NCCL 2.29.7 introduced a top-level "ncclGinIb" dispatcher that
-// picks between GDAKI and Proxy at runtime. AMD doesn't ship the GDAKI
-// driver-mode kernels yet, so we point ncclGinIb at the Proxy
-// implementation -- this is a strict subset that always works on ROCm
-// HCAs and matches the existing behaviour. Whoever wires up GDAKI in
-// the future can replace this with a real dispatcher.
-ncclGin_t ncclGinIb = {
-  "GIN_IB",
-  ncclGinIbInit,
-  ncclIbDevices,
-  ncclGinIbProxyGetProperties,
-  ncclIbListen,
-  ncclGinIbConnect,
-  NULL,
-  ncclGinIbProxyRegMrSym,
-  ncclGinIbProxyRegMrSymDmaBuf,
-  ncclGinIbProxyDeregMrSym,
-  NULL,
-  ncclGinIbCloseColl,
-  ncclIbCloseListen,
-  ncclGinIbProxyIPut,
-  ncclGinIbProxyIPutSignal,
-  ncclGinIbProxyTest,
-  NULL,
-  NULL,
-  ncclGinIbFinalize
-};
