+++ projects/rccl/src/transport/net_ib/gin.cc
@@ -35,6 +35,7 @@ static ncclResult_t ncclGinIbGdrGpuSupport(bool gdaki) {
      ncclIbGdrSupport() == ncclSuccess;
   if (peerMemSupport) return ncclSuccess;
 
+#if !defined(__HIP_PLATFORM_AMD__)
   int cudaDev;
   CUDACHECK(cudaGetDevice(&cudaDev));
   int dmaBufSupportOnDevice = 1;
@@ -43,9 +44,15 @@ static ncclResult_t ncclGinIbGdrGpuSupport(bool gdaki) {
 
   WARN("Unable to use GIN: Peermem is not supported, and device %d does not support DMA-BUF.", cudaDev);
   return ncclInvalidUsage;
+#else
+  if (ncclIbDmaBufSupport(0) == ncclSuccess) return ncclSuccess;
+
+  WARN("Unable to use GIN: Peermem is not supported, nor DMA-BUF.");
+  return ncclInvalidUsage;
+#endif
 }
 
-NCCL_PARAM(GinType, "GIN_TYPE", -1);
+extern int64_t ncclParamGinType();
 
 static std::mutex ncclGinIbGdakiLockMutex;
 static int ncclGinIbGdakiNDevs = -1;
@@ -67,7 +74,9 @@ ncclResult_t ncclGinIbGdakiInit() {
 }
 
 extern ncclGin_t ncclGinIb;
+#if !defined(__HIP_PLATFORM_AMD__)
 extern ncclGin_t ncclGinIbGdaki;
+#endif // !defined(__HIP_PLATFORM_AMD__)
 extern ncclGin_t ncclGinIbProxy;
 
 // Initlialize GDAKI or PROXY backend. ginType can force a particular backend.
@@ -76,7 +85,9 @@ ncclResult_t ncclGinIbInitType(void** ctx, uint64_t commId, ncclDebugLogger_t lo
   NCCLCHECK(ncclIbInitDevices(logFunction, nullptr));
   if (ncclNIbDevs == 0) return ncclInternalError; // Caught in plugin init code, not propagated to user.
 
+#if !defined(__HIP_PLATFORM_AMD__)
   if (ginType == NCCL_GIN_TYPE_GDAKI) goto try_gdaki;
+#endif // !defined(__HIP_PLATFORM_AMD__)
   if (ginType == NCCL_GIN_TYPE_PROXY) goto try_proxy;
   if (ginType != -1) {
     INFO(NCCL_INIT|NCCL_NET, "NET_IB: no support for GIN type %ld", ncclParamGinType());
@@ -85,6 +96,7 @@ ncclResult_t ncclGinIbInitType(void** ctx, uint64_t commId, ncclDebugLogger_t lo
 
   bool gdrSupport;
 
+#if !defined(__HIP_PLATFORM_AMD__)
   // First try GDAKI
 try_gdaki:
   NCCLCHECK(ncclGinIbGdakiInit());
@@ -94,6 +106,7 @@ try_gdaki:
   if (!gdrSupport) return ncclInternalError;
   if (ginIb) memcpy(ginIb, &ncclGinIbGdaki, sizeof(ncclGinIb));
   goto end;
+#endif // !defined(__HIP_PLATFORM_AMD__)
 
   // Then Proxy
 try_proxy:
@@ -111,11 +124,50 @@ ncclResult_t ncclGinIbInit(void** ctx, uint64_t commId, ncclDebugLogger_t logFun
   return ncclGinIbInitType(ctx, commId, logFunction, ncclParamGinType(), &ncclGinIb);
 }
 
-// GIN Entry point, which will then morph into either the GDAKI or PROXY backend
+// Forward declarations for Proxy functions referenced in ncclGinIb dispatcher below.
+ncclResult_t ncclGinIbProxyGetProperties(int dev, ncclNetProperties_t* props);
+ncclResult_t ncclGinIbProxyRegMrSym(void* collComm, void* data, size_t size, int type, uint64_t mr_flags, void** mhandle, void **ginHandle);
+ncclResult_t ncclGinIbProxyRegMrSymDmaBuf(void* collComm, void* data, size_t size, int type, uint64_t offset, int fd, uint64_t mr_flags, void** mhandle, void **ginHandle);
+ncclResult_t ncclGinIbProxyDeregMrSym(void* collComm, void* mhandle);
+ncclResult_t ncclGinIbProxyIPut(void *collComm, uint64_t srcOff, void *srcMhandle, size_t size,
+                                uint64_t dstOff, void *dstMhandle, uint32_t rank, int connectionId,
+                                void **request);
+ncclResult_t ncclGinIbProxyIPutSignal(void *collComm, uint64_t srcOff, void *srcMhandle,
+                                      size_t size, uint64_t dstOff, void *dstMhandle, uint32_t rank,
+                                      uint64_t signalOff, void *signalMhandle, uint64_t signalValue,
+                                      uint32_t signalOp, int connectionId, void **request);
+ncclResult_t ncclGinIbProxyTest(void *collComm, void *request, int *done);
+ncclResult_t ncclGinIbFinalize(void *ctx);
+ncclResult_t ncclGinIbConnect(void *ctx, void *handles[], int nranks, int rank, int nConnections,
+                              int queueDepth, void *listenComm, void **collComm);
+ncclResult_t ncclGinIbCloseColl(void* collComm);
+
+// [RCCL] NCCL 2.29.7 introduced a top-level "ncclGinIb" dispatcher that
+// picks between GDAKI and Proxy at runtime. AMD doesn't ship the GDAKI
+// driver-mode kernels yet, so we point ncclGinIb at the Proxy
+// implementation -- this is a strict subset that always works on ROCm
+// HCAs and matches the existing behaviour. Whoever wires up GDAKI in
+// the future can replace this with a real dispatcher.
 ncclGin_t ncclGinIb = {
   "GIN_IB",
   ncclGinIbInit,
-  NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL
+  ncclIbDevices,
+  ncclGinIbProxyGetProperties,
+  ncclIbListen,
+  ncclGinIbConnect,
+  NULL,
+  ncclGinIbProxyRegMrSym,
+  ncclGinIbProxyRegMrSymDmaBuf,
+  ncclGinIbProxyDeregMrSym,
+  NULL,
+  ncclGinIbCloseColl,
+  ncclIbCloseListen,
+  ncclGinIbProxyIPut,
+  ncclGinIbProxyIPutSignal,
+  ncclGinIbProxyTest,
+  NULL,
+  NULL,
+  ncclGinIbFinalize
 };
 
 ncclResult_t ncclGinIbFinalize(void *ctx) {
@@ -303,6 +355,7 @@ ncclResult_t ncclGinIbCloseColl(void* collComm) {
   return ncclSuccess;
 }
 
+#if !defined(__HIP_PLATFORM_AMD__)
 #include "gdaki/gin_host_gdaki.h"
 
 ncclResult_t ncclGinIbGdakiInit(void** ctx, uint64_t commId, ncclDebugLogger_t logFunction) {
@@ -398,6 +451,7 @@ ncclGin_t ncclGinIbGdaki = {
   ncclGinIbGdakiQueryLastError,
   ncclGinIbFinalize
 };
+#endif // !defined(__HIP_PLATFORM_AMD__)
 
 
 struct ncclIbGinProxyMrHandle {
@@ -438,7 +492,7 @@ ncclResult_t ncclGinIbProxyRegMrSymDmaBuf(void* collComm, void* data, size_t siz
   struct ncclIbGinProxyMrHandle *ginMrHandle;
   NCCLCHECK(ncclCalloc(&ginMrHandle, 1));
 
-  NCCLCHECKNOWARN(ncclIbRegMrDmaBufInternal(cComm->recvComm, data, size, type, offset, fd, mr_flags, (void **)&ginMrHandle->mrHandle), NCCL_NET);
+  NCCLCHECKNOWARN(ncclIbRegMrDmaBuf(cComm->recvComm, data, size, type, offset, fd, (void**)&ginMrHandle->mrHandle), NCCL_NET);
 
   NCCLCHECK(ncclCalloc(&ginMrHandle->base_vas, cComm->nranks));
   NCCLCHECK(ncclCalloc(&ginMrHandle->rkeys, cComm->nranks));
